Extracted Ion Chromatogram parameters

Parameters for creation of extracted ion chromatograms.

getDefEICParams(...)

Arguments

...

optional named arguments that override defaults.

Details

To configure the creation of extracted ion chromatograms (EICs) several parameters exist:

rtWindow Retention time (in seconds) that will be subtracted/added to respectively the minimum and maximum retention time of the feature. Thus, setting this value to >0 will 'zoom out' on the retention time axis.

topMost Only create EICs for this number of top most intense features. If NULL then EICs are created for all features.

topMostByRGroup If set to TRUE and topMost is set: only create EICs for the top most features in each replicate group. For instance, when topMost=1 and topMostByRGroup=TRUE, then EICs will be plotted for the most intense feature of each replicate group.

onlyPresent If TRUE then EICs are created only for analyses in which a feature was detected. If onlyPresent=FALSE then EICs are generated for all analyses. The latter is handy to evaluate if a peak was 'missed' during feature detection or removed during e.g. filtering.

if onlyPresent=FALSE then the following parameters are also relevant:

mzExpWindow To create EICs for analyses in which no feature was found, the m/z value is derived from the min/max values of all features in the feature group. The value of mzExpWindow further expands this window.

setsAdductPos,setsAdductNeg (sets workflow) In sets workflows the adduct must be known to calculate the ionized m/z. If a feature is completely absent in a particular set then it follows no adduct annotations are available and the value of setsAdductPos (positive ionization data) or setsAdductNeg (negative ionization data) will be used instead.

These parameters are passed as a named list as the EICParams argument to functions that use EICs. The getDefEICParams function can be used to generate such parameter list with defaults.