Generation of MS Peak Lists

Functionality to generate MS peak lists.

# S4 method for featureGroups
generateMSPeakLists(fGroups, algorithm, ...)

generateMSPeakListsDA(
  fGroups,
  bgsubtr = TRUE,
  maxMSRtWindow = 5,
  minMSIntensity = 500,
  minMSMSIntensity = 500,
  clear = TRUE,
  close = TRUE,
  save = close,
  MSMSType = "MSMS",
  avgFGroupParams = getDefAvgPListParams()
)

generateMSPeakListsDAFMF(
  fGroups,
  minMSIntensity = 500,
  minMSMSIntensity = 500,
  close = TRUE,
  save = close,
  avgFGroupParams = getDefAvgPListParams()
)

generateMSPeakListsMzR(
  fGroups,
  maxMSRtWindow = 5,
  precursorMzWindow = 4,
  topMost = NULL,
  avgFeatParams = getDefAvgPListParams(),
  avgFGroupParams = getDefAvgPListParams()
)

getDefAvgPListParams(...)

Arguments

fGroups
The featureGroups object from which MS peak lists should be extracted.

algorithm
A character string describing the algorithm that should be used: "bruker", "brukerfmf", "mzr"

...
For generateMSPeakLists: Any parameters to be passed to the selected MS peak lists generation algorithm.

For getDefAvgPListParams: Optional named arguments that override defaults.

bgsubtr
If TRUE background will be subtracted using the 'spectral' algorithm.

maxMSRtWindow
Maximum chromatographic peak window used for spectrum averaging (in seconds, +/- retention time). If NULL all spectra from a feature will be taken into account. Lower to decrease processing time.

minMSIntensity, minMSMSIntensity
Minimum intensity for peak lists obtained with DataAnalysis. Highly recommended to set >0 as DA tends to report many very low intensity peaks.

clear
Remove any existing chromatogram traces/mass spectra prior to making new ones.

close, save
If TRUE then Bruker files are closed and saved after processing with DataAnalysis, respectively. Setting close=TRUE prevents that many analyses might be opened simultaneously in DataAnalysis, which otherwise may use excessive memory or become slow. By default save is TRUE when close is TRUE, which is likely what you want as otherwise any processed data is lost.

MSMSType
The type of MS/MS experiment performed: "MSMS" for MRM/AutoMSMS or "BBCID" for broadband CID.

precursorMzWindow
The m/z window (in Da) to find MS/MS spectra of a precursor. This is typically used for Data-Dependent like MS/MS data and should correspond to the isolation m/z window (i.e. +/- the precursor m/z) that was used to collect the data. For Data-Independent MS/MS experiments, where precursor ions are not isolated prior to fragmentation (e.g. bbCID, MSe, all-ion, ...) the value should be NULL.

topMost
Only extract MS peak lists from a maximum of topMost analyses with highest intensity. If NULL all analyses will be used.

avgFeatParams, avgFGroupParams
A list with parameters used for averaging of peak lists for individual features and feature groups, respectively (see below).

Value

A MSPeakLists object that can be used for formulae calculation and compound identification.

Details

Formula calculation and identification tools rely on mass spectra that belong to features of interest. For processing, MS (and MS/MS) spectra are typically reduced to a table with a column containing measured m/z values and a column containing their intensities. These 'MS peak lists' can then be used for formula generation and compound generation.

MS and MS/MS peak lists are first generated for all features (or a subset, if the topMost argument is set). During this step multiple spectra over the feature elution profile are averaged. Subsequently, peak lists will be generated for each feature group by averaging peak lists of the features within the group. Functionality that uses peak lists will either use data from individual features or from group averaged peak lists. For instance, the former may be used by formulae calculation, while compound identification and plotting functionality typically uses group averaged peak lists.

Several functions exist to automatically extract MS peak lists for feature groups.

generateMSPeakLists is a generic function that will generate MS peak lists using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateMSPeakListsMzR and generateMSPeakListsDA. While these functions may be called directly, generateMSPeakLists provides a generic interface and is therefore usually preferred.

generateMSPeakListsDA uses Bruker DataAnalysis to generate MS peak lists. Naturally, this only works with analyses in the Bruker data format (.d). This function leverages DataAnalysis functionality to support averaging of spectra, background subtraction and identification of isotopes. In order to obtain mass spectra TICs will be added of the MS and relevant MS/MS signals.

generateMSPeakListsDAFMF is similar to generateMSPeakListsDA, but uses compounds that were generated by the Find Molecular Features (FMF) algorithm to extract MS peak lists. This is generally much faster than generateMSPeakListsDA, however, it only works when features were obtained using the findFeaturesBruker function. Since all MS spectra are generated in advance by Bruker DataAnalysis, no further parameters exist to customize its operation.

generateMSPeakListsMzR uses the mzR package to extract MS peak lists. For this analyses should be either in .mzXML or .mzML format. This function averages multiple spectra over a chromatgraphic peak to improve accuracy.

The getDefAvgPListParams is used to create a parameter list for peak list averaging (discussed below).

Note

generateMSPeakListsDA requires that the Component column is active (Method-->Parameters-->Layouts-->Mass List Layout) in order to add isotopologue information.

If any errors related to DCOM appear it might be necessary to terminate DataAnalysis (note that DataAnalysis might still be running as a background process). The ProcessCleaner application installed with DataAnalayis can be used for this.

Peak list averaging parameters

The parameters set used for averaging peak lists are set by the avgFeatParams and avgFGroupParams arguments. This should be a named list with the following values:

clusterMzWindow m/z window (in Da) used for clustering m/z values when spectra are averaged. For method="hclust" this corresponds to the cluster height, while for method="distance" this value is used to find nearby masses (+/- window). Too small windows will prevent clustering m/z values (thus erroneously treating equal masses along spectra as different), whereas too big windows may cluster unrelated m/z values from different or even the same spectrum together.

topMost Only retain this maximum number of MS peaks when generating averaged spectra. Lowering this number may exclude more irrelevant (noisy) MS peaks and decrease processing time, whereas higher values may avoid excluding lower intense MS peaks that may still be of interest.

minIntensityPre MS peaks with intensities below this value will be removed (applied prior to selection by topMost) before averaging.

minIntensityPost MS peaks with intensities below this value will be removed after averaging.

avgFun Function that is used to calculate average m/z values.

method Method used for producing averaged MS spectra. Valid values are "hclust", used for hierarchical clustering (using the fastcluster package), and "distance", to use the between peak distance. The latter method may reduces processing time and memory requirements, at the potential cost of reduced accuracy.

pruneMissingPrecursorMS For MS data only: if TRUE then peak lists without a precursor peak are removed. Note that even when this is set to FALSE, functionality that relies on MS (not MS/MS) peak lists (e.g. formulae calulcation) will still skip calculation if a precursor is not found.

retainPrecursorMSMS For MS/MS data only: if TRUE then always retain the precursor mass peak even if is not amongst the topMost peaks. Note that MS precursor mass peaks are always kept. Furthermore, note that precursor peaks in both MS and MS/MS data may still be removed by intensity thresholds (this is unlike the filter method function).

Note that when Bruker algorithms are used these parameters only control generation of feature groups averaged peak lists: how peak lists for features are generated is controlled by DataAnalysis.

The getDefAvgPListParams function can be used to generate a default parameter list. The current defaults are:

clusterMzWindow=0.005; topMost=50; minIntensityPre=500; minIntensityPost=500; avgFun=mean; method="hclust"; pruneMissingPrecursorMS=TRUE; retainPrecursorMSMS=TRUE

Source

Averaging of mass spectra algorithms used by are based on the msProcess R package (now archived on CRAN).

References

A cross-platform toolkit for mass spectrometry and proteomics Chambers, Matthew C. and Maclean, Brendan and Burke, Robert and Amodei, Dario and Ruderman, Daniel L. and Neumann, Steffen and Gatto, Laurent and Fischer, Bernd and Pratt, Brian and Egertson, Jarrett and Hoff, Katherine and Kessner, Darren and Tasman, Natalie and Shulman, Nicholas and Frewen, Barbara and Baker, Tahmina A. and Brusniak, Mi-Youn and Paulse, Christopher and Creasy, David and Flashner, Lisa and Kani, Kian and Moulding, Chris and Seymour, Sean L. and Nuwaysir, Lydia M. and Lefebvre, Brent and Kuhlmann, Frank and Roark, Joe and Rainer, Paape and Detlev, Suckau and Hemenway, Tina and Huhmer, Andreas and Langridge, James and Connolly, Brian and Chadick, Trey and Holly, Krisztina and Eckels, Josh and Deutsch, Eric W. and Moritz, Robert L. and Katz, Jonathan E. and Agus, David B. and MacCoss, Michael and Tabb, David L. and Mallick, Parag Nat Biotechnol. 2012 NOct;30(10):918-920.

Mol Cell Proteomics. 2010 Aug 17. mzML - a Community Standard for Mass Spectrometry Data. Martens L, Chambers M, Sturm M, Kessner D, Levander F, Shofstahl J, Tang WH, Rompp A, Neumann S, Pizarro AD, Montecchi-Palazzi L, Tasman N, Coleman M, Reisinger F, Souda P, Hermjakob H, Binz PA, Deutsch EW.

Nat Biotechnol. 2004 Nov;22(11):1459-66. A common open representation of mass spectrometry data and its application to proteomics research. Pedrioli PG, Eng JK, Hubley R, Vogelzang M, Deutsch EW, Raught B, Pratt B, Nilsson E, Angeletti RH, Apweiler R, Cheung K, Costello CE, Hermjakob H, Huang S, Julian RK, Kapp E, McComb ME, Oliver SG, Omenn G, Paton NW, Simpson R, Smith R, Taylor CF, Zhu W, Aebersold R.

Mol Syst Biol. 2005;1:2005.0017. Epub 2005 Aug 2. A uniform proteomics MS/MS analysis platform utilizing open XML file formats. Keller A, Eng J, Zhang N, Li XJ, Aebersold R.

Bioinformatics. 2008 Nov 1;24(21):2534-6. Epub 2008 Jul 7. ProteoWizard: open source software for rapid proteomics tools development. Kessner D, Chambers M, Burke R, Agus D, Mallick P.

Daniel Müllner (2013). fastcluster: Fast Hierarchical, Agglomerative Clustering Routines for R and Python. Journal of Statistical Software, 53(9), 1-18. URL http://www.jstatsoft.org/v53/i09/.

See also