Generation of transformation products (TPs)

Functionality to automatically obtain transformation products for a given list of parent compounds.

generateTPs(algorithm, ...)

generateTPsBioTransformer(
  parents,
  type = "env",
  steps = 2,
  extraOpts = NULL,
  skipInvalid = TRUE,
  fpType = "extended",
  fpSimMethod = "tanimoto",
  MP = FALSE
)

generateTPsLibrary(
  parents = NULL,
  TPLibrary = NULL,
  skipInvalid = TRUE,
  matchParentsBy = "InChIKey"
)

# S4 method for featureGroups
generateTPsLogic(fGroups, minMass = 40, adduct = NULL, transformations = NULL)

# S4 method for featureGroupsSet
generateTPsLogic(fGroups, minMass = 40, transformations = NULL)

Arguments

algorithm
A character string describing the algorithm that should be used: "biotransformer", "logic", "library"

...
Any parameters to be passed to the selected TP generation algorithm.

parents
The parents for which transformation products should be obtained. This can be (1) a suspect list (see suspect screening for more information), (2) the resulting output screenSuspects or (3) a compounds annotation object. In the former two cases, the suspect (hits) are used as parents, whereas in the latter case all candidates are used as parents.

For generateTPsLibrary: can be NULL, in this case TPs for all parents in the library are obtained.

type
The type of prediction. Valid values are: "env", "ecbased", "cyp450", "phaseII", "hgut", "superbio", "allHuman". Sets the -b command line option.

steps
The number of steps for the predictions. Sets the -s command line option.

extraOpts
A character with extra command line options passed to the biotransformer.jar tool.

skipInvalid
If set to TRUE then the parents will be skipped (with a warning) for which insufficient information (e.g. SMILES) is available.

fpType
The type of structural fingerprint that should be calculated. See the type argument of the get.fingerprint function of rcdk.

fpSimMethod
The method for calculating similarities (i.e. not dissimilarity!). See the method argument of the fp.sim.matrix function of the fingerprint package.

MP
If TRUE then multiprocessing is enabled. Since BioTransformer supports native parallelization, additional multiprocessing generally doesn't lead to significant reduction in computational times. Furthermore, enabling multiprocessing can lead to very high CPU/RAM usage.

TPLibrary
If NULL, a default PubChem based library is used. Otherwise, TPLibrary should be a data.frame. See section below.

matchParentsBy
A character that specifies how the input parents are matched with the data from the TP library. Valid options are: "InChIKey", "InChIKey1", "InChI", "SMILES".

fGroups
A featureGroups object for which TPs should be calculated.

minMass
A numeric that specifies the minimum mass of calculated TPs. If below this mass it will be removed.

adduct
An adduct object (or something that can be converted to it with as.adduct). Examples: "[M-H]-", "[M+Na]+". If the featureGroups object has adduct annotations then these are used if adducts=NULL.

(sets workflow) The adduct argument is not supported for sets workflows, since the adduct annotations will then always be used.

transformations
A data.frame with transformation reactions to be used for calculating the TPs (see section below). If NULL, a default table from Schollee et al. is used (see references).

Details

Several algorithms are available that obtain TPs for a set of parent compounds. Depending on the algorithm, the parents can be a suspect list, all the candidates in a compounds object or the screening results or all feature groups from a featureGroups object. The resulting objects can (again depending on the used algorithm) then be used for conversion to suspect lists (convertToSuspects), a MetFrag database (convertToMFDB) or further processed to generate parent to TP links from features (generateComponentsTPs).

generateTPs is a generic function that will generate transformation products using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateTPsBioTransformer and generateTPsLogic. While these functions may be called directly, generateTPs provides a generic interface and is therefore usually preferred.

generateTPsBioTransformer uses BioTransformer to predict TPs. An important advantage of this algorithm is that it provides full structural information for TPs. However, to do so, structural information also needs to be present for the parents. Afterwards, structural similarities between the parent and its TPs are calculated, which can be used to filter the results. In order to use this function the .jar command line utility should be installed and specified in the patRoon.path.BioTransformer option. The .jar file can be obtained via https://bitbucket.org/djoumbou/biotransformer/src/master.

generateTPsLibrary obtains transformation products from a library. Similar to generateTPsBioTransformer, this algorithm relies and provides structural information for parents/TPs (SMILES). By default, a library is used that is based on data from PubChem. However, it also possible to use your own library.

generateTPsLogic applies metabolic logic to predict transformation products. With this algorithm, TPs are predicted from common (environmental) chemical reactions, such as hydroxylation, demethylation etc. The generated TPs result from calculating the mass differences between a parent feature after it underwent the reaction. While this only results in little information on chemical properties of the TP, an advantage of this method is that it does not rely on structural information of the parent, which may be unknown in a full non-target analysis.

Note

When the parents argument is a compounds object, the candidate library identifier is used in case the candidate has no defined compoundName.

Custom TP libraries and transformations

The TPLibrary argument is used to specify a custom TP library for generateTPsLibrary. This should be a data.frame where each row specifies a TP for a parent, with the following columns:

parent_name and TP_name: The name of the parent/TP.

parent_SMILES and TP_SMILES The SMILES of the parent structure.

parent_LogP and TP_LogP The log P values for the parent/TP. (optional)

LogPDiff The difference between parent and TP Log P values. Ignored if both parent_LogP and TP_LogP are specified. (optional)

Other columns are allowed, and will be included in the final object. Multiple TPs for a single parent are specified by repeating the value within parent_ columns.

Similarly, the transformations argument to generateTPsLogic is used to specify custom rules to calculate transformation products. This should be a data.frame with the following columns:

transformation The name of the chemical transformation

add The elements that are added by this reaction (e.g. "O").

sub The elements that are removed by this reaction (e.g. "H2O").

retDir The expected retention time direction relative to the parent (assuming a reversed phase like LC separation). Valid values are: -1 (elutes before the parent), 1 (elutes after the parent) or 0 (no significant change or unknown).

Parallelization

generateTPsBioTransformer uses multiprocessing to parallelize computations. Please see the parallelization section in the handbook for more details and patRoon options for configuration options.

Source

The algorithm of generateTPsLogic is directly based on the work done by Schollee et al. (see references).

References

Djoumbou-Feunang Y, Fiamoncini J, Gil-de-la-Fuente A, Greiner R, Manach C, Wishart DS (2019). “BioTransformer: a comprehensive computational tool for small molecule metabolism prediction and metabolite identification.” Journal of Cheminformatics, 11(1). doi: 10.1186/s13321-018-0324-5 .

Wicker J, Lorsbach T, Gutlein M, Schmid E, Latino D, Kramer S, Fenner K (2015). “enviPath - The environmental contaminant biotransformation pathway resource.” Nucleic Acids Research, 44(D1), D502--D508. doi: 10.1093/nar/gkv1229 .

Guha, R. (2007). 'Chemical Informatics Functionality in R'. Journal of Statistical Software 6(18)

Schollee JE, Schymanski EL, Avak SE, Loos M, Hollender J (2015). “Prioritizing Unknown Transformation Products from Biologically-Treated Wastewater Using High-Resolution Mass Spectrometry, Multivariate Statistics, and Metabolic Logic.” Analytical Chemistry, 87(24), 12121--12129. doi: 10.1021/acs.analchem.5b02905 .

See also