Hierarchical clustering information

Objects from this class hold information required for hierarchical clustering.

# S4 method for clusterInfo
length(x)

# S4 method for clusterInfo
show(object)

# S4 method for clusterInfo
clusterProperties(obj)

Arguments

x, object, cInfo
A clusterInfo object.

Value

clusterProperties returns a list

Details

Objects from this class are returned by makeHCluster.

Methods (by generic)

length: length returns the total number of clusters.

show: show prints general information about this object.

clusterProperties: returns a list with data and properties relevant to the clustering that was performed (see Slots section for a description).

Slots

clusterm

Numeric matrix with normalized feature group intensities that was used for clustering.

distm

Distance matrix that was used for clustering (obtained with daisy).

clust

Object returned by hclust.

metric

Distance metric that was used (e.g. euclidean).

method

Clustering method that was used.

average

If TRUE then feature groups from the same replicate groups were averaged prior to clustering.