Grouping feature groups in components

Functionality to automatically group related feature groups (e.g. isotopes, adducts and homologues) to assist and simplify compound annotation.

# S4 method for featureGroups
generateComponents(fGroups, algorithm, ...)

generateComponentsCAMERA(fGroups, ionization, onlyIsotopes = FALSE,
  extraOpts = NULL)

generateComponentsNontarget(fGroups, ionization, rtRange = c(-120, 120),
  mzRange = c(5, 120), elements = c("C", "H", "O"), maxRTDev = 30,
  maxMzDev = 0.002, extraOpts = NULL)

generateComponentsRAMClustR(fGroups, st = NULL, sr = NULL, maxt = 12,
  hmax = 0.3, normalize = "TIC", ionization, absMzDev = 0.002,
  relMzDev = 5, RCExperimentVals = list(design = list(platform = "LC-MS"),
  instrument = list(ionization = ionization, MSlevs = 1)), extraOptsRC = NULL,
  extraOptsFM = NULL)

Arguments

fGroups
featureGroups object for which components should be generated.

algorithm
A character string describing the algorithm that should be used: "ramclustr", "camera", "nontarget"

...
Any parameters to be passed to the selected component generation algorithm.

ionization
Which ionization polarity was used to generate the data: should be "positive" or "negative".

onlyIsotopes
Logical value. If TRUE only isotopes are considered when generating components (faster). Corresponds to quick argument of CAMERA::annotate.

extraOpts
Named character vector with extra arguments directly passed to homol.search (generateComponentsNontarget) or CAMERA::annotate (generateComponentsCAMERA). Set to NULL to ignore.

rtRange
A numeric vector containing the minimum and maximum retention time (in seconds) between homologues. Series are always considered from low to high m/z, thus, a negative minimum retention time allows detection of homologous series with increasing m/z and decreasing retention times. These values set the minrt and maxrt arguments of homol.search.

mzRange
A numeric vector specifying the minimum and maximum m/z increment of a homologous series. Sets the minmz and maxmz arguments of homol.search.

elements
A character vector with elements to be considered for detection of repeating units. Sets the elements argument of homol.search function.

maxRTDev, maxMzDev
Maximum retention and (absolute) m/z deviation for detection of homologues within series. These arguments set the rttol and mztol arguments of homol.search.

st, sr, maxt, hmax, normalize
Arguments to tune the behaviour of feature group clustering. See their documentation from ramclustR. When st is NULL it will be automatically calculated as the half of the median for all chromatographic peak widths.

absMzDev, relMzDev
Maximum absolute/relative m/z deviation for annotation. Passed as the mzabs.error and ppm.error arguments to do.findmain.

RCExperimentVals
A named list containing two more lists: design and instrument. These are used to construct the ExpDes argument passed to ramclustR.

extraOptsRC, extraOptsFM
Named list with further arguments to be passed to ramclustR and do.findmain. Set to NULL to ignore.

Value

A components object containing all generated components.

Details

Several algorithms are provided to group feature groups that are related in some (chemical) way to each other. These components generally include adducts, isotopes, in-source fragments and homologues. The linking of this data is generally useful to provide more information for compound annotation and reduce the data size and thus complexity. generateComponents is a generic function that will generate components using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateComponentsRAMClustR and generateComponentsNontarget. While these functions may be called directly, generateComponents provides a generic interface and is therefore usually preferred. generateComponentsCAMERA provides an interface to CAMERA which is used to generate components from known adducts, isotopes and in-source fragments. The specified featureGroups object is automatically converted to an xcmsSet object using getXcmsSet. generateComponentsNontarget uses the nontarget R package to generate components by unsupervised detection of homologous series. In the first step the homol.search function is used to detect all homologues within each replicate group (analyses within each replicate group are averaged prior to detection). Then, homologous series across replicate groups are merged in case of full overlap or when merging of partial overlapping series causes no conflicts. generateComponentsRAMClustR uses RAMClustR to generate components from feature groups which follow similar chromatographic retention profiles, but are not necessarily restricted to known rules (e.g. adducts or isotopes). This method uses the ramclustR functions for generating the components, whereas do.findmain is used for annotation.

References

Kuhl, C., Tautenhahn, R., Boettcher, C., Larson, T. R. and Neumann, S. CAMERA: an integrated strategy for compound spectra extraction and annotation of liquid chromatography/mass spectrometry data sets. Analytical Chemistry, 84:283-289 (2012) Loos, M., Singer, H. (2017). Nontargeted homologue series extraction from hyphenated high resolution mass spectrometry, Journal of Cheminformatics 9(12), . Loos, M., Gerber, C., Corona, F., Hollender, J., Singer, H. (2015). Accelerated isotope fine structure calculation using pruned transition trees, Analytical Chemistry 87(11), 5738-5744. C. D. Broeckling, F. A. Afsar, S. Neumann , A. Ben-Hur, and J. E. Prenni: RAMClust: A Novel Feature Clustering Method Enables Spectral-Matching-Based Annotation for Metabolomics Data XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 2014, 86 (14), pp 6812–6817 Broeckling, C. D.; Heuberger, A. L.; Prince, J. A.; Ingelsson, E.; Prenni, J. E. Assigning precursor–product ion relationships in indiscriminant MS/MS data from non-targeted metabolite profiling studies. Metabolomics 2013, 9, 33-43 (DOI: 10.1007/s11306-012-0426-4).