Components class for network-based componentization.

This class is derived from components and is used to store results from network-based componentization.

Usage

# S4 method for class 'componentsNet'
plotGraph(obj, analysis, width = NULL, height = NULL)

# S4 method for class 'componentsNetSet'
plotGraph(obj, analysis, set, ...)

# S4 method for class 'componentsNetSet'
unset(obj, set)

Arguments

obj

The componentsNet object to plot.

analysis

The name of the analysis to plot.

width, height

Passed to visNetwork.

set

(sets workflow) The name of the set.

...

(sets workflow) Further arguments passed to the non-sets workflow method.

Value

plotGraph returns the result of visNetwork.

Details

Objects from this class are generated by generateComponentsNet.

Methods (by generic)

plotGraph(componentsNet): Plots an interactive network graph for the feature components of an analysis.

Slots

featureComponents

A list with feature components for each analysis.

featureGraphs

A list with feature network graphs (igraph) for each analysis.

annotationObjects

A list with annotation objects for each component (from nontarget or InterpretMSSpectrum).

Sets workflows

The componentsNetSet class is applicable for sets workflows. This class is derived from componentsNet and therefore largely follows the same user interface.

The following methods are specifically defined for sets workflows:

unset Converts the object data for a specified set into a 'non-set' object (componentsNetUnset), which allows it to be used in 'regular' workflows. Only the components in the specified set are kept. Furthermore, the component names are restored to non-set specific names (see generateComponents for more details).

The following methods are changed or with new functionality:

plotGraph Currently can only create graph networks from one set (specified by the set argument).

Note that the componentsNetSet class does not have featureComponents and featureGraphs slots. Instead, the setObjects method can be used to access this data for a specific set.

References

Antonov M, Csárdi G, Horvát S, Müller K, Nepusz T, Noom D, Salmon M, Traag V, Welles BF, Zanini F (2023). “igraph enables fast and robust network analysis across programming languages.” arXiv preprint arXiv:2311.10260. doi:10.48550/arXiv.2311.10260 .

Csárdi G, Nepusz T (2006). “The igraph software package for complex network research.” InterJournal, Complex Systems, 1695. https://igraph.org.

Csárdi G, Nepusz T, Traag V, Horvát S, Zanini F, Noom D, Müller K, Schoch D, Salmon M (2026). igraph: Network Analysis and Visualization in R. doi:10.5281/zenodo.7682609 . R package version 2.3.3, https://CRAN.R-project.org/package=igraph.

See also