Automatic compound identification

Functionality to automatically identify chemical compounds from feature groups.

# S4 method for featureGroups
generateCompounds(fGroups, MSPeakLists, algorithm,
  ...)

generateCompoundsMetfrag(fGroups, MSPeakLists, method = "CL",
  logPath = file.path("log", "metfrag"), timeout = 300,
  timeoutRetries = 2, topMost = 100, dbRelMzDev = 5, fragRelMzDev = 5,
  fragAbsMzDev = 0.002, isPositive, adduct, database = "PubChem",
  chemSpiderToken = "", scoreTypes = c("FragmenterScore",
  "OfflineMetFusionScore"), scoreWeights = 1,
  preProcessingFilters = c("UnconnectedCompoundFilter", "IsotopeFilter"),
  postProcessingFilters = c("InChIKeyFilter"), maxCandidatesToStop = 2500,
  addTrivialNames = TRUE, identifiers = NULL, extraOpts = NULL,
  maxProcAmount = getOption("patRoon.maxProcAmount"))

generateCompoundsSirius(fGroups, MSPeakLists, maxMzDev = 5,
  adduct = "[M+H]+", elements = "CHNOP", profile = "qtof",
  formulaDatabase = NULL, fingerIDDatabase = "pubchem", noise = NULL,
  topMost = 100, logPath = file.path("log", "sirius"),
  maxProcAmount = getOption("patRoon.maxProcAmount"))

Arguments

fGroups
featureGroups object for which compounds should be identified. This should be the same or a subset of the object that was used to create the specified MSPeakLists. In the case of a subset only the remaining feature groups in the subset are considered.

MSPeakLists
A MSPeakLists object that was generated for the supplied fGroups.

algorithm
A character string describing the algorithm that should be used: "metfrag", "sirius"

...
Any parameters to be passed to the selected compound generation algorithm.

method
Which method should be used for MetFrag execution: "CL" for MetFragCL and "R" for MetFragR. The former might be faster.

logPath
Destination directory for log files with output from executed commands. Will be created if non-existant. Set to NULL to disable logging.

timeout
Maximum time (in seconds) before a metFrag query for a feature group is stopped. Also see timeoutRetries argument.

timeoutRetries
Maximum number of retries after reaching a timeout before completely skipping the metFrag query for a feature group. Also see timeout argument.

topMost
Only keep this number of candidates (per feature group) with highest score. Set to NULL to always keep all candidates, however, please note that this may result in significant usage of CPU/RAM resources for large numbers of candidates.

dbRelMzDev
Relative mass deviation (in ppm) for database search. Sets the DatabaseSearchRelativeMassDeviation option.

fragRelMzDev
Relative mass deviation (in ppm) for fragment matching. Sets the FragmentPeakMatchRelativeMassDeviation option.

fragAbsMzDev
Absolute mass deviation (in Da) for fragment matching. Sets the FragmentPeakMatchAbsoluteMassDeviation option.

isPositive
Set to TRUE for data measured with positive ionization. Sets the IsPositiveIonMode option.

adduct
The assumed MS adduct. For generateCompoundsMetfrag: a numeric value (e.g. 1 for [M+H]+, -1 for [M-H]- and 0 for [M]+/-, see the MetFrag CL homepage for more). Please note that isPositive also needs to be set. For generateCompoundsSirius, a character string such as "[M+H]+", "[M-H]-", "[M+Na]+" and "[M]+". Sets the PrecursorIonMode and --ion options, respectively.

database
Compound database to use. Valid values are: "PubChem", "ExtendedPubChem", "ChemSpider" and "KEGG". The "ExtendedPubChem" is a PubChem database which includes number of patents and references information, which an be used for further scoring (see scoreTypes parameter.) In order to use the ChemSpider database the chemSpiderToken should be set. Sets the MetFragDatabaseType option.

chemSpiderToken
A character string with the ChemSpider security token that should be set when the ChemSpider database is used. Sets the ChemSpiderToken option.

scoreTypes
A character vector defining the scoring methods (e.g. "FragmenterScore", "OfflineMetFusionScore", "RetentionTimeScore"). Some methods require further options to be set. Additional scoring methods become available when the "ExtendedPubChem" (i.e. "PubChemNumberPatents" and "PubChemNumberPubMedReferences") or "ChemSpider" (i.e. "ChemSpiderReferenceCount", "ChemSpiderNumberExternalReferences", "ChemSpiderRSCCount", "ChemSpiderNumberPubMedReferences" and "ChemSpiderDataSourceCount") database type is used. For all scoring types and more information refer to the Candidate Scores section on the MetFragR homepage. Sets the MetFragScoreTypes option.

scoreWeights
Numeric vector containing weights of the used scoring types. Order is the same as set in scoreTypes. Values are recycled if necessary. Sets the MetFragScoreWeights option.

preProcessingFilters, postProcessingFilters
A character vector defining pre/post filters applied before/after fragmentation and scoring (e.g. "UnconnectedCompoundFilter", "IsotopeFilter", "ElementExclusionFilter"). Some methods require further options to be set. For all filters and more information refer to the Candidate Filters section on the MetFragR homepage. Sets the MetFragPreProcessingCandidateFilter and MetFragPostProcessingCandidateFilter options.

maxCandidatesToStop
If more than this number of candidate structures are found then processing will be aborted and no results this feature group will be reported. Low values increase the chance of missing data, whereas too high values will use too much computer resources and signficantly slowdown the process. Sets the MaxCandidateLimitToStop option.

addTrivialNames
If TRUE and the PubChem database is used then trivial names will be added after compound search.

identifiers
A list containing for each feature group a character vector with database identifiers that should be used to find candidates for a feature group (the list should be named by feature group names). If NULL all relevant candidates will be retrieved from the specified database. An example usage scenario is to obtain the list of candidate identifiers from a compounds object obtained with generateCompoundsSirius using the identifiers method. This way, only those candidates will be searched by MetFrag that were generated by SIRIUS+CSI:FingerID. Sets the PrecursorCompoundIDs option.

extraOpts
A named list containing further settings to be passed to run.metfrag. See the MetFragR and MetFrag CL homepages for all available options.

maxProcAmount
Maximum number of processes to run for parallelization. Usually a number close to the amount of physical cores yields most efficient results.

maxMzDev
Maximum deviation between the measured and candidate formula m/z values (in ppm). Sets the --ppm-max commandline option.

elements
Elements to be considered for formulae calculation. This will heavily affects the number of candidates! Always try to work with a minimal set by excluding elements you don't expect. The minimum/maximum number of elements can also be specified, for example: a value of "C[5]H[10-15]O" will only consider formulae with up to five carbon atoms, between ten and fifteen hydrogen atoms and any amount of oxygen atoms. Sets the --elements commandline option.

profile
Name of the configuration profile, for example: "qtof", "orbitrap", "fticr". Sets the --profile commandline option.

formulaDatabase
If not NULL, use a database for retrieval of formula candidates. Possible values are: "pubchem", "bio", "kegg", "hmdb". Sets the --database commandline option.

fingerIDDatabase
Database specifically used for CSI:FingerID. If NULL, the value of the formulaDatabase parameter will be used or "pubchem" when that is also NULL. Sets the --fingerid-db option.

noise
Median intensity of the noise (NULL ignores this parameter). Sets the --noise commandline option.

Value

A compounds object containing all tentatively identified compounds.

Details

Several algorithms are provided to automatically identify compounds for given feature groups. To this end, each measured masses for all feature groups are searched within online database(s) (e.g. PubChem) to retrieve a list of potential candidate chemical compounds. Depending on the algorithm and its parameters, further scoring of candidates is then performed using, for instance, matching of measured and theoretical isotopic patterns, presence within other data sources such as patent databases and similarity of measured and in-silico predicted MS/MS fragments. Note that this process is often quite time consuming, especially for large feature group sets. Therefore, this is often one of the last steps within the workflow and not performed before feature groups have been prioritized. generateCompounds is a generic function that will generate compounds using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateCompoundsMetfrag and generateCompoundsSirius. While these functions may be called directly, generateCompounds provides a generic interface and is therefore usually preferred. generateCompoundsMetfrag uses the metfRag package for compound identification (see http://c-ruttkies.github.io/MetFrag/). Several online compound databases such as PubChem and ChemSpider may be chosen for retrieval of candidate structures. In addition, many options exist to score and filter resulting data, and it is highly suggested to optimize these to improve results. While MS/MS data is not mandatory, it will usually greatly improve candidate scoring. The MetFrag options PeakList, IonizedPrecursorMass and ExperimentalRetentionTimeValue (in minutes) fields are automatically set from feature data. generateCompoundsSirius uses SIRIUS in combination with CSI:FingerID for compound identification. Similar to generateFormulasSirius, candidate formulae are generated with SIRIUS. These results are then feed to CSI:FingerID to acquire candidate structures. Candidates are then scored on similarity of measured and in-silico predicted MS/MS fragments and if the PubChem database is used, on the number of references in patents. This method requires the availability of MS/MS data, and feature groups without it will be ignored.

Note

At this point only mass spectrometric data from the most intense feature is used for compound identification.

References

Ruttkies C, Schymanski EL, Wolf S, Hollender J and Neumann S (2016). “MetFrag relaunched: incorporating strategies beyond in silico fragmentation.” Journal of Cheminformatics, 8(1). doi: 10.1186/s13321-016-0115-9 . Dührkop K and Böcker S (2015). “Fragmentation Trees Reloaded.” In Przytycka TM (ed.), Research in Computational Molecular Biology, pp. 65--79. ISBN 978-3-319-16706-0. Dührkop K, Shen H, Meusel M, Rousu J and Böcker S (2015). “Searching molecular structure databases with tandem mass spectra using CSI:FingerID.” Proceedings of the National Academy of Sciences, 112(41), pp. 12580--12585. doi: 10.1073/pnas.1509788112 . Böcker S, Letzel MC, Lipták Z and Pervukhin A (2008). “SIRIUS: decomposing isotope patterns for metabolite identification.” Bioinformatics, 25(2), pp. 218--224. doi: 10.1093/bioinformatics/btn603 .

See also

compounds-class