Compounds list class for MetFrag results.

This class is derived from compounds and contains additional specific MetFrag data.

Usage

settings(compoundsMF)

# S4 method for class 'compoundsMF'
settings(compoundsMF)

Arguments

compoundsMF

A compoundsMF object.

Details

Objects from this class are generated by generateCompoundsMetFrag

Methods (by generic)

settings(compoundsMF): Accessor method for the settings slot.

Slots

settings

A list with all general configuration settings passed to MetFrag. Feature specific items (e.g. spectra and precursor masses) are not contained in this list.

S4 class hierarchy

compounds

compoundsMF

References

Ruttkies C, Schymanski EL, Wolf S, Hollender J, Neumann S (2016). “MetFrag relaunched: incorporating strategies beyond in silico fragmentation.” Journal of Cheminformatics, 8(1). doi:10.1186/s13321-016-0115-9 .

See also