Conversion of analyses between several open formats

This function uses the FileConverter command from OpenMS to convert analyses files from one open format to another. The supported file formats are (both directions): .mzXML, .mzML and .mzData.

convertMSFiles(anaInfo, formatFrom, formatTo, outPath = anaInfo$path,
  overWrite = FALSE)

Arguments

anaInfo
A table with analysis information that contains the analyses to be converted.

formatFrom, formatTo
A character that specifies the source and destination format, respectively. Valid options are: "mzXML", "mzML" and "mzData".

outPath
A directory path that should be used for the output. Usually this is the same as the source (otherwise the analysis information should be changed for further processing).

overWrite
Should existing destination file be overwritten (TRUE) or not (FALSE)?

References

Rost HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmstrom L, Aebersold R, Reinert K and Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), pp. 741--748. doi: 10.1038/nmeth.3959 .