Finding features

Functions and classes for collection of features.

findFeatures(analysisInfo, algorithm, ...)

findFeaturesBruker(analysisInfo, doFMF = "auto", startRange = 0,
  endRange = 0)

findFeaturesEnviPick(analysisInfo, ...)

importFeaturesEnviMass(analysisInfo, enviProjPath)

findFeaturesOpenMS(analysisInfo, thr = 1000, comfwhm = 5, minfwhm = 3,
  maxfwhm = 60, minlength = 3, maxlength = -1, mzppm = 10,
  extraOpts = NULL, logPath = file.path("log", "openms"),
  maxProcAmount = getOption("patRoon.maxProcAmount"))

findFeaturesXCMS(analysisInfo, method = "centWave", ...)

importFeaturesXCMS(xs, analysisInfo)

Arguments

analysisInfo
Analysis info table.

algorithm
A character string describing the algorithm that should be used: "bruker", "openms", "xcms", "envipick"

...
further parameters passed to xcmsSet (findFeaturesXCMS), enviPickwrap (featurefinderEnviPick) or to selected feature finding algorithms (findFeatures).

doFMF
Run the 'Find Molecular Features' algorithm before loading compounds. Valid options are: TRUE (always), FALSE (never) and "auto" (automatically if no compounds present).

startRange, endRange
Start/End retention range (seconds) from which to collect features. A 0 (zero) for codeendRange marks the end of the analysis.

enviProjPath
The path of the enviMass project.

thr
Noise intensity threshold. Sets algorithm:common:noise_threshold_int option.

comfwhm, minfwhm, maxfwhm
Expected/Minimal/Maximal chromatographic peak width (seconds). Sets algorithm:common:chrom_fwhm, algorithm:epd:min_fwhm and algorithm:epd:max_fwhm, respectively.

minlength, maxlength
Minimum/Maximum length of mass trace (seconds). Set negative value for maxlength to disable maximum. Sets algorithm:mtd:min_trace_length and algorithm:mtd:min_trace_length, respectively.

mzppm
Allowed mass deviation (ppm) for trace detection. Sets algorithm:mtd:mass_error_ppm.

extraOpts
Named character vector containing extra options that will be passed to FeatureFinderMetabo. Any options specified here will override any of the above.

logPath
Destination directory for log files with output from executed commands. Will be created if non-existant. Set to NULL to disable logging.

maxProcAmount
Maximum number of processes to run for parallelization. Usually a number close to the amount of physical cores yields most efficient results.

method
The method setting used by XCMS peak finding, see xcms::findPeaks

xs
An xcmsSet object.

Value

An object of a class which is derived from features.

Details

Several functions exist to collect features (i.e. retention and MS information that represent potential compounds) from a set of analyses. All 'feature finders' return an object derived from the features base class. The next step in a general workflow is to group and align these features across analyses by feature groupers. Note that some feature finders have a plethora of options which sometimes may have a large effect on the quality of results. Fine-tuning parameters is therefore important, and the optimum is largely dependent upon applied analysis methodology and instrumentation.

findFeatures is a generic function that will find features using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as findFeaturesOpenMS and findFeaturesBruker. While these functions may be called directly, findFeatures provides a generic interface and is therefore usually preferred.

findFeaturesBruker uses the 'Find Molecular Features' (FMF) algorithm of Bruker DataAnalysis vendor software to find features. The resulting 'compounds' are then transferred from DataAnalysis and stored as features.

findFeaturesEnviPick uses the enviPickwrap. function from the enviPick R package to extract features.

importFeaturesEnviMass imports features from a project generated by the enviMass package.

findFeaturesOpenMS uses the FeatureFinderMetabo TOPP tool (see http://www.openms.de).

findFeaturesXCMS uses the xcmsSet function from the xcms package to find features.

importFeaturesXCMS converts features from an existing xcmsSet object (obtained with the xcms package) to a new features object.

Note

findFeaturesBruker only works with Bruker data files (.d extension) and requires Bruker DataAnalysis and the RDCOMClient package to be installed.

findFeaturesEnviPick Requires analysis files to be in the mzXML format.

The file format of analyses for findFeaturesOpenMS must be mzML. This functionality has been tested with OpenMS version >= 2.0. Please make sure it is installed and its binaries are added to the PATH environment variable or the patRoon.path.OpenMS option is set.

The file format of analyses for findFeaturesXCMS must be mzML or mzXML.

References

Rost HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmstrom L, Aebersold R, Reinert K and Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), pp. 741--748. doi: 10.1038/nmeth.3959 .

Smith, C.A. and Want, E.J. and O'Maille, G. and Abagyan,R. and Siuzdak, G.: XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 78:779-787 (2006) Ralf Tautenhahn, Christoph Boettcher, Steffen Neumann: Highly sensitive feature detection for high resolution LC/MS BMC Bioinformatics, 9:504 (2008) H. Paul Benton, Elizabeth J. Want and Timothy M. D. Ebbels Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data Bioinformatics, 26:2488 (2010)

See also

features-class and analysis-information