Automatic chemical formula generation

Functionality to automatically calculate chemical formulae from feature groups and generate a consensus over all analyses.

# S4 method for featureGroups
generateFormulas(fGroups, algorithm, ...)

generateFormulasDA(fGroups, precursorMzSearchWindow = 0.002,
  MSMode = "both")

generateFormulasGenForm(fGroups, MSPeakLists, maxMzDev = 5, adduct = "M+H",
  elements = "CHNOP", hetero = TRUE, MSMode = "both", extraOpts = NULL,
  maxProcAmount = getOption("patRoon.maxProcAmount"), maxCmdsPerProc = 25)

generateFormulasSirius(fGroups, MSPeakLists, maxMzDev = 5,
  adduct = "[M+H]+", elements = "CHNOP", profile = "qtof",
  database = NULL, noise = NULL, logPath = file.path("log", "sirius"),
  maxProcAmount = getOption("patRoon.maxProcAmount"))

Arguments

fGroups
featureGroups object for which formulae should be generated. This should be the same or a subset of the object that was used to create the specified MSPeakLists (only relevant for algorithms using MSPeakLists). In the case of a subset only the remaining feature groups in the subset are considered.

algorithm
A character string describing the algorithm that should be used: "bruker", "genform", "sirius"

...
Any parameters to be passed to the selected formula generation algorithm.

precursorMzSearchWindow
Maximum m/z deviation (Da) to find back feature data of precursor/parent ions from MS/MS spectra (this data is not readily available from SmartFormula3D results).

MSMode
Whether formulae should be generated only from MS data ("ms"), MS/MS data ("msms") or both ("both"). Using the latter option, unique formulae from MS data will still be reported if not predicted from MS/MS data. Formulae calculated from MS data that were also generated from MS/MS data will be removed.

MSPeakLists
An MSPeakLists object that was generated for the supplied fGroups.

maxMzDev
Maximum deviation between the measured and candidate formula m/z values (in ppm). Sets the ppm and --ppm-max commandline options for GenForm and SIRIUS, respectively.

adduct
Sets the adduct form of measured ions. For generateFormulasGenForm the format is without brackets and charges (based on this website), examples: "M+H", "M-H", "M+Na", "M-e". For generateFormulasSirius the format is with brackets and charge, for example: "[M+H]+", "[M-H]-", "[M+Na]+", "[M]+". Sets the ion and --ion commandline options for GenForm and SIRIUS, respectively.

elements
Elements to be considered for formulae calculation. This will heavily affects the number of candidates! Always try to work with a minimal set by excluding elements you don't expect. For generateFormulasSirius, the minimum/maximum number of elements can also be specified, for example: a value of "C[5]H[10-15]O" will only consider formulae with up to five carbon atoms, between ten and fifteen hydrogen atoms and any amount of oxygen atoms. Sets the el and --elements commandline options for GenForm and SIRIUS, respectively.

hetero
Only consider formulae with at least one hetero atom. Sets the het commandline option.

extraOpts
An optional character vector with any other commandline options that will be passed to GenForm. See the GenForm options section for all available commandline options.

maxProcAmount
Maximum number of processes to run for parallelization. Usually a number close to the amount of physical cores yields most efficient results.

maxCmdsPerProc
Maximum number of commands that should be combined for each executed process. Combining commands with short runtimes (such as GenForm) can significantly increase parallel performance.

profile
Name of the configuration profile, for example: "qtof", "orbitrap", "fticr". Sets the --profile commandline option.

database
If not NULL, use a database for retrieval of formula candidates. Possible values are: "pubchem", "bio", "kegg", "hmdb". Sets the --database commandline option.

noise
Median intensity of the noise (NULL ignores this parameter). Sets the --noise commandline option.

logPath
Destination directory for log files with output from executed commands. Will be created if non-existant. Set to NULL to disable logging.

Value

A formulas object containing all generated formulae.

Details

Several algorithms are provided to automatically generate formulae for given feature groups. All tools use the accurate mass of a feature to back-calculate candidate formulae. Depending on the algorithm and data availability, other data such as isotopic pattern and MS/MS fragments may be used to further improve formula assignment.

generateFormulas is a generic function that will generate formulae using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateFormulasDA and generateFormulasGenForm. While these functions may be called directly, generateFormulas provides a generic interface and is therefore usually preferred.

generateFormulasDA uses Bruker DataAnalysis to generate chemical formulae. This method supports scoring based on overlap between measured and theoretical isotopic patterns (both MS and MS/MS data) and the presence of 'fitting' MS/MS fragments. The method will iterate through all features (or "Compounds" in DataAnalysis terms) and call SmartFormula (and SmartFormula3D if MS/MS data is available) to generate all formulae. Parameters affecting formula calculation have to be set in advance within the DataAnalysis method for each analysis (e.g. by setDAMethod). This method requires that features were obtained with findFeaturesBruker. Unlike other algorithms, there is no prior need to generate MS peak lists.

generateFormulasGenForm uses GenForm to generate chemical formulae. When MS/MS data is available it will be used to score candidate formulae by presence of 'fitting' fragments.

generateFormulasSirius uses SIRIUS to generate chemical formulae. Similarity of measured and theoretical isotopic patterns will be used for scoring candidates. Note that SIRIUS requires availability of MS/MS data.

Note

generateFormulasGenForm always sets the exist and oei GenForm commandline options.

GenForm options

Below is a list of options (generated by running GenForm without commandline options) which can be set by the extraOpts parameter.

References

Meringer M, Reinker S, Zhang J, Muller A (2011). “MS/MS Data Improves Automated Determination of Molecular Formulas by Mass Spectrometry.” MATCH Commun. Math. Comput. Chem., 65(2), 259--290.

Duhrkop K, Bocker S (2015). “Fragmentation Trees Reloaded.” In Przytycka TM (ed.), Research in Computational Molecular Biology, 65--79. ISBN 978-3-319-16706-0.

Bocker S, Letzel MC, Liptak Z, Pervukhin A (2008). “SIRIUS: decomposing isotope patterns for metabolite identification.” Bioinformatics, 25(2), 218--224. doi: 10.1093/bioinformatics/btn603 .

See also