Formula consensus class

Contains all formulae for given feature groups after a consensus was made.

# S4 method for formulaConsensus
formulaTable(obj)

# S4 method for formulaConsensus
algorithm(obj)

# S4 method for formulaConsensus
length(x)

# S4 method for formulaConsensus
show(object)

# S4 method for formulaConsensus
plotSpec(obj, precursor, groupName, MSPeakLists,
  useGGPlot2 = FALSE)

Arguments

obj, x, object
The formulaConsensus object.

precursor
The formula of the precursor (in ionic form, i.e. as detected by the MS).

groupName
The name of the feature group to which the candidate belongs.

MSPeakLists
The MSPeakLists object that was used to generate the candidate

Value

formulaTable returns a data.table containing all generated formulae, scoring information, MS/MS fragments, etc. plotSpec will return a ggplot object if useGGPlot2 is TRUE.

Details

Objects of formulaConsensus are generated by the consensus method.

Methods (by generic)

formulaTable: Accessor method to obtain generated formulae.

algorithm: Accessor method for the algorithm (a character string) used to generate formulae.

length: Obtain total number of formulae entries.

show: Show summary information for this object.

plotSpec: Plots an annotated spectrum for a given candidate formula of a feature group.

Slots

formulas

A data.table containing an overview of all consensus formulae. Use the formulaTable method for access.

algorithm

The algorithm that was used for generation of formulae. Use the algorithm method for access.