Generation of transformation products (TPs)

Functionality to automatically obtain transformation products for a given set of parent compounds.

Usage

generateTPs(algorithm, ...)

Arguments

algorithm

A character string describing the algorithm that should be used: "biotransformer", "logic", "library", "library_formula", "cts", "ann_form", "ann_comp"

...

Any parameters to be passed to the selected TP generation algorithm.

Value

A transformationProducts (derived) object containing all generated TPs.

Details

generateTPs is a generic function that will generate transformation products by one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateTPsBioTransformer and generateTPsLogic. While these functions may be called directly, generateTPs provides a generic interface and is therefore usually preferred.

See also