Hierarchical clustering of feature groups

Utilities to perform and process hierarchical clustering with feature groups.

# S4 method for featureGroups
makeHCluster(fGroups, normFunc = max,
  metric = "euclidean", method = "ward.D2", average = TRUE)

# S4 method for featureGroups,clusterInfo
hClusterFilter(fGroups, cInfo, k, c)

# S4 method for clusterInfo
drawHeatMap(cInfo, col = colorRampPalette(c("black",
  "yellow"))(100), interactive = FALSE, ...)

# S4 method for clusterInfo
getSilhouetteInfo(cInfo, ranges = seq(10, 100, 10))

# S4 method for clusterInfo
plotInt(obj, k, c, ...)

# S4 method for clusterInfo,ANY
plot(x, k = NULL, ...)

# S4 method for silhouetteInfo,ANY
plot(x, y, ...)

Arguments

fGroups
A featureGroups object.

normFunc
Function that should be used for normalization of data. Intensitity values of a feature group will be divided by the result of this function when it is called with all intensity values of that feature group. For example, when max is used normalized intensities will be between zero and one.

metric
Distance metric used to calculate the distance matrix (passed to daisy).

method
Clustering method that should be applied (passed to hclust).

average
If TRUE then all intensitity data will be averaged for each replicate group.

cInfo
A clusterInfo object with relevant clustering information.

k
The desired number of branches from a given cluster object.

c
The numeric identifier of the desired branch from a cut cluster.

col
Colours used for drawing the heatmap. See heatmap / d3heatmap.

interactive
If TRUE an interactive heatmap will be drawn (with d3heatmap).

…
Further options passed to heatmap / d3heatmap (drawHeatMap), barplot (plot, silhouetteInfo signature) or plot (plot, clusterInfo signature and plotInt).

ranges
An integer vector containing a sequence of desired amounts of branches to which silhouette information should be calculated.

x, obj
A clusterInfo/silhouetteInfo object.

y
Ignored.

Value

makeHCluster Returns an object of the clusterInfo class. drawHeatMap returns the same as heatmap or d3heatmap. getSilhouetteInfo returns a silhouetteInfo object. plot (silhouetteInfo signature) returns the result of barplot.

Details

These functions provide an interface to common hierarchical clustering tools such as hclust that can be used for clustering feature groups. makeHCluster generates hierarchical clustering information for feature groups. Intensity data will optionally be averaged and then normalized. A distance matrix is calculated with daisy and clustering is performed with hclust. hClusterFilter isolates feature groups within a featureGroups object which are present within a given branch of a cluster object. drawHeatMap draws a heatmap using the heatmap or d3heatmap function. getSilhouetteInfo is used to obtain silhouette information of a clusterInfo object which is 'cut' by a given sequence of desired number of branches. plotInt makes a plot for all (normalized) intensity profiles of the feature groups within a given cluster. plot (clusterInfo signature) generates a dendrogram from a given cluster object and optionally highlights resulting branches when the cluster is cut. plot (silhouetteInfo signature) plots the various mean silhouette widths stored inside a silhouetteInfo object. The resulting barplot can be used to assess the optimal number of clusters.