Import feature groups from files

Generic function to import feature groups produced by other software from files.

Usage

importFeatureGroups(input, type, ...)

Arguments

input

The input object or path that should be imported. See the algorithm specific functions for more details.

type

What type of data should be imported: "xcms", "xcms3", "kpic2", "table", "brukerpa" (Bruker ProfileAnalysis), "brukertasq" (Bruker TASQ) or "envimass".

...

Further arguments passed to the selected import algorithm function.

Value

An object of a class which is derived from featureGroups.

Details

importFeatureGroups is a generic function that will import feature groups from files by one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as importFeatureGroupsXCMS3 and importFeatureGroupsTable. While these functions may be called directly, importFeatureGroups provides a generic interface and is therefore usually preferred.

See also