Imports feature groups from Bruker ProfileAnalysis

Imports a 'bucket table' produced by Bruker ProfileAnalysis (PA)

Usage

importFeatureGroupsBrukerPA(
  input,
  feat,
  rtWindow = defaultLim("retention", "medium"),
  mzWindow = defaultLim("mz", "medium"),
  intWindow = 5,
  warn = TRUE
)

Arguments

input

The file path to a exported 'bucket table' .txt file from PA.

feat

The features object obtained with findFeaturesBruker.

rtWindow, mzWindow, intWindow

Search window values for retention time (seconds), m/z (Da) and intensity used to find back features within feature groups from PA (+/- the retention/mass/intensity value of a feature).

warn

Warn about missing or duplicate features when relating them back from grouped features.

Value

An object of a class which is derived from featureGroups.

Details

This function imports data from Bruker ProfileAnalysis. This function is called when calling importFeatureGroups with type="brukerpa".

The 'bucket table' should be exported as .txt file. Please note that this function only supports features generated by findFeaturesBruker and it is crucial that DataAnalysis files remain unchanged when features are collected and the bucket table is generated. Furthermore, please note that PA does not retain information about originating features for generated buckets. For this reason, this function tries to find back the original features and care must be taken to correctly specify search parameters (rtWindow, mzWindow, intWindow).

See also