Imports feature groups from enviMass

Imports a 'profiles' produced by enviMass.

Usage

importFeatureGroupsEnviMass(input, feat, positive)

Arguments

input

The path of the enviMass project.

feat

The features object obtained with importFeaturesEnviMass.

positive

Whether data from positive (TRUE) or negative (FALSE) should be loaded.

Value

An object of a class which is derived from featureGroups.

Details

This function imports data from enviMass. This function is called when calling importFeatureGroups with type="envimass".

This function only imports 'raw' profiles, not any results from further componentization steps performed in enviMass. Furthermore, this functionality has only been tested with older versions of enviMass. Finally, please note that this function only supports features imported by importFeaturesEnviMass (obviously, the same project should be used for both importing functions).

See also