Silhouette information

Objects from this class hold silhouette information from a clusterInfo object.

Details

Objects from this class are returned by the getSilhouetteInfo method.

Slots

ranges

Numeric vector of k values that were used to cut the cluster and generate silhouette information from it.

maxk

k value (from ranges parameter) that generated the largest average silhouette width.

maxmw

Largest average silhouette width from all tested ranges.

silInfo

A list containing for each test iteration the fields k (the tested k value), meanw (mean of all silhouette widths) and cmeans (mean silhouette width for each cluster).