Class to store transformation products (TPs) predicted by BioTransformer

This class is used to store prediction results that are generated with BioTransformer.

# S4 method for transformationProductsBT
convertToMFDB(TPs, out, includeParents = FALSE)

# S4 method for transformationProductsBT
filter(
  obj,
  removeParentIsomers = FALSE,
  removeTPIsomers = FALSE,
  removeDuplicates = FALSE,
  minSimilarity = NULL,
  negate = FALSE
)

Arguments

out
The file name of the the output MetFrag database.

includeParents
Set to TRUE to include the parents in the database.

obj, TPs
transformationProductsBTs object to be accessed

removeParentIsomers
If TRUE then TPs with an equal formula as their parent (isomers) are removed.

removeTPIsomers
If TRUE then all TPs with equal formula as any sibling TPs (isomers) are removed. Unlike removeDuplicates, all TP candidates are removed (including the first match). This filter automatically sets removeDuplicates=TRUE to avoid complete removal of TPs with equal structure.

removeDuplicates
If TRUE then the TPs of a parent with duplicate structures (SMILES) are removed. Such duplicates may occur when different transformation pathways yield the same TPs. The first TP candidate with duplicate structure will be kept.

minSimilarity
Minimum structure similarity (0-1) that a TP should have relative to its parent. For details on how these similarities are calculated, see the generateTPsBioTransformer function. May be useful under the assumption that parents and TPs who have a high structural similarity, also likely have a high MS/MS spectral similarity (which can be evaluated after componentization with generateComponentsTPs.

negate
If TRUE then filters are performed in opposite manner.

Value

filter returns a filtered transformationProductsBT object.

Details

Objects from this class are generate with generateTPsBioTransformer. This class is derived from the transformationProducts base class, please see its documentation for more details.

Methods (by generic)

convertToMFDB: Exports this object as a .csv file that can be used as a MetFrag local database.

filter: Performs rule-based filtering of the BioTransformer predictions. Useful to simplify and clean-up the data.

S4 class hierarchy

transformationProducts

transformationProductsBT

References

Djoumbou-Feunang Y, Fiamoncini J, Gil-de-la-Fuente A, Greiner R, Manach C, Wishart DS (2019). “BioTransformer: a comprehensive computational tool for small molecule metabolism prediction and metabolite identification.” Journal of Cheminformatics, 11(1). doi: 10.1186/s13321-018-0324-5 .

Wicker J, Lorsbach T, Gutlein M, Schmid E, Latino D, Kramer S, Fenner K (2015). “enviPath - The environmental contaminant biotransformation pathway resource.” Nucleic Acids Research, 44(D1), D502--D508. doi: 10.1093/nar/gkv1229 .

See also