Class to store transformation products (TPs) obtained from a library

This class is used to store prediction results that are available in a TP library.

# S4 method for transformationProductsLibrary
convertToMFDB(TPs, out, includeParents = FALSE)

Arguments

TPs

transformationProductsLibrary object to be accessed

out

The file name of the the output MetFrag database.

includeParents

Set to TRUE to include the parents in the database.

Details

Objects from this class are generate with generateTPsLibrary. This class is derived from the transformationProducts base class, please see its documentation for more details.

Methods (by generic)

convertToMFDB: Exports this object as a .csv file that can be used as a MetFrag local database.

S4 class hierarchy

transformationProducts

transformationProductsLibrary

See also