(Virtual) Base class for all workflow objects.

All workflow objects (e.g. featureGroups, compounds, etc) are derived from this class. Objects from this class are never created directly.

Usage

# S4 method for class 'workflowStep'
algorithm(obj)

# S4 method for class 'workflowStep'
as.data.table(x, keep.rownames = FALSE, ...)

# S4 method for class 'workflowStep'
as.data.frame(x, row.names = NULL, optional = FALSE, ...)

# S4 method for class 'workflowStep'
show(object)

Arguments

obj, x, object

An object (derived from) this class.

keep.rownames

Ignored.

...

Method specific arguments. Please see the documentation of the derived classes.

row.names, optional

Ignored.

Methods (by generic)

algorithm(workflowStep): Returns the algorithm that was used to generate an object.

as.data.table(workflowStep): Summarizes the data in this object and returns this as a data.table.

as.data.frame(workflowStep): This method simply calls as.data.table and converts the result to a classic a data.frame.

show(workflowStep): Shows summary information for this object.

Slots

algorithm

The algorithm that was used to generate this object. Use the algorithm method for access.

S4 class hierarchy

workflowStep

transformationProducts

transformationProductsStructure

transformationProductsStructureConsensus

transformationProductsCTS

transformationProductsAnnComp

transformationProductsBT

transformationProductsLibrary

transformationProductsFormula

transformationProductsAnnForm

transformationProductsLibraryFormula

transformationProductsLogic

features

featuresSet

featuresUnset

featuresFromFeatGroups

featuresConsensus

featuresBruker

featuresEnviPick

featuresKPIC2

featuresOpenMS

featuresPiek

featuresSAFD

featuresSIRIUS

featuresTable

featuresBrukerTASQ

featuresXCMS

featuresXCMS3

featureGroups

featureGroupsSet

featureGroupsScreeningSet

featureGroupsUnset

featureGroupsScreening

featureGroupsSetScreeningUnset

featureGroupsBruker

featureGroupsConsensus

featureGroupsEnviMass

featureGroupsGreedy

featureGroupsIMS

featureGroupsKPIC2

featureGroupsOpenMS

featureGroupsSIRIUS

featureGroupsTable

featureGroupsBrukerTASQ

featureGroupsXCMS

featureGroupsXCMS3

components

componentsCamera

componentsFeatures

componentsCliqueMS

componentsOpenMS

componentsClust

componentsIntClust

componentsSpecClust

componentsSet

componentsNetSet

componentsNTSet

componentsUnset

componentsNet

componentsNetUnset

componentsNT

componentsNTUnset

componentsRC

componentsTPs

featureAnnotations

formulas

formulasConsensus

formulasSet

formulasUnset

formulasSIRIUS

compounds

compoundsConsensus

compoundsMF

compoundsSet

compoundsUnset

compoundsSIRIUS

MSPeakLists

MSPeakListsSet

MSPeakListsUnset

MSLibrary