Interactive GUI utilities

Interactive utilities using shiny to provide a graphical user interface (GUI).

# S4 method for featureGroups
checkChromatograms(fGroups, mzWindow = 0.005,
  enabledFGroups = NULL)

# S4 method for featureGroups,MSPeakLists,compounds
compoundViewer(fGroups,
  MSPeakLists, compounds)

newProject()

Arguments

fGroups
A featureGroups object.

mzWindow
Default m/z window to be used for creating extracted ion chromatograms (EICs).

enabledFGroups
A logical vector that states for each feature group whether it should be kept (TRUE) or not (FALSE). The order is the same as the fGroups parameter. If NULL then all feature groups are considered to be kept.

MSPeakLists
A MSPeakLists object.

compounds
A compounds object.

Value

checkChromatograms returns a logical vector for all feature groups that were selected to be kept (TRUE) or not (FALSE). This result can be passed to the enabledFGroups parameter for subsequent calls to checkChromatograms in order to restore the keep/not keep state from a previous call. To actually remove unwanted feature groups the object should be subset by the subsetting ([) operator to which the return value should be passed as the second parameter.

Details

checkChromatograms is used to review chromatographic information for feature groups. This is especially useful to get a visual impression of the quality of detected features. In addition, this function may be used to remove unwanted (e.g. outlier) features. Better performance is often obtained when an external browser is used to use this Shiny application. Furthermore, when a large featureGroups object is used it is recommended to limit the number of analyses/feature groups by subsetting the object.

The compoundViewer method is used to view compound identification results. It will display available candidate information such as scorings and identifiers, MS/MS spectra with explained peaks and chemical structures.

The newProject function is used to quickly generate a processing R script. This tool allows the user to quickly select the targeted analyses, workflow steps and configuring some of their common parameters. This function requires to be run within a RStudio session. The resulting script is either added to the current open file or to a new file. The analysis information will be written to a .csv file so that it can easily be modified afterwards.