Class containing MS Peak Lists

Contains all MS (and MS/MS where available) peak lists for a featureGroups object.

# S4 method for MSPeakLists
peakLists(obj)

# S4 method for MSPeakLists
algorithm(obj)

# S4 method for MSPeakLists
length(x)

# S4 method for MSPeakLists
show(object)

# S4 method for MSPeakLists
filter(obj, absMSIntThr = NULL,
  absMSMSIntThr = NULL, relMSIntThr = NULL, relMSMSIntThr = NULL,
  topMSPeaks = NULL, topMSMSPeaks = NULL, deIsotopeMS = FALSE,
  deIsotopeMSMS = FALSE)

Arguments

obj, x, object
The MSPeakLists object to access.

absMSIntThr, absMSMSIntThr, relMSIntThr, relMSMSIntThr
Absolute/relative intensity threshold for MS or MS/MS peak lists. NULL for none.

topMSPeaks, topMSMSPeaks
Only consider this amount of MS or MS/MS peaks with highest intensity. NULL to consider all.

deIsotopeMS, deIsotopeMSMS
Remove any isotopic peaks in MS or MS/MS peak lists. This may improve data processing steps which do not assume the presence of isotopic peaks (e.g. MetFrag for MS/MS). Note that getMzRPeakLists does not (yet) support flagging of isotopes.

Value

peakLists returns a nested list containing MS (and MS/MS where available) peak lists per feature group and per analysis. The format is: [[analysis]][[featureGroupName]][[MSType]][[PeakLists]] where MSType is either "MS" or "MSMS" and PeakLists a data.table containing all m/z values (mz column) and their intensities (intensity column). In addition, the peak list tables may contain a cmp column which contains an unique alphabatical identifier to which isotopic cluster (or "compound") a mass belongs (only supported by MS peak lists generated by Bruker tools at the moment).

Details

Objects for this class are returned by MS peak lists generators.

Methods (by generic)

peakLists: Accessor method to obtain the MS peak lists.

algorithm: Retrieve the algorithm (a character string) used to generate the MS peak lists.

length: Obtain total number of m/z values.

show: Shows summary information for this object.

filter: provides post filtering of generated MS peak lists, which may further enhance quality of subsequent workflow steps (e.g. formulae calculation and compounds identification) and/or speed up these processes.

Slots

peakLists

Contains a list of all MS (and MS/MS) peak lists. Use the peakLists method for access.

algorithm

The algorithm that was used to generate the MS peak lists. Use the algorithm method for access.