Class containing MS Peak Lists

Contains all MS (and MS/MS where available) peak lists for a featureGroups object.

# S4 method for MSPeakLists
peakLists(obj)

# S4 method for MSPeakLists
averagedPeakLists(obj)

# S4 method for MSPeakLists
analyses(obj)

# S4 method for MSPeakLists
groupNames(obj)

# S4 method for MSPeakLists
length(x)

# S4 method for MSPeakLists
show(object)

# S4 method for MSPeakLists,ANY,ANY,missing
[(x, i, j, ..., reAverage = TRUE,
  drop = TRUE)

# S4 method for MSPeakLists,ANY,ANY
[[(x, i, j)

# S4 method for MSPeakLists
$(x, name)

# S4 method for MSPeakLists
as.data.table(x, fGroups = NULL,
  averaged = TRUE)

# S4 method for MSPeakLists
filter(obj, absMSIntThr = NULL,
  absMSMSIntThr = NULL, relMSIntThr = NULL, relMSMSIntThr = NULL,
  topMSPeaks = NULL, topMSMSPeaks = NULL, isolatePrec = NULL,
  deIsotopeMS = FALSE, deIsotopeMSMS = FALSE, withMSMS = FALSE,
  retainPrecursorMSMS = TRUE, negate = FALSE)

# S4 method for MSPeakLists
plotSpec(obj, groupName, analysis = NULL,
  MSLevel = 1, title = NULL, useGGPlot2 = FALSE, xlim = NULL,
  ylim = NULL, ...)

getDefIsolatePrecParams(...)

Arguments

obj, x, object
The MSPeakLists object to access.

i, j
A numeric or character value which is used to select analyses/feature groups by their index or name, respectively (for the order/names see analyses()/groupNames()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all analyses/feature groups are selected.

For [[: should be a scalar value. If j is not specified, i selects by feature groups instead.

…
Further arguments passed to plot.

reAverage
Set to TRUE to regenerate averaged MS peak lists after subsetting analyses.

drop
ignored.

name
The feature group name (partially matched).

fGroups
The featureGroups object that was used to generate this object. If not NULL it is used to add feature group information (retention and m/z values).

averaged
If TRUE then feature group averaged peak list data is used.

absMSIntThr, absMSMSIntThr, relMSIntThr, relMSMSIntThr
Absolute/relative intensity threshold for MS or MS/MS peak lists. NULL for none.

topMSPeaks, topMSMSPeaks
Only consider this amount of MS or MS/MS peaks with highest intensity. NULL to consider all.

isolatePrec
If not NULL then value should be a list with parameters used for isolating the precursor and its isotopes in MS peak lists (see Isolating precursor data). Alternatively, TRUE to apply the filter with default settings (as given with getDefIsolatePrecParams).

deIsotopeMS, deIsotopeMSMS
Remove any isotopic peaks in MS or MS/MS peak lists. This may improve data processing steps which do not assume the presence of isotopic peaks (e.g. MetFrag for MS/MS). Note that getMzRPeakLists does not (yet) support flagging of isotopes.

withMSMS
If set to TRUE then only results will be retained for which MS/MS data is available. if negate=TRUE then only results without MS/MS data will be retained.

retainPrecursorMSMS
If TRUE then precursor peaks will never be filtered out from MS/MS peak lists (note that precursors are never removed from MS peak lists). The negate argument does not affect this setting.

negate
If TRUE then filters are applied in opposite manner.

groupName
The name of the feature group for which a plot should be made.

analysis
The name of the analysis for which a plot should be made. If NULL then data from the feature group averaged peak list is used.

MSLevel
The MS level: 1 for regular MS, 2 for MSMS.

title
The title of the plot. If NULL a title will be automatically made.

useGGPlot2
If TRUE then ggplot2 is used for plotting, otherwise base plot used. For plotSpec, ggplot2 allows nicely repelled text for annotation. However, base plot is generally faster.

xlim, ylim
Sets the plot size limits used by plot. Set to NULL for automatic plot sizing.

Value

peakLists returns a nested list containing MS (and MS/MS where available) peak lists per feature group and per analysis. The format is: [[analysis]][[featureGroupName]][[MSType]][[PeakLists]] where MSType is either "MS" or "MSMS" and PeakLists a data.table containing all m/z values (mz column) and their intensities (intensity column). In addition, the peak list tables may contain a cmp column which contains an unique alphabatical identifier to which isotopic cluster (or "compound") a mass belongs (only supported by MS peak lists generated by Bruker tools at the moment).

averagedPeakLists returns a nested list of feature group averaged peak lists in a similar format as peakLists.

plotSpec will return a ggplot object if useGGPlot2 is TRUE.

Details

Objects for this class are returned by MS peak lists generators.

The getDefIsolatePrecParams is used to create a parameter list for isolating the precursor and its isotopes (see Isolating precursor data).

Methods (by generic)

peakLists: Accessor method to obtain the MS peak lists.

averagedPeakLists: Accessor method to obtain the feature group averaged MS peak lists.

analyses: returns a character vector with the names of the analyses for which data is present in this object.

groupNames: returns a character vector with the names of the feature groups for which data is present in this object.

length: Obtain total number of m/z values.

show: Shows summary information for this object.

[: Subset on analyses/feature groups.

[[: Extract a list with MS and MS/MS (if available) peak lists. If the second argument (j) is not specified the averaged peak lists for the group specified by the first argument (i) will be returned.

$: Extract group averaged MS peaklists for a feature group.

as.data.table: Returns all MS peak list data in a table.

filter: provides post filtering of generated MS peak lists, which may further enhance quality of subsequent workflow steps (e.g. formulae calculation and compounds identification) and/or speed up these processes.

plotSpec: Plots a spectrum using MS or MS/MS peak lists for a given feature group.

Slots

peakLists

Contains a list of all MS (and MS/MS) peak lists. Use the peakLists method for access.

metadata

Metadata for all spectra used to generate peak lists. Follows the format of the peakLists slot.

averagedPeakLists

A list with averaged MS (and MS/MS) peak lists for each feature group.

avgPeakListArgs

A list with arguments used to generate feature group averaged MS(/MS) peak lists.

Isolating precursor data

Formula calculation typically relies on evaluating the measured isotopic pattern from the precursor to score candidates. Some algorithms (currently only GenForm) penalize candidates if mass peaks are present in MS1 spectra that do not contribute to the isotopic pattern. Since these spectra are typically very 'noisy' due to background and co-eluting ions, an additional filtering step may be recommended prior to formula calculation. During this precursor isolation step all mass peaks are removed that are (1) not the precursor and (2) not likely to be an isotopologue of the precursor. To determine potential isotopic peaks the following parameters are used:

maxIsotopes The maximum number of isotopes to consider. For instance, a value of 5 means that M+0 (i.e. the monoisoptic peak) till M+5 is considered. All mass peaks outside this range are removed.

mzDefectRange A two-sized vector specifying the minimum (can be negative) and maximum m/z defect deviation compared to the precursor m/z defect. When chlorinated, brominated or other compounds with strong m/z defect in their isotopologues are to be considered a higher range may be desired. On the other hand, for natural compounds this range may be tightened. Note that the search range is propegated with increasing distance from the precursor, e.g. the search range is doubled for M+2, tripled for M+3 etc.

intRange A two-sized vector specifying the minimum and maximum relative intensity range compared to the precursor. For instance, c(0.001, 2) removes all peaks that have an intensity below 0.1% are above 200% of that of the precursor.

z The z value (i.e. absolute charge) to be considerd. For instance, a value of 2 would look for M+0.5, M+1 etc. Note that the mzDefectRange is adjusted accordingly (e.g. halved if z=2).

maxGap The maximum number of missing adjacent isotopic peaks ('gaps'). If the (rounded) m/z difference to the previous peak exceeds this value then this and all next peaks will be removed. Similar to z, the maximum gap is automatically adjusted for charge.

These parameters should be in a list that is passed to the isolatePrec argument to filter. The default values can be obtained with the getDefIsolatePrecParams function:

maxIsotopes=5; mzDefectRange=c(-0.01, 0.01); intRange=c(0.001, 2); z=1; maxGap=2

S4 class hierarchy

workflowStep

MSPeakLists