Generation of MS Peak Lists

Functionality to generate MS peak lists.

# S4 method for featureGroups
generateMSPeakLists(fGroups, algorithm, ...)

generateDAMSPeakLists(fGroups, bgsubtr = TRUE, maxRtMSWidth = NULL,
  clear = TRUE, save = TRUE, MSMSType = "MSMS")

generateDAFMFMSPeakLists(fGroups)

generateMzRPeakLists(fGroups, avgMzWindow = 0.001, maxRtMSWidth = NULL,
  avgTopPeaks = 50, avgMinIntensity = 500, avgMassFun = mean,
  avgMethod = "distance", precursorMzWindow = 8, topMost = NULL)

Arguments

fGroups
The featureGroups object from which MS peak lists should be extracted.

algorithm
A character string describing the algorithm that should be used: "bruker", "brukerfmf", "mzr"

…
Any parameters to be passed to the selected MS peak lists generation algorithm.

bgsubtr
If TRUE background will be subtracted using the 'spectral' algorithm.

maxRtMSWidth
Maximum total chromatographic peak width (seconds) used for spectrum averaging. If NULL all spectra from a feature will be taken into account. Lower to decrease processing time.

clear
Remove any existing chromatogram traces/mass spectra prior to making new ones.

save
Save data files after completion.

MSMSType
The type of MS/MS experiment performed: "MSMS" for MRM/AutoMSMS or "BBCID" for broadband CID.

avgMzWindow
m/z window (in Da) used for clustering m/z values when spectra are averaged. Too small windows will prevent clustering m/z values (thus erroneously creating 'extra' values), whereas too big windows may cluster unrelated m/z values from different or even the same spectrum together.

avgTopPeaks
Only retain a maximum number of avgTopPeaks MS peaks when generating averaged spectra. Lowering this number may exclude more irrelevant (noisy) MS peaks and decrease processing time, whereas higher values may avoid excluding lower intense MS peaks that may still be of interest.

avgMinIntensity
MS peaks with intensities below this value will be completely excluded (overrides avgTopPeaks).

avgMassFun
Function that is used to calculate average m/z values.

avgMethod
Method used for producing averaged MS spectra. Valid values are "hclust", used for hierarchical clustering, and "distance", to use the between peak distance. The latter method significantly reduces processing time and memory requirements, at the (potential) cost of slightly reducing accuracy.

precursorMzWindow
The m/z window (in Da) to find MS/MS spectra of a precursor. This is typically used for Data-Dependent like MS/MS data and should correspond to the isolation m/z width that was used to collect the data. For Data-Independent MS/MS experiments, where precursor ions are not isolated prior to fragmentation (e.g. bbCID, MSe, all-ion, ...) the value should be NULL.

topMost
Only extract MS peak lists from a maximum of topMost analyses with highest intensity. If NULL all analyses will be used.

Value

A MSPeakLists object that can be used for formulae calculation and compound identification.

Details

Formula calculation and identification tools rely on mass spectra that belong to features of interest. For processing, MS (and MS/MS) spectra are typically reduced to a table with a column containing measured m/z values and a column containing their intensities. These 'MS peak lists' can then be used for formula generation and compound generation. By default, peak lists are generated from all analyses containing a particular feature group. The main advantage is that this will allow subsequent processes (at the moment only formulae calculation) to generate a consensus from feature group data originating from multiple analyses. Nevertheless, it might be useful to limit to maximum amount of analyses (topMost parameter) to signficantly reduce the processing time required for MS peak list extraction.

Several functions exist to automatically extract MS peak lists for feature groups.

generateMSPeakLists is a generic function that will generate MS peak lists using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateFormulasDA and generateFormulasGenForm. While these functions may be called directly, generateMSPeakLists provides a generic interface and is therefore usually preferred.

generateDAMSPeakLists uses Bruker DataAnalysis to generate MS peak lists. Naturally, this only works with analyses in the Bruker data format (.d). This function leverages DataAnalysis functionality to support averaging of spectra, background subtraction and identification of isotopes. In order to obtain mass spectra TICs will be added of the MS and relevant MS/MS signals.

generateDAFMFMSPeakLists is similar to generateDAMSPeakLists, but uses compounds that were generated by the Find Molecular Features (FMF) algorithm to extract MS peak lists. This is generally much faster than generateDAMSPeakLists, however, it only works when features were obtained using the findFeaturesBruker function. Since all MS spectra are generated in advance by Bruker DataAnalysis, no further parameters exist to customize its operation.

generateMzRPeakLists uses the mzR package to extract MS peak lists. For this analyses should be either in .mzXML or .mzML format. This function can optionally average multiple spectra over a chromatgraphic peak to improve accuracy.

Note

generateDAMSPeakLists requires that the Component column is active (Method-->Parameters-->Layouts-->Mass List Layout).

References

Averaging of mass spectra algorithms used by generateMzRPeakLists were based on the msProcess R package (now archived on CRAN). A cross-platform toolkit for mass spectrometry and proteomics Chambers, Matthew C. and Maclean, Brendan and Burke, Robert and Amodei, Dario and Ruderman, Daniel L. and Neumann, Steffen and Gatto, Laurent and Fischer, Bernd and Pratt, Brian and Egertson, Jarrett and Hoff, Katherine and Kessner, Darren and Tasman, Natalie and Shulman, Nicholas and Frewen, Barbara and Baker, Tahmina A. and Brusniak, Mi-Youn and Paulse, Christopher and Creasy, David and Flashner, Lisa and Kani, Kian and Moulding, Chris and Seymour, Sean L. and Nuwaysir, Lydia M. and Lefebvre, Brent and Kuhlmann, Frank and Roark, Joe and Rainer, Paape and Detlev, Suckau and Hemenway, Tina and Huhmer, Andreas and Langridge, James and Connolly, Brian and Chadick, Trey and Holly, Krisztina and Eckels, Josh and Deutsch, Eric W. and Moritz, Robert L. and Katz, Jonathan E. and Agus, David B. and MacCoss, Michael and Tabb, David L. and Mallick, Parag Nat Biotechnol. 2012 NOct;30(10):918-920. Mol Cell Proteomics. 2010 Aug 17. mzML - a Community Standard for Mass Spectrometry Data. Martens L, Chambers M, Sturm M, Kessner D, Levander F, Shofstahl J, Tang WH, Rompp A, Neumann S, Pizarro AD, Montecchi-Palazzi L, Tasman N, Coleman M, Reisinger F, Souda P, Hermjakob H, Binz PA, Deutsch EW. Nat Biotechnol. 2004 Nov;22(11):1459-66. A common open representation of mass spectrometry data and its application to proteomics research. Pedrioli PG, Eng JK, Hubley R, Vogelzang M, Deutsch EW, Raught B, Pratt B, Nilsson E, Angeletti RH, Apweiler R, Cheung K, Costello CE, Hermjakob H, Huang S, Julian RK, Kapp E, McComb ME, Oliver SG, Omenn G, Paton NW, Simpson R, Smith R, Taylor CF, Zhu W, Aebersold R. Mol Syst Biol. 2005;1:2005.0017. Epub 2005 Aug 2. A uniform proteomics MS/MS analysis platform utilizing open XML file formats. Keller A, Eng J, Zhang N, Li XJ, Aebersold R. Bioinformatics. 2008 Nov 1;24(21):2534-6. Epub 2008 Jul 7. ProteoWizard: open source software for rapid proteomics tools development. Kessner D, Chambers M, Burke R, Agus D, Mallick P.

See also

MSPeakLists-class