Analysis information

Required information for analyses that should be processed and utilities to automatically generate this information.

generateAnalysisInfo(paths, groups = "", refs = "",
  fileTypes = c("Bruker", "mzXML", "mzML"))

generateAnalysisInfoFromEnviMass(path)

Arguments

paths
A character vector containing one or more file paths that should be used for finding the analyses.

groups, refs
An (optional) character vector containing replicate groups and references, respectively (will be recycled). If groups is an empty character string ("") the analysis name will be set as replicate group.

fileTypes
A character vector of analyses file types. Valid values are: Bruker, mzXML and mzML.

path
The path of the enviMass project.

Details

Several properties need to be known about analyses that should be processed during various workflow steps such as finding features, averaging intensities of feature groups and blank subtraction. This information should be made available with an 'analysis info' object, which is a data.frame containing the following columns:

path the full path to the analysis.

analysis the filename without extension. Must be unique, even if the path is different.

group name of replicate group. A replicate group is used to group analyses together that are replicates of each other. Thus, the group column for all analyses considered to be belonging to the same replicate group should have an equal (but unique) value. Used for e.g. avaraging and filter.

ref name of the reference: all analyses within this replicate group are used by filter for blank subtraction. Multiple entries can be entered by separation with a comma.

conc a numeric value specifying the 'concentration' of the analysis. This can be actually any kind of quantitative value such as exposure time, dilution factor or anything else which may be used to form a linear relationship.

Most functionality requires the data files to be in either the .mzXML or .mzML format. Functionality that utilizes Bruker DataAnalysis (e.g. findFeaturesBruker) may only work with data files in the proprietary .d format. Therefore, when tools with varying requirements are mixed (a common scenario), the data files also need to be present in the required formats. To deal with this situation the data files with varying formats should all be placed in the same path that was used to specify the location of the analysis.

Whether analysis data files are present in multiple formats or not, each analysis should only be entered once. The analysis column is used as a basename to automatically find back the data file with the required format, hence, analysis names should be specified as the file name without its file extension.

The group column is mandatory and needs to be filled in for each analysis. The ref column should also be present, however, these may contain empty character strings ("") for analyses where no blank subtraction should occur. The conc column is only required when using the regression method and optional for the screenTargets method. generateAnalysisInfo is an utility function that automatically generates an analysis information object. It will collect all datafiles from given file paths and convert the filenames into valid analysis names (i.e. without extensions such as .d and .mzML). Duplicate analyses, which may appear when datafiles with different file extension (.d, .mzXML and/or .mzML) are present, will be automatically removed. generateAnalysisInfoFromEnviMass loads analysis information from an enviMass project.