Bruker DataAnalysis utilities

Miscellaneous utility functions which interface with Bruker DataAnalysis

showDataAnalysis()

setDAMethod(anaInfo, method)

recalibrarateDAFiles(anaInfo)

getDACalibrationError(anaInfo)

exportDAFiles(anaInfo, format = "mzML", exportLine = TRUE,
  outPath = anaInfo$path, overWrite = FALSE)

addDAEIC(analysis, path, mz, mzWidth, ctype = "EIC", mtype = "MS",
  polarity = "both", bgsubtr = FALSE, fragpath = "", name = NULL,
  hideDA = TRUE)

Arguments

anaInfo
Analysis info table

method
The full path of the DataAnalysis method.

format
The output format of exported files. Should be either "mzXML", "mzML" or "mzData".

exportLine
Export line spectra (TRUE) or profile spectra (FALSE). Usually line spectra are preferred, since profile spectra use signficantly more disk space and increase required memory during processing.

outPath
Character vector of output paths for exported analyses. Will be recycled if necessary.

overWrite
If TRUE existing files will be overwritten.

analysis
Analysis name (without file extension).

path
path of the analysis.

mz
m/z (Da) value used for the chromatographic trace (if applicable).

mzWidth
m/z width (in Da) used for the chromatographic trace (if applicable).

ctype
Type of the chromatographic trace. Valid options are: "EIC" (extracted ion chromatogram), "TIC" (total ion chromatogram) and "BPC" (Base Peak Chromatogram).

mtype
MS filter for chromatographic trace. Valid values are: "all", "MS", "MSMS", "allMSMS" and "BBCID".

polarity
Polarity filter for chromatographic trace. Valid values: "both", "positive" and "negative".

bgsubtr
If TRUE then background subtraction ('Spectral' algorithm) will be performed.

fragpath
Precursor m/z used for MS/MS traces ("" for none).

name
The name for the chromatographic trace. NULL for none.

hideDA
Hides DataAnalysis while adding the chromatographic trace (faster).

Value

getDACalibrationError returns a data.frame with a column of all analyses (named analysis) and their mass error (named error).

Details

These functions communicate directly with Bruker DataAnalysis to provide various functionality, such as calibrating and exporting data and adding chromatographic traces. For this the RDCOMClient package is required to be installed. showDataAnalysis makes a hidden DataAnalysis window visible again. Most functions using DataAnalysis will hide the window during processing for efficiency reasons. If the window remains hidden (e.g. because there was an error) this function can be used to make it visible again. setDAMethod Sets a given DataAnalysis method (.m file) to a set of analyses. recalibrarateDAFiles Performs automatic mass recalibration of a given set of analyses. The current method settings for each analyses will be used. getDACalibrationError is used to obtain the standard deviation of the current mass calibration (in ppm). exportDAFiles will export a set of analyses either in .mzXML or .mzML formats. addDAEIC adds an Extracted Ion Chromatogram (EIC) or other chromatographic trace to a given analysis which can be used directly with DataAnalysis.

See also

analysis-information