Grouping feature groups in components

Functionality to automatically group related feature groups (e.g. isotopes, adducts and homologues) to assist and simplify compound annotation.

# S4 method for featureGroups
generateComponents(fGroups, algorithm, ...)

generateComponentsCAMERA(
  fGroups,
  ionization,
  onlyIsotopes = FALSE,
  minSize = 2,
  relMinReplicates = 0.5,
  extraOpts = NULL
)

generateComponentsIntClust(
  fGroups,
  method = "complete",
  metric = "euclidean",
  normFunc = max,
  average = TRUE,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)

generateComponentsNontarget(
  fGroups,
  ionization,
  rtRange = c(-120, 120),
  mzRange = c(5, 120),
  elements = c("C", "H", "O"),
  rtDev = 30,
  absMzDev = 0.002,
  absMzDevLink = absMzDev * 2,
  extraOpts = NULL,
  traceHack = all(R.Version()[c("major", "minor")] >= c(3, 4))
)

generateComponentsRAMClustR(
  fGroups,
  st = NULL,
  sr = NULL,
  maxt = 12,
  hmax = 0.3,
  normalize = "TIC",
  ionization,
  absMzDev = 0.002,
  relMzDev = 5,
  minSize = 2,
  relMinReplicates = 0.5,
  RCExperimentVals = list(design = list(platform = "LC-MS"), instrument =
    list(ionization = ionization, MSlevs = 1)),
  extraOptsRC = NULL,
  extraOptsFM = NULL
)

Arguments

fGroups
featureGroups object for which components should be generated.

algorithm
A character string describing the algorithm that should be used: "ramclustr", "camera", "nontarget", "intclust"

...
Any parameters to be passed to the selected component generation algorithm.

ionization
Which ionization polarity was used to generate the data: should be "positive" or "negative".

onlyIsotopes
Logical value. If TRUE only isotopes are considered when generating components (faster). Corresponds to quick argument of CAMERA::annotate.

minSize
The minimum size of a component. Smaller components than this size will be removed. For RAMClustR: sets the minModuleSize argument to ramclustR. See note below.

relMinReplicates
Feature groups within a component are only kept when they contain data for at least this (relative) amount of replicate analyses. For instance, 0.5 means that at least half of the replicates should contain data for a particular feature group in a component. In this calculation replicates that are fully absent within a component are not taken in to account. See note below.

extraOpts
Named character vector with extra arguments directly passed to homol.search (generateComponentsNontarget) or CAMERA::annotate (generateComponentsCAMERA). Set to NULL to ignore.

method
Clustering method that should be applied (passed to hclust).

metric
Distance metric used to calculate the distance matrix (passed to daisy).

normFunc
Function that should be used for normalization of data. Intensitity values of a feature group will be divided by the result of this function when it is called with all intensity values of that feature group. For example, when max is used normalized intensities will be between zero and one.

average
If TRUE then all intensitity data will be averaged for each replicate group.

maxTreeHeight, deepSplit, minModuleSize
Arguments used by cutreeDynamicTree.

rtRange
A numeric vector containing the minimum and maximum retention time (in seconds) between homologues. Series are always considered from low to high m/z, thus, a negative minimum retention time allows detection of homologous series with increasing m/z and decreasing retention times. These values set the minrt and maxrt arguments of homol.search.

mzRange
A numeric vector specifying the minimum and maximum m/z increment of a homologous series. Sets the minmz and maxmz arguments of homol.search.

elements
A character vector with elements to be considered for detection of repeating units. Sets the elements argument of homol.search function.

rtDev, absMzDev, relMzDev
Maximum deviation for retention time or absolute/relative m/z.

For generateComponentsRAMClustR: Sets the mzabs.error and ppm.error arguments to do.findmain.

For generateComponentsNontarget: Sets the rttol and mztol arguments of homol.search.

absMzDevLink
Maximum absolute m/z deviation when linking series. This should usually be a bit higher than absMzDev to ensure proper linkage.

traceHack
Currently homol.search does not work with R >3.3.3. This flag, which is enabled by default on these R versions, implements a (messy) workaround (more details here).

st, sr, maxt, hmax, normalize
Arguments to tune the behaviour of feature group clustering. See their documentation from ramclustR. When st is NULL it will be automatically calculated as the half of the median for all chromatographic peak widths.

RCExperimentVals
A named list containing two more lists: design and instrument. These are used to construct the ExpDes argument passed to ramclustR.

extraOptsRC, extraOptsFM
Named list with further arguments to be passed to ramclustR and do.findmain. Set to NULL to ignore.

Value

A components (derived) object containing all generated components.

Details

Several algorithms are provided to group feature groups that are related in some (chemical) way to each other. These components generally include adducts, isotopes, in-source fragments and homologues. The linking of this data is generally useful to provide more information for compound annotation and reduce the data size and thus complexity.

generateComponents is a generic function that will generate components using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateComponentsRAMClustR and generateComponentsNontarget. While these functions may be called directly, generateComponents provides a generic interface and is therefore usually preferred.

generateComponentsCAMERA provides an interface to CAMERA which is used to generate components from known adducts, isotopes and in-source fragments. The specified featureGroups object is automatically converted to an xcmsSet object using getXCMSSet.

generateComponentsIntClust generates components based on intensity profiles of feature groups. Hierarchical clustering is performed on normalized (and optionally replicate averaged) intensity data and the resulting dendrogram is automatically cut with cutreeDynamicTree. The distance matrix is calculated with daisy and clustering is performed with hclust. The clustering of the resulting components can be further visualized and modified using the methods defined for componentsIntClust.

generateComponentsNontarget uses the nontarget R package to generate components by unsupervised detection of homologous series. In the first step the homol.search function is used to detect all homologues within each replicate group (analyses within each replicate group are averaged prior to detection). Then, homologous series across replicate groups are merged in case of full overlap or when merging of partial overlapping series causes no conflicts.

generateComponentsRAMClustR uses RAMClustR to generate components from feature groups which follow similar chromatographic retention profiles, but are not necessarily restricted to known rules (e.g. adducts or isotopes). This method uses the ramclustR functions for generating the components, whereas do.findmain is used for annotation.

Note

For generateComponentsCAMERA and generateComponentsRAMClustR: the minSize and relMinReplicates arguments provide additional filtering functionality not provided by CAMERA or RAMClustR (except minSize). Note that these filters are enabled by default, hence, final results may be different than what CAMERA/RAMClustR normally would return.

References

Kuhl, C., Tautenhahn, R., Boettcher, C., Larson, T. R. and Neumann, S. CAMERA: an integrated strategy for compound spectra extraction and annotation of liquid chromatography/mass spectrometry data sets. Analytical Chemistry, 84:283-289 (2012)

Martin Loos (2016). nontarget: Detecting Isotope, Adduct and Homologue Relations in LC-MS Data. R package version 1.9.

Loos, M., Gerber, C., Corona, F., Hollender, J., Singer, H. (2015). Accelerated isotope fine structure calculation using pruned transition trees, Analytical Chemistry 87(11), 5738-5744.

Broeckling, Heuberger CD, Prince AL, Ingelsson JA, Prenni E, E. J (2013). “Assigning precursor-product ion relationships in indiscriminant MS/MS data from non-targeted metabolite profiling studies.” Analytical Chemistry, 9, 33-43.

Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE (2014). “RAMClust: A Novel Feature Clustering Method Enables Spectral-Matching-Based Annotation for Metabolomics Data.” Analytical Chemistry, 86 (14), 6812–6817.