Grouping feature groups in components

Functionality to automatically group related feature groups (e.g. isotopes, adducts and homologues) to assist and simplify compound annotation.

# S4 method for featureGroups
generateComponents(fGroups, algorithm, ...)

# S4 method for featureGroups
generateComponentsCAMERA(
  fGroups,
  ionization = NULL,
  onlyIsotopes = FALSE,
  minSize = 2,
  relMinReplicates = 0.5,
  extraOpts = NULL
)

# S4 method for featureGroupsSet
generateComponentsCAMERA(fGroups, ionization = NULL, ...)

# S4 method for featureGroups
generateComponentsCliqueMS(
  fGroups,
  ionization = NULL,
  maxCharge = 1,
  maxGrade = 2,
  ppm = 10,
  adductInfo = NULL,
  absMzDev = 0.005,
  minSize = 2,
  relMinAdductAbundance = 0.75,
  adductConflictsUsePref = TRUE,
  NMConflicts = c("preferential", "mostAbundant", "mostIntense"),
  prefAdducts = c("[M+H]+", "[M-H]-"),
  extraOptsCli = NULL,
  extraOptsIso = NULL,
  extraOptsAnn = NULL,
  parallel = TRUE
)

# S4 method for featureGroupsSet
generateComponentsCliqueMS(fGroups, ionization = NULL, ...)

# S4 method for featureGroups
generateComponentsIntClust(
  fGroups,
  method = "complete",
  metric = "euclidean",
  normFunc = max,
  average = TRUE,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)

# S4 method for featureGroups
generateComponentsNontarget(
  fGroups,
  ionization = NULL,
  rtRange = c(-120, 120),
  mzRange = c(5, 120),
  elements = c("C", "H", "O"),
  rtDev = 30,
  absMzDev = 0.002,
  absMzDevLink = absMzDev * 2,
  traceHack = all(R.Version()[c("major", "minor")] >= c(3, 4)),
  ...
)

# S4 method for featureGroupsSet
generateComponentsNontarget(fGroups, ionization = NULL, ...)

# S4 method for featureGroups
generateComponentsOpenMS(
  fGroups,
  ionization = NULL,
  chargeMin = 1,
  chargeMax = 1,
  chargeSpan = 3,
  qTry = "heuristic",
  potentialAdducts = NULL,
  minRTOverlap = 0.66,
  retWindow = 1,
  absMzDev = 0.005,
  minSize = 2,
  relMinAdductAbundance = 0.75,
  adductConflictsUsePref = TRUE,
  NMConflicts = c("preferential", "mostAbundant", "mostIntense"),
  prefAdducts = c("[M+H]+", "[M-H]-"),
  extraOpts = NULL
)

# S4 method for featureGroupsSet
generateComponentsOpenMS(
  fGroups,
  ionization = NULL,
  chargeMin = 1,
  chargeMax = 1,
  chargeSpan = 3,
  qTry = "heuristic",
  potentialAdducts = NULL,
  ...
)

defaultOpenMSAdducts(ionization)

# S4 method for featureGroups
generateComponentsRAMClustR(
  fGroups,
  ionization = NULL,
  st = NULL,
  sr = NULL,
  maxt = 12,
  hmax = 0.3,
  normalize = "TIC",
  absMzDev = 0.002,
  relMzDev = 5,
  minSize = 2,
  relMinReplicates = 0.5,
  RCExperimentVals = list(design = list(platform = "LC-MS"), instrument =
    list(ionization = ionization, MSlevs = 1)),
  extraOptsRC = NULL,
  extraOptsFM = NULL
)

# S4 method for featureGroupsSet
generateComponentsRAMClustR(fGroups, ionization = NULL, ...)

# S4 method for featureGroups
generateComponentsSpecClust(
  fGroups,
  MSPeakLists,
  method = "complete",
  specSimParams = getDefSpecSimParams(),
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)

# S4 method for featureGroups
generateComponentsTPs(
  fGroups,
  fGroupsTPs = fGroups,
  ignoreParents = FALSE,
  TPs = NULL,
  MSPeakLists = NULL,
  formulas = NULL,
  compounds = NULL,
  minRTDiff = 20,
  specSimParams = getDefSpecSimParams()
)

# S4 method for featureGroupsSet
generateComponentsTPs(
  fGroups,
  fGroupsTPs = fGroups,
  ignoreParents = FALSE,
  TPs = NULL,
  MSPeakLists = NULL,
  formulas = NULL,
  compounds = NULL,
  minRTDiff = 20,
  specSimParams = getDefSpecSimParams()
)

Arguments

fGroups
featureGroups object for which components should be generated.

For generateComponentsTPs: also see the fGroupsTPs argument.

algorithm
A character string describing the algorithm that should be used: "ramclustr", "camera", "nontarget", "intclust", "openms", "cliquems", "specclust", "tp"

...
Any parameters to be passed to the selected component generation algorithm.

For sets workflows: further arguments directly passed to the non-sets method.

ionization
Which ionization polarity was used to generate the data: should be "positive" or "negative". If the featureGroups object has adduct annotations, and ionization=NULL, the ionization will be detected automatically.

(sets workflow) This parameter is not supported for sets workflows, as the ionization will always be detected automatically.

onlyIsotopes
Logical value. If TRUE only isotopes are considered when generating components (faster). Corresponds to quick argument of CAMERA::annotate.

minSize
The minimum size of a component. Smaller components than this size will be removed. For RAMClustR: sets the minModuleSize argument to ramclustR. See note below.

relMinReplicates
Feature groups within a component are only kept when they contain data for at least this (relative) amount of replicate analyses. For instance, 0.5 means that at least half of the replicates should contain data for a particular feature group in a component. In this calculation replicates that are fully absent within a component are not taken in to account. See note below.

extraOpts
Named character vector with extra arguments directly passed to CAMERA::annotate (generateComponentsCAMERA) or as extra command line parameters to MetaboliteAdductDecharger (generateComponentsCAMERA). Set to NULL to ignore.

maxCharge, maxGrade, ppm
Arguments passed to cliqueMS::getIsotopes and/or cliqueMS::getAnnotation.

adductInfo
Sets the adinfo argument to cliqueMS::getAnnotation. If NULL then the default adduct information from cliqueMS is used (i.e. the positive.adinfo/negative.adinfo package datasets).

relMinAdductAbundance
The minimum relative abundance (0-1) that an adduct should be assigned to features within the same feature group. See the Feature components section for more details.

adductConflictsUsePref
If set to TRUE, and not all adduct assigments to the features within a feature group are equal and at least one of those adducts is a preferential adduct (prefAdducts argument), then only the features with (the lowest ranked) preferential adduct are considered. In all other cases or when adductConflictsUsePref=FALSE only features with the most frequently assigned adduct is considered. See the Feature components section for more details.

NMConflicts
The strategies to employ when not all neutral masses within a component are equal. Valid options are: "preferential", "mostAbundant" and "mostIntense". Multiple strategies are possible, and will be executed in the given order until one succeeds. See the Feature components section for more details.

prefAdducts
A character vector with one or more preferential adducts. See the Feature components section for more details.

extraOptsCli, extraOptsIso, extraOptsAnn
Named list with further arguments to be passed to cliqueMS::getCliques, cliqueMS::getIsotopes and cliqueMS::getAnnotation, respectively. Set to NULL to ignore.

parallel
If set to TRUE then code is executed in parallel through the futures package. Please see the parallelization section in the handbook for more details.

method
Clustering method that should be applied (passed to fastcluster::hclust).

metric
Distance metric used to calculate the distance matrix (passed to daisy).

normFunc, average
Passed to as.data.table to perform normalization and averaging of data.

maxTreeHeight, deepSplit, minModuleSize
Arguments used by cutreeDynamicTree.

rtRange
A numeric vector containing the minimum and maximum retention time (in seconds) between homologues. Series are always considered from low to high m/z, thus, a negative minimum retention time allows detection of homologous series with increasing m/z and decreasing retention times. These values set the minrt and maxrt arguments of homol.search.

mzRange
A numeric vector specifying the minimum and maximum m/z increment of a homologous series. Sets the minmz and maxmz arguments of homol.search.

elements
A character vector with elements to be considered for detection of repeating units. Sets the elements argument of homol.search function.

rtDev, absMzDev, relMzDev
Maximum deviation for retention time or absolute/relative m/z.

For generateComponentsRAMClustR: Sets the mzabs.error and ppm.error arguments to do.findmain.

For generateComponentsNontarget: Sets the rttol and mztol arguments of homol.search.

For generateComponentsOpenMS: Sets the algorithm:MetaboliteFeatureDeconvolution:mass_max_diff option.

absMzDevLink
Maximum absolute m/z deviation when linking series. This should usually be a bit higher than absMzDev to ensure proper linkage.

traceHack
Currently homol.search does not work with R >3.3.3. This flag, which is enabled by default on these R versions, implements a (messy) workaround (more details here).

chargeMin, chargeMax
The minimum/maximum charge to consider. Corresponds to the algorithm:MetaboliteFeatureDeconvolution:charge_min/algorithm:MetaboliteFeatureDeconvolution:charge_min options.

chargeSpan
The maximum charge span for a single analyte. Corresponds to algorithm:MetaboliteFeatureDeconvolution:charge_span_max.

qTry
Sets how charges are determined. Corresponds to algorithm:MetaboliteFeatureDeconvolution:q_try. Valid options are "heuristic" and "all" (the "feature" option from OpenMS is currently not supported).

potentialAdducts
The adducts to consider. Should be a numeric vector with probabilities for each adduct, e.g. potentialAdducts=c("[M+H]+" = 0.8, "[M+Na]+" = 0.2). Note that the sum of probabilities should always be 1. Furthermore, note that additions of multiple adducts should be controlled by the chargeMin/chargeMax arguments (and not with potentialAdducts), e.g. if chargeMax=2 then both [M+H]+ and [2M+H]2+ may be considered. Please see the algorithm:MetaboliteFeatureDeconvolution:potential_adducts option of MetaboliteAdductDecharger for more details. If NULL then the a default is chosen with defaultOpenMSAdducts (which is not the same as OpenMS).

(sets workflow) Should be a list where each entry specifies the potential adducts for a set. Should either be named with the sets names or follow the same order as sets(fGroups). Example: potentialAdducts=list(positive=c("[M+H]+" = 0.8, "[M+Na]+" = 0.2), negative=c("[M-H]-" = 0.8, "[M-H2O-H]-" = 0.2))

minRTOverlap, retWindow
Sets feature retention tolerances when grouping features. Sets the "algorithm:MetaboliteFeatureDeconvolution:retention_max_diff" and algorithm:MetaboliteFeatureDeconvolution:min_rt_overlap options.

st, sr, maxt, hmax, normalize
Arguments to tune the behaviour of feature group clustering. See their documentation from ramclustR. When st is NULL it will be automatically calculated as the half of the median for all chromatographic peak widths.

RCExperimentVals
A named list containing two more lists: design and instrument. These are used to construct the ExpDes argument passed to ramclustR.

extraOptsRC, extraOptsFM
Named list with further arguments to be passed to ramclustR and do.findmain. Set to NULL to ignore.

MSPeakLists
The MSPeakLists object for the given feature groups that should be used for MS spectral similarity calculations. For generateComponentsTPs it can be NULL, in which case no calculations are performed.

specSimParams
A named list with parameters that influence the calculation of MS spectra similarities. See the spectral similarity parameters documentation for more details.

fGroupsTPs
A featureGroups object containing the feature groups that are expected to be transformation products. If a distinction between parents and TPs is not yet known, fGroupsTPs should equal the fGroups argument. Otherwise, fGroups should only contain the parent feature groups, and both fGroups and fGroupsTPs must be a subset of the same featureGroups object.

ignoreParents
If TRUE then feature groups present in both fGroups and fGroupsTPs are not considered as TPs.

TPs
A transformationProducts object. Set to NULL to perform linking without this data.

formulas, compounds
A formulas/compounds object to calculate annotation similarities between parents and TPs. If NULL then this data is not calculated. For more details see the Linking parents and transformation products section below.

minRTDiff
Minimum retention time (in seconds) difference between the parent and a TP to determine whether a TP elutes prior/after the parent (to calculate retDir values, see Details in componentsTPs))

Value

A components (derived) object containing all generated components.

The algorithms based on 'feature components' return a componentsFeatures derived object.

generateComponentsIntClust and generateComponentsSpecClust return objects derived from componentsSpecClust.

generateComponentsTPs returns objects derived from componentsTPs.

Details

Several algorithms are provided to group feature groups that are related in some (chemical) way to each other. These components generally include adducts, isotopes, in-source fragments and homologues. The linking of this data is generally useful to provide more information for compound annotation and reduce the data size and thus complexity.

generateComponents is a generic function that will generate components using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateComponentsRAMClustR and generateComponentsNontarget. While these functions may be called directly, generateComponents provides a generic interface and is therefore usually preferred.

generateComponentsCAMERA provides an interface to CAMERA which is used to generate components from known adducts, isotopes and in-source fragments. The specified featureGroups object is automatically converted to an xcmsSet object using getXCMSSet.

generateComponentsCliqueMS uses cliqueMS to generate components using the cliqueMS::getCliques function. The grouping of features in each component ('clique') is based on high similarity of chromatographic elution profiles. All features in each component are then annotated with the cliqueMS::getIsotopes and cliqueMS::getAnnotation functions.

generateComponentsIntClust generates components based on intensity profiles of feature groups. Hierarchical clustering is performed on normalized (and optionally replicate averaged) intensity data and the resulting dendrogram is automatically cut with cutreeDynamicTree. The distance matrix is calculated with daisy and clustering is performed with fastcluster::hclust. The clustering of the resulting components can be further visualized and modified using the methods defined for componentsIntClust.

generateComponentsNontarget uses the nontarget R package to generate components by unsupervised detection of homologous series. In the first step the homol.search function is used to detect all homologues within each replicate group (analyses within each replicate group are averaged prior to detection). Then, homologous series across replicate groups are merged in case of full overlap or when merging of partial overlapping series causes no conflicts.

generateComponentsOpenMS uses the MetaboliteAdductDecharger utility (see http://www.openms.de) to generate components. Features that show highly similar chromatographic elution profiles are grouped, and subsequently annotated with their adducts.

defaultOpenMSAdducts returns the default adducts and their probabilities when the OpenMS algorithm is used for componentization. See the potentialAdducts argument for more details.

generateComponentsRAMClustR uses RAMClustR to generate components from feature groups which follow similar chromatographic retention profiles, but are not necessarily restricted to known rules (e.g. adducts or isotopes). This method uses the ramclustR functions for generating the components, whereas do.findmain is used for annotation.

generateComponentsSpecClust generates components based on MS/MS similarity between feature groups. The similarities are converted to a distance matrix and used as input for hierarchical clustering, and the resulting dendrogram is automatically cut with cutreeDynamicTree. The clustering is performed with fastcluster::hclust.

generateComponentsTPs generates components by linking feature groups of transformation products and their parents. Moreover, this method typically employs data from generated transformation products to find parents and their TPs. However, this data is not necessary, and components can also be made based on MS/MS similarity and/or other annotation similarities between the parent and its TPs. For more details see the Linking parents and transformation products section below.

Note

For generateComponentsCAMERA and generateComponentsRAMClustR: the minSize and relMinReplicates arguments provide additional filtering functionality not provided by CAMERA or RAMClustR (except minSize). Note that these filters are enabled by default, hence, final results may be different than what CAMERA/RAMClustR normally would return.

The shift parameter of specSimParams is ignored by generateComponentsTPs, since it always calculates similarities with all supported options.

Feature components

Components resulting from generateComponentsCliqueMS and generateComponentsOpenMS are based on so called feature components. Unlike other algorithms, components are first made on a feature level (per analysis), instead of for complete feature groups. In the final step the feature components are converted to 'regular' components by employing a consensus approach with the following steps:

If an adduct assigned to a feature only occurs as a minority compared to other adduct assigments within the same feature group, it is considered as an outlier and removed accordingly (controlled by the relMinAdductAbundance argument).

For features within a feature group, only keep their adduct assignment if it occurs as the most frequent or is preferential (controlled by adductConflictsUsePref and prefAdducts arguments).

Components are made by combining the feature groups for which at least one of their features are jointly present in the same feature component.

Conflicts of neutral mass assignments within a component (i.e. not all are the same) are dealt with. Firstly, all feature groups with an unknown neutral mass are split in another component. Then, if conflicts still occur, the feature groups with similar neutral mass (determined by absMzDev argument) are grouped. Depending on the NMConflicts argument, the group with one or more preferential adduct(s) or that is the largest or most intense is selected, whereas others are removed from the component. In case multiple groups contain preferential adducts, and >1 preferential adducts are available, the group with the adduct that matches first in prefAdducts 'wins'. In case of ties, one of the next strategies in NMConflicts is tried.

If a feature group occurs in multiple components it will be removed completely.

the minSize filter is applied.

Sets workflows

In a sets workflow the componentization data is generated differently depending on the used algorithm:

The componentization algorithms that support annotation of adducts/isotopes, e.g. RAMClustR, CAMERA and OpenMS, first perform componentization for each set independently. This is necessary, because e.g. adducts are specific to the MS ionization mode that was used. The resulting components are then all combined in a componentsSet object. Note that the components themselves are never merged. The components are renamed to include the set name for which they were generated (e.g. "CMP1" becomes "CMP1-positive").

For componentization with nontarget a similar strategy is applied as above. However, the output is stored in a componentsNTSet object, which supports additional methods such as plotGraph.

For the componentization based on hierarchical clustering, i.e. the "intclust" and "specclust" algorithms, the output format is not different than for a non-sets workflow. However, for intensity based clustering, normalization of feature intensities occurs per set. For MS/MS based clustering, spectral similarities from each sets are combined as is described for the spectrumSimilarity method for sets workflows.

For componentization of transformation products, the output class is the same (componentsTPs). However, the component tables are amended with extra information such as overall/specific set spectrum similarities. As sets data is mixed, transformation products are able to be linked with a parent, even if they were not measured in the same set.

Parallelization

generateComponentsOpenMS uses multiprocessing to parallelize computations. Please see the parallelization section in the handbook for more details and patRoon options for configuration options.

Note that for caching purposes, the analyses files must always exist on the local host computer, even if it is not participating in computations.

Linking parents and transformation products

With generateComponentsTPs, each component consists of feature groups that are considered to be transformation products for one parent (the parent that 'belongs' to the component can be retrieved with the componentInfo method). The parent feature groups are taken from the fGroups parameter, while the feature groups for TPs are taken from fGroupsTPs. If a feature group occurs in both variables, it may therefore be considered as both a parent or TP.

If transformation product data is given, i.e. the TPs argument is set, then a suspect screening of the TPs must be performed in advance (see screenSuspects and convertToSuspects to create the suspect list). Furthermore, if TPs were generated with generateTPsBioTransformer or generateTPsLibrary then the suspect screening must also include the parents (e.g. by setting includeParents=TRUE when calling convertToSuspects or by amending results by setting amend=TRUE to screenSuspects). The suspect screening is necessary for the componentization algorithm to map the feature groups of the parent or TP. If the the suspect screening yields multiple TP hits, all will be reported. Similarly, if the suspect screening contains multiple hits for a parent, a component is made for each of the parent hits.

In case no transformation product data is provided (TPs=NULL), the componentization algorithm simply assumes that each feature group from fGroupsTPs is a potential TP for every parent feature group in fGroups. For this reason, it is highly recommended to specify which feature groups are parents/TPs (see the fGroupsTPs argument description above) and crucial that the data is post-processed, for instance by only retaining TPs that have high annotation similarity with their parents (see the filter method for componentsTPs).

An typical way to distinguish which feature groups are parents or TPs from two different (groups of) samples is by calculating Fold Changes (see the as.data.table method for feature groups and plotVolcano). Of course, other statistical techniques from R are also suitable.

During componentization, several characteristics are calculated which may be useful for post-processing:

specSimilarity: the MS/MS spectral similarity between the feature groups of the TP and its parent (0-1).

specSimilarityPrec,specSimilarityBoth: as specSimilarity, but calculated with binned data using the "precursor" and "both" method, respectively (see MS spectral similarity parameters for more details).

fragmentMatches The number of MS/MS fragment formula annotations that overlap between the TP and parent. If both the formulas and compounds arguments are specified then the annotation data is pooled prior to calculation. Note that only unique matches are counted. Furthermore, note that annotations from all candidates are considered, even if the formula/structure of the parent/TP is known. Hence, fragmentMatches is mainly useful when little or no chemical information is known on the parents/TPs, i.e., when TPs=NULL or originates from generateTPsLogic. Since annotations for all candidates are used, it is highly recommended that the annotation objects are first processed with the filter method, for instance, to select only the top ranked candidates.

neutralLossMatches As fragmentMatches, but counting overlapping neutral loss formulae.

retDir The retention time direction of the TP relative to its parent. See Details in componentsTPs. If TP data was specified, the expected direction is stored in TP_retDir.

retDiff,mzDiff,formulaDiff The retention time, m/z and formula difference between the parent and TP (latter only available if data TP formula is available).

References

Schollee JE, Bourgin M, von Gunten U, McArdell CS, Hollender J (2018). “Non-target screening to trace ozonation transformation products in a wastewater treatment train including different post-treatments.” Water Research, 142, 267--278. doi: 10.1016/j.watres.2018.05.045 .

Daniel Müllner (2013). fastcluster: Fast Hierarchical, Agglomerative Clustering Routines for R and Python. Journal of Statistical Software, 53(9), 1-18. URL https://www.jstatsoft.org/v53/i09/.

Kuhl, C., Tautenhahn, R., Boettcher, C., Larson, T. R. and Neumann, S. CAMERA: an integrated strategy for compound spectra extraction and annotation of liquid chromatography/mass spectrometry data sets. Analytical Chemistry, 84:283-289 (2012)

Senan O, Aguilar-Mogas A, Navarro M, Capellades J, Noon L, Burks D, Yanes O, Guimera R, Sales-Pardo M (2019). “CliqueMS: a computational tool for annotating in-source metabolite ions from LC-MS untargeted metabolomics data based on a coelution similarity network.” Bioinformatics, 35(20), 4089--4097. doi: 10.1093/bioinformatics/btz207 .

Martin Loos (2016). nontarget: Detecting Isotope, Adduct and Homologue Relations in LC-MS Data. R package version 1.9.

Loos, M., Gerber, C., Corona, F., Hollender, J., Singer, H. (2015). Accelerated isotope fine structure calculation using pruned transition trees, Analytical Chemistry 87(11), 5738-5744.

Bielow C, Ruzek S, Huber CG, Reinert K (2010). “Optimal Decharging and Clustering of Charge Ladders Generated in ESI-MS.” Journal of Proteome Research, 9(5), 2688--2695. doi: 10.1021/pr100177k .

Broeckling, Heuberger CD;, Prince AL;, Ingelsson JA;, Prenni E;, E. J (2013). “Assigning precursor-product ion relationships in indiscriminant MS/MS data from non-targeted metabolite profiling studies.” Analytical Chemistry, 9, 33-43.

Broeckling CD, Afsar FA, Neumann S, Ben-Hur A, Prenni JE (2014). “RAMClust: A Novel Feature Clustering Method Enables Spectral-Matching-Based Annotation for Metabolomics Data.” Analytical Chemistry, 86 (14), 6812–6817.

Author

Rick Helmus <r.helmus@uva.nl> and Bas van de Velde (major contributions to spectral binning and similarity calculation).