Components based on clustered intensity profiles.

This class is derived from components and is used to store hierarchical clustering information from intensity profiles of feature groups.

# S4 method for componentsIntClust
clusters(obj)

# S4 method for componentsIntClust
cutClusters(obj)

# S4 method for componentsIntClust
clusterProperties(obj)

# S4 method for componentsIntClust
treeCut(obj, k = NULL, h = NULL)

# S4 method for componentsIntClust
treeCutDynamic(obj, maxTreeHeight, deepSplit, minModuleSize)

# S4 method for componentsIntClust
plotHeatMap(
  obj,
  interactive = FALSE,
  col = NULL,
  margins = c(6, 2),
  cexCol = 1,
  ...
)

# S4 method for componentsIntClust
plotInt(obj, index, pch = 20, type = "b", lty = 3, col = NULL, ...)

# S4 method for componentsIntClust,ANY
plot(
  x,
  pal = "Paired",
  numericLabels = TRUE,
  colourBranches = length(x) < 50,
  showLegend = length(x) < 20,
  ...
)

# S4 method for componentsIntClust
plotSilhouettes(obj, kSeq, pch = 16, type = "b", ...)

Arguments

k, h
Desired number of clusters or tree height to be used for cutting the dendrogram, respectively. One or the other must be specified. Analogous to cutree.

maxTreeHeight, deepSplit, minModuleSize
Arguments used by cutreeDynamicTree.

interactive
If TRUE an interactive heatmap will be drawn (with heatmaply).

col
The colour used for plotting. Set to NULL for automatic colours.

margins, cexCol
Passed to heatmap.2

...
Further options passed to heatmap.2 / heatmaply (plotHeatMap), plot.dendrogram (plot) or plot (plotInt).

index
Numeric component/cluster index.

pch, type
Passed to plot.

lty
Passed to lines.

x, obj
A componentsIntClust object.

pal
Colour palette to be used from RColorBrewer.

numericLabels
Set to TRUE to label with numeric indices instead of (long) feature group names.

colourBranches
Whether branches from cut clusters (and their labels) should be coloured. Might be slow with large numbers of clusters, hence, the default is only TRUE when this is not the case.

showLegend
If TRUE and colourBranches is also TRUE then a legend will be shown which outlines cluster numbers and their colours. By default TRUE for small amount of clusters to avoid overflowing the plot.

kSeq
An integer vector containing the sequence that should be used for average silhouette width calculation.

Value

plotHeatMap returns the same as heatmap.2 or heatmaply.

Details

Objects from this class are generated by generateComponentsIntClust

Methods (by generic)

clusters: Accessor method to the clust slot, which was generated by hclust.

cutClusters: Accessor method to the cutClusters slot. Returns a vector with cluster membership for each candidate (format as cutree).

clusterProperties: Returns a list with properties on how the clustering was performed.

treeCut: Manually (re-)cut the dendrogram.

treeCutDynamic: Automatically (re-)cut the dendrogram using the cutreeDynamicTree function from dynamicTreeCut.

plotHeatMap: draws a heatmap using the heatmap.2 or heatmaply function.

plotInt: makes a plot for all (normalized) intensity profiles of the feature groups within a given cluster.

plot: generates a dendrogram from a given cluster object and optionally highlights resulting branches when the cluster is cut.

plotSilhouettes: Plots the average silhouette width when the clusters are cut by a sequence of k numbers. The k value with the highest value (marked in the plot) may be considered as the optimal number of clusters.

Slots

clusterm

Numeric matrix with normalized feature group intensities that was used for clustering.

distm

Distance matrix that was used for clustering (obtained with daisy).

clust

Object returned by hclust.

cutClusters

A list with assigned clusters (same format as what cutree returns).

gInfo

The groupInfo of the feature groups object that was used.

properties

A list containing general properties and parameters used for clustering.

Note

The intensity values for components (used by plotSpec) are set to a dummy value (1) as no single intensity value exists for this kind of components.

References

Schollee JE, Bourgin M, von Gunten U, McArdell CS, Hollender J (2018-oct). “Non-target screening to trace ozonation transformation products in a wastewater treatment train including different post-treatments.” Water Research, 142, 267--278. doi: 10.1016/j.watres.2018.05.045 .

See also