Components based on clustered intensity profiles.

This class is derived from componentsClust and is used to store hierarchical clustering information from intensity profiles of feature groups.

# S4 method for componentsIntClust
plotHeatMap(
  obj,
  interactive = FALSE,
  col = NULL,
  margins = c(6, 2),
  cexCol = 1,
  ...
)

# S4 method for componentsIntClust
plotInt(obj, index, pch = 20, type = "b", lty = 3, col = NULL, ...)

Arguments

obj
A componentsIntClust object.

interactive
If TRUE an interactive heatmap will be drawn (with heatmaply).

col
The colour used for plotting. Set to NULL for automatic colours.

margins, cexCol
Passed to heatmap.2

...
Further options passed to heatmap.2 / heatmaply (plotHeatMap), plot (plotInt).

index
Numeric component/cluster index.

pch, type, lty
Passed to lines.

Value

plotHeatMap returns the same as heatmap.2 or heatmaply.

Details

Objects from this class are generated by generateComponentsIntClust

Methods (by generic)

plotHeatMap: draws a heatmap using the heatmap.2 or heatmaply function.

plotInt: makes a plot for all (normalized) intensity profiles of the feature groups within a given cluster.

Slots

clusterm

Numeric matrix with normalized feature group intensities that was used for clustering.

Note

When the object is altered (e.g. by filtering or subsetting it), methods that need the original clustered data such as plotting methods do not work anymore and stop with an error.

S4 class hierarchy

componentsClust

componentsIntClust

See also