Components class for homologous series.

This class is derived from components and is used to store results from unsupervised homolog detection with the nontarget package.

# S4 method for componentsNT
plotGraph(obj, onlyLinked)

# S4 method for componentsNTSet
plotGraph(obj, onlyLinked, set)

# S4 method for componentsNTSet
unset(obj, set)

Arguments

obj

The componentsNT object to plot.

onlyLinked

If TRUE then only components with links are plotted.

set

(sets workflow) The name of the set.

Value

plotGraph returns the result of visNetwork.

Details

Objects from this class are generated by generateComponentsNontarget

Methods (by generic)

plotGraph: Plots an interactive network graph for linked homologous series (i.e. series with (partial) overlap which could not be merged). The resulting graph can be browsed interactively and allows quick inspection of series which may be related. The graph is constructed with the igraph package and rendered with visNetwork.

Slots

homol

A list with homol objects for each replicate group as returned by homol.search

Sets workflows

The componentsNTSet class is applicable for sets workflows. This class is derived from componentsNT and therefore largely follows the same user interface.

The following methods are specifically defined for sets workflows:

unset Converts the object data for a specified set into a 'non-set' object (componentsNTUnset), which allows it to be used in 'regular' workflows. Only the components in the specified set are kept. Furthermore, the component names are restored to non-set specific names (see generateComponents for more details).

The following methods are changed or with new functionality:

plotGraph Currently can only create graph networks from one set (specified by the set argument).

Note that the componentsNTSet class does not have a homol slot. Instead, the setObjects method can be used to access this data for a specific set.

References

Martin Loos (2016). nontarget: Detecting Isotope, Adduct and Homologue Relations in LC-MS Data. R package version 1.9.

Loos, M., Gerber, C., Corona, F., Hollender, J., Singer, H. (2015). Accelerated isotope fine structure calculation using pruned transition trees, Analytical Chemistry 87(11), 5738-5744.

Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. https://igraph.org

Almende B.V. and Contributors, Benoit Thieurmel and Titouan Robert (2021). visNetwork: Network Visualization using 'vis.js' Library. R package version 2.1.0. http://datastorm-open.github.io/visNetwork/

See also