Automatic compound identification

Functionality to automatically identify chemical compounds from feature groups.

# S4 method for featureGroups
generateCompounds(fGroups, MSPeakLists, algorithm, ...)

compoundScorings(
  algorithm = NULL,
  database = NULL,
  includeSuspectLists = TRUE,
  onlyDefault = FALSE,
  includeNoDB = TRUE
)

# S4 method for featureGroups
generateCompoundsMetFrag(
  fGroups,
  MSPeakLists,
  method = "CL",
  timeout = 300,
  timeoutRetries = 2,
  errorRetries = 2,
  topMost = 100,
  dbRelMzDev = 5,
  fragRelMzDev = 5,
  fragAbsMzDev = 0.002,
  adduct = NULL,
  database = "pubchem",
  extendedPubChem = "auto",
  chemSpiderToken = "",
  scoreTypes = compoundScorings("metfrag", database, onlyDefault = TRUE)$name,
  scoreWeights = 1,
  preProcessingFilters = c("UnconnectedCompoundFilter", "IsotopeFilter"),
  postProcessingFilters = c("InChIKeyFilter"),
  maxCandidatesToStop = 2500,
  identifiers = NULL,
  extraOpts = NULL
)

# S4 method for featureGroupsSet
generateCompoundsMetFrag(
  fGroups,
  MSPeakLists,
  method = "CL",
  timeout = 300,
  timeoutRetries = 2,
  errorRetries = 2,
  topMost = 100,
  dbRelMzDev = 5,
  fragRelMzDev = 5,
  fragAbsMzDev = 0.002,
  adduct = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0
)

# S4 method for featureGroups
generateCompoundsSIRIUS(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  elements = "CHNOP",
  profile = "qtof",
  formulaDatabase = NULL,
  fingerIDDatabase = "pubchem",
  noise = NULL,
  errorRetries = 2,
  cores = NULL,
  topMost = 100,
  topMostFormulas = 5,
  extraOptsGeneral = NULL,
  extraOptsFormula = NULL,
  verbose = TRUE,
  splitBatches = FALSE
)

# S4 method for featureGroupsSet
generateCompoundsSIRIUS(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = NULL,
  ...,
  setThreshold = 0,
  setThresholdAnn = 0
)

Arguments

fGroups
featureGroups object for which compounds should be identified. This should be the same or a subset of the object that was used to create the specified MSPeakLists. In the case of a subset only the remaining feature groups in the subset are considered.

MSPeakLists
A MSPeakLists object that was generated for the supplied fGroups.

algorithm
A character string describing the algorithm that should be used: "metfrag", "sirius"

...
Any parameters to be passed to the selected compound generation algorithm.

For sets workflows: further arguments directly passed to the non-sets method.

database
Compound database to use. Valid values are: "pubchem", "chemspider", "for-ident", "comptox", "pubchemlite", "kegg", "sdf", "psv" and "csv". See section below for more information. Sets the MetFragDatabaseType option.

includeSuspectLists, onlyDefault, includeNoDB
A logical specifying whether scoring terms related to suspect lists, default scoring terms and non-database specific scoring terms should be included in the output, respectively.

method
Which method should be used for MetFrag execution: "CL" for MetFragCL and "R" for MetFragR. The former might be faster.

timeout
Maximum time (in seconds) before a metFrag query for a feature group is stopped. Also see timeoutRetries argument.

timeoutRetries
Maximum number of retries after reaching a timeout before completely skipping the metFrag query for a feature group. Also see timeout argument.

errorRetries
Maximum number of retries after an error occurred. This may be useful to handle e.g. connection errors.

topMost
Only keep this number of candidates (per feature group) with highest score. Set to NULL to always keep all candidates, however, please note that this may result in significant usage of CPU/RAM resources for large numbers of candidates.

dbRelMzDev
Relative mass deviation (in ppm) for database search. Sets the DatabaseSearchRelativeMassDeviation option.

fragRelMzDev
Relative mass deviation (in ppm) for fragment matching. Sets the FragmentPeakMatchRelativeMassDeviation option.

fragAbsMzDev
Absolute mass deviation (in Da) for fragment matching. Sets the FragmentPeakMatchAbsoluteMassDeviation option.

adduct
An adduct object (or something that can be converted to it with as.adduct). Examples: "[M-H]-", "[M+Na]+". If the featureGroups object has adduct annotations then these are used if adducts=NULL.

(sets workflow) The adduct argument is not supported for sets workflows, since the adduct annotations will then always be used.

extendedPubChem
If database="pubchem": whether to use the extended database that includes information for compound scoring (i.e. number of patents/PubMed references). Note that downloading candidates from this database might take extra time. Valid values are: FALSE (never use it), TRUE (always use it) or "auto" (default, use if specified scorings demand it).

chemSpiderToken
A character string with the ChemSpider security token that should be set when the ChemSpider database is used. Sets the ChemSpiderToken option.

scoreTypes
A character vector defining the scoring types. See the Scorings section below for more information. Note that both generic and MetFrag specific names are accepted (i.e. name and metfrag columns returned by compoundScorings). When a local database is used, the name should match what is given there (e.g column names when database=csv). Note that MetFrag may still report other scoring data, however, these are not used for ranking. Sets the MetFragScoreTypes option.

scoreWeights
Numeric vector containing weights of the used scoring types. Order is the same as set in scoreTypes. Values are recycled if necessary. Sets the MetFragScoreWeights option.

preProcessingFilters, postProcessingFilters
A character vector defining pre/post filters applied before/after fragmentation and scoring (e.g. "UnconnectedCompoundFilter", "IsotopeFilter", "ElementExclusionFilter"). Some methods require further options to be set. For all filters and more information refer to the Candidate Filters section on the MetFragR homepage. Sets the MetFragPreProcessingCandidateFilter and MetFragPostProcessingCandidateFilter options.

maxCandidatesToStop
If more than this number of candidate structures are found then processing will be aborted and no results this feature group will be reported. Low values increase the chance of missing data, whereas too high values will use too much computer resources and signficantly slowdown the process. Sets the MaxCandidateLimitToStop option.

identifiers
A list containing for each feature group a character vector with database identifiers that should be used to find candidates for a feature group (the list should be named by feature group names). If NULL all relevant candidates will be retrieved from the specified database. An example usage scenario is to obtain the list of candidate identifiers from a compounds object obtained with generateCompoundsSIRIUS using the identifiers method. This way, only those candidates will be searched by MetFrag that were generated by SIRIUS+CSI:FingerID. Sets the PrecursorCompoundIDs option.

extraOpts
For MetFrag: A named list containing further settings to be passed to run.metfrag. See the MetFragR and MetFrag CL homepages for all available options.

For SIRIUS: a character vector with any extra commandline parameters for formula prediction. See the SIRIUS manual for more details.

Set to NULL to ignore.

setThreshold
(sets workflow) Minimum abundance for a candidate among all sets (0-1). For instance, a value of 1 means that the candidate needs to be present in all the set data.

setThresholdAnn
(sets workflow) As setThreshold, but only taking into account the set data that contain annotations for the feature group of the candidate.

relMzDev
Maximum relative deviation between the measured and candidate formula m/z values (in ppm). Sets the --ppm-max commandline option.

elements
Elements to be considered for formulae calculation. This will heavily affects the number of candidates! Always try to work with a minimal set by excluding elements you don't expect. The minimum/maximum number of elements can also be specified, for example: a value of "C[5]H[10-15]O" will only consider formulae with up to five carbon atoms, between ten and fifteen hydrogen atoms and any amount of oxygen atoms. Sets the --elements commandline option.

profile
Name of the configuration profile, for example: "qtof", "orbitrap", "fticr". Sets the --profile commandline option.

formulaDatabase
If not NULL, use a database for retrieval of formula candidates. Possible values are: "pubchem", "bio", "kegg", "hmdb". Sets the --database commandline option.

fingerIDDatabase
Database specifically used for CSI:FingerID. If NULL, the value of the formulaDatabase parameter will be used or "pubchem" when that is also NULL. Sets the --fingerid-db option.

noise
Median intensity of the noise (NULL ignores this parameter). Sets the --noise commandline option.

cores
The number of cores SIRIUS will use. If NULL then the default of all cores will be used.

topMostFormulas
Do not return more than this number of candidate formulae. Note that only compounds for these formulae will be searched. Sets the --candidates commandline option.

extraOptsGeneral, extraOptsFormula
a character vector with any extra commandline parameters for SIRIUS. For SIRIUS versions <4.4 there is no distinction between general and formula options. Otherwise commandline options specified in extraOptsGeneral are added prior to the formula command, while options specified in extraOptsFormula are added in afterwards. See the SIRIUS manual for more details. Set to NULL to ignore.

verbose
If TRUE then more output is shown in the terminal.

splitBatches
If TRUE then the calculations done by SIRIUS will be evenly split over multiple SIRIUS calls (which may be run in parallel depending on the set package options). If splitBatches=FALSE then all feature calculations are performed from a single SIRIUS execution, which is often the fastest if calculations are performed on a single computer.

Value

compoundScorings returns a data.frame with information on which scoring terms are used, what their algorithm specific name is and other information such as to which database they apply and short remarks.

generateCompoundsMetFrag returns a compoundsMF object.

generateCompoundsSIRIUS returns a compounds object.

Details

Several algorithms are provided to automatically identify compounds for given feature groups. To this end, each measured masses for all feature groups are searched within online database(s) (e.g. PubChem) to retrieve a list of potential candidate chemical compounds. Depending on the algorithm and its parameters, further scoring of candidates is then performed using, for instance, matching of measured and theoretical isotopic patterns, presence within other data sources such as patent databases and similarity of measured and in-silico predicted MS/MS fragments. Note that this process is often quite time consuming, especially for large feature group sets. Therefore, this is often one of the last steps within the workflow and not performed before feature groups have been prioritized.

generateCompounds is a generic function that will generate compounds using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateCompoundsMetFrag and generateCompoundsSIRIUS. While these functions may be called directly, generateCompounds provides a generic interface and is therefore usually preferred.

compoundScorings displays an overview of scorings may be applied to rank candidate compounds (see Scorings section below).

generateCompoundsMetFrag uses the metfRag package or MetFrag CL for compound identification (see http://ipb-halle.github.io/MetFrag/). Several online compound databases such as PubChem and ChemSpider may be chosen for retrieval of candidate structures. In addition, many options exist to score and filter resulting data, and it is highly suggested to optimize these to improve results. While MS/MS data is not mandatory, it will usually greatly improve candidate scoring. The MetFrag options PeakList, IonizedPrecursorMass and ExperimentalRetentionTimeValue (in minutes) fields are automatically set from feature data.

generateCompoundsSIRIUS uses SIRIUS in combination with CSI:FingerID for compound identification. Similar to generateFormulasSIRIUS, candidate formulae are generated with SIRIUS. These results are then feed to CSI:FingerID to acquire candidate structures. This method requires the availability of MS/MS data, and feature groups without it will be ignored.

Note

For annotations performed with SIRIUS it is often the fastest to keep the default splitBatches=FALSE. In this case, all SIRIUS output will be printed to the terminal (unless verbose=FALSE or patRoon.MP.method="future"). Furthermore, please note that only annotations to be performed for the same adduct are grouped in a single batch execution.

Scorings

Each algorithm implements their own scoring system. Their names have been simplified and harmonized where possible and are used for reporting and in the case MetFrag is used to specify how compounds should be scored (scoreTypes argument). The compoundScorings function can be used to get an overview of both the algorithm specific and generic scoring names. For instance, the table below shows all scorings for MetFrag: (some columns are omitted)

name	metfrag	database

score	Score

fragScore	FragmenterScore

metFusionScore	OfflineMetFusionScore

individualMoNAScore	OfflineIndividualMoNAScore

numberPatents	PubChemNumberPatents	pubchem

numberPatents	Patent_Count	pubchemlite

pubMedReferences	PubChemNumberPubMedReferences	pubchem

pubMedReferences	ChemSpiderNumberPubMedReferences	chemspider

pubMedReferences	NUMBER_OF_PUBMED_ARTICLES	comptox

pubMedReferences	PubMed_Count	pubchemlite

extReferenceCount	ChemSpiderNumberExternalReferences	chemspider

dataSourceCount	ChemSpiderDataSourceCount	chemspider

referenceCount	ChemSpiderReferenceCount	chemspider

RSCCount	ChemSpiderRSCCount	chemspider

smartsInclusionScore	SmartsSubstructureInclusionScore

smartsExclusionScore	SmartsSubstructureExclusionScore

suspectListScore	SuspectListScore

retentionTimeScore	RetentionTimeScore

CPDATCount	CPDAT_COUNT	comptox

TOXCASTActive	TOXCAST_PERCENT_ACTIVE	comptox

dataSources	DATA_SOURCES	comptox

pubChemDataSources	PUBCHEM_DATA_SOURCES	comptox

EXPOCASTPredExpo	EXPOCAST_MEDIAN_EXPOSURE_PREDICTION_MG/KG-BW/DAY	comptox

ECOTOX	ECOTOX	comptox

NORMANSUSDAT	NORMANSUSDAT	comptox

MASSBANKEU	MASSBANKEU	comptox

TOX21SL	TOX21SL	comptox

TOXCAST	TOXCAST	comptox

KEMIMARKET	KEMIMARKET	comptox

MZCLOUD	MZCLOUD	comptox

pubMedNeuro	PubMedNeuro	comptox

CIGARETTES	CIGARETTES	comptox

INDOORCT16	INDOORCT16	comptox

SRM2585DUST	SRM2585DUST	comptox

SLTCHEMDB	SLTCHEMDB	comptox

THSMOKE	THSMOKE	comptox

ITNANTIBIOTIC	ITNANTIBIOTIC	comptox

STOFFIDENT	STOFFIDENT	comptox

KEMIMARKET_EXPO	KEMIMARKET_EXPO	comptox

KEMIMARKET_HAZ	KEMIMARKET_HAZ	comptox

REACH2017	REACH2017	comptox

KEMIWW_WDUIndex	KEMIWW_WDUIndex	comptox

KEMIWW_StpSE	KEMIWW_StpSE	comptox

KEMIWW_SEHitsOverDL	KEMIWW_SEHitsOverDL	comptox

ZINC15PHARMA	ZINC15PHARMA	comptox

PFASMASTER	PFASMASTER	comptox

peakFingerprintScore	AutomatedPeakFingerprintAnnotationScore

lossFingerprintScore	AutomatedLossFingerprintAnnotationScore

agroChemInfo	AgroChemInfo	pubchemlite

bioPathway	BioPathway	pubchemlite

drugMedicInfo	DrugMedicInfo	pubchemlite

foodRelated	FoodRelated	pubchemlite

pharmacoInfo	PharmacoInfo	pubchemlite

safetyInfo	SafetyInfo	pubchemlite

toxicityInfo	ToxicityInfo	pubchemlite

knownUse	KnownUse	pubchemlite

disorderDisease	DisorderDisease	pubchemlite

identification	Identification	pubchemlite

annoTypeCount	FPSum	pubchemlite

annoTypeCount	AnnoTypeCount	pubchemlite

In addition, the compoundScorings function is also useful to programatically generate a set of scorings to be used by MetFrag. For instance, the following can be given to the scoreTypes argument to use all default scorings for PubChem: compoundScorings("metfrag", "pubchem", onlyDefault=TRUE)$name.

For all MetFrag scoring types refer to the Candidate Scores section on the MetFragR homepage.

Sets workflows

With a sets workflow, annotation is first performed for each set. This is important, since the annotation algorithms typically cannot work with data from mixed ionization modes. The annotation results are then combined to generate a sets consensus:

The annotation tables for each feature group from the set specific data are combined. Rows with overlapping candidates (determined by the first-block InChIKey) are merged.

Set specific data (e.g. the ionic formula) is retained by renaming their columns with set specific names.

The MS/MS fragment annotations (fragInfo column) from each set are combined.

The scorings for each set are averaged to calculate overall scores.

The candidates are re-ranked based on their average ranking among the set data (if a candidate is absent in a set it is assigned the poorest rank in that set).

The coverage of each candidate among sets is calculated. Depending on the setThreshold and setThresholdAnn arguments, candidates with low abundance are removed.

Parallelization

generateCompoundsMetFrag and generateCompoundsSIRIUS uses multiprocessing to parallelize computations. Please see the parallelization section in the handbook for more details and patRoon options for configuration options.

When local database files are used with generateCompoundsMetFrag (e.g. when database is set to "pubchemlite", "csv" etc.) and patRoon.MP.method="future", then the database file must be present on all the nodes. When pubchemlite or comptox is used, the location for these databases can be configured on the host with the respective package options (patRoon.path.MetFragPubChemLite and patRoon.path.MetFragCompTox). Note that these files must also be present on the local host computer, even if it is not participating in computations.

Usage of MetFrag databases

When database="chemspider" setting the chemSpiderToken argument is mandatory.

When a local database is set (i.e. sdf, psv, csv, comptox, pubchemlite) the file location of the database should be set in the LocalDatabasePath value via the extraOpts argument or using the patRoon.path.MetFragCompTox/patRoon.path.MetFragPubChemLite option (only when database="comptox" or database="pubchemlite").

Examples: options(patRoon.path.MetFragCompTox = "C:/CompTox_17March2019_SelectMetaData.csv") extraOpts = list(LocalDatabasePath = "C:/myDB.csv").

For database="comptox" the files can be obtained from here. Furthermore, the files with additions for https://zenodo.org/record/3364464#.XnjM-XLvKUk and https://zenodo.org/record/3472781#.XnjMAHLvKUk metadata are also supported. Note that only recent MetFrag versions (>= 2.4.5) support these libraries.

References

Ruttkies C, Schymanski EL, Wolf S, Hollender J, Neumann S (2016). “MetFrag relaunched: incorporating strategies beyond in silico fragmentation.” Journal of Cheminformatics, 8(1). doi: 10.1186/s13321-016-0115-9 .

Duhrkop K, Fleischauer M, Ludwig M, Aksenov AA, Melnik AV, Meusel M, Dorrestein PC, Rousu J, Bocker S (2019). “SIRIUS 4: a rapid tool for turning tandem mass spectra into metabolite structure information.” Nature Methods, 16(4), 299--302. doi: 10.1038/s41592-019-0344-8 .

Duhrkop K, Bocker S (2015). “Fragmentation Trees Reloaded.” In Przytycka TM (ed.), Research in Computational Molecular Biology, 65--79. ISBN 978-3-319-16706-0.

Duhrkop K, Shen H, Meusel M, Rousu J, Bocker S (2015). “Searching molecular structure databases with tandem mass spectra using CSI:FingerID.” Proceedings of the National Academy of Sciences, 112(41), 12580--12585. doi: 10.1073/pnas.1509788112 .

Bocker S, Letzel MC, Liptak Z, Pervukhin A (2008). “SIRIUS: decomposing isotope patterns for metabolite identification.” Bioinformatics, 25(2), 218--224. doi: 10.1093/bioinformatics/btn603 .

See also