Compound lists class

Contains data of generated chemical compounds for given feature groups.

# S4 method for compounds
compoundTable(obj)

# S4 method for compounds
algorithm(obj)

# S4 method for compounds
length(x)

# S4 method for compounds
show(object)

# S4 method for compounds
identifiers(compounds)

# S4 method for compounds
filter(obj, minExplainedPeaks = NULL, minScore = NULL,
  minFragScore = NULL, minFormulaScore = NULL, topMost = NULL)

# S4 method for compounds
addFormulaScoring(compounds, formConsensus,
  updateScore = FALSE, formulaScoreWeight = 1)

# S4 method for compounds
plotStructure(obj, index, groupName, width = 500,
  height = 500, useGGPlot2 = FALSE)

# S4 method for compounds
plotScores(obj, index, groupName,
  normalizeScores = TRUE, useGGPlot2 = FALSE)

# S4 method for compounds
plotSpec(obj, index, groupName, MSPeakLists,
  formConsensus = NULL, normalizeScores = TRUE, useGGPlot2 = FALSE)

# S4 method for compounds
consensus(obj, ..., compThreshold = 0,
  mergeScoresFunc = sum)

Arguments

obj, object, x, compounds
The compound object.

minExplainedPeaks, minScore, minFragScore, minFormulaScore
Minimal number of explained peaks, overall score, in-silico fragmentation score and formula score, respectively. Set to NULL to ignore.

topMost
Only keep a maximum of topMost candidates with highest score. Set to NULL to ignore.

formConsensus
The formulaConsensus object that should be used for scoring/annotation. For plotSpec: set to NULL to ignore.

updateScore
If set to TRUE then the score column is updated by adding the normalized formulaScore (weighted by formulaScoreWeight). Currently, this only makes sense for MetFrag results!

formulaScoreWeight
Weight used to update scoring (see updateScore parameter).

index
The (numeric) index of the candidate to be plotted.

groupName
The name of the feature group to which the candidate belongs.

width, height
The dimensions (in pixels) of the raster image that should be plotted.

useGGPlot2
If TRUE then ggplot2 is used for plotting, otherwise base plot used. For plotSpec, ggplot2 allows nicely repelled text for annotation. However, base plot is generally faster.

normalizeScores
Normalize scores to maximum values.

MSPeakLists
The MSPeakLists object that was used to generate the candidate

…
compounds objects that should be used to generate the consensus.

compThreshold
Minimal fraction (0-1) that a candidate must be present within all given compounds objects.

mergeScoresFunc
Function used to calculate the total score for all (merged) score columns.

Value

plotSpec and plotStructure will return a ggplot object if useGGPlot2 is TRUE.

compoundTable returns a list containing for each feature group a data.table with an overview of all candidate compounds and other data such as candidate scoring, matched MS/MS fragments, etc.

filter returns a filtered compounds object.

addFormulaScoring returns a compounds object updated with formula scoring.

consensus returns a compounds object that is produced by merging multiple specified compounds objects.

Details

compounds objects are obtained from compound generators.

Methods (by generic)

compoundTable: Accessor method to obtain generated compounds.

algorithm: Accessor method for the algorithm (a character string) used to generate compounds.

length: Obtain total number of candidate compounds.

show: Show summary information for this object.

identifiers: Returns a list containing for each feature group a character vector with database identifiers for all candidate compounds. The list is named by feature group names, and is typically used with the identifiers option of generateCompoundsMetfrag.

filter: Provides rule based filtering for generated compounds. Useful to eliminate unlikely candidates and speed up further processing.

addFormulaScoring: Provides compound scoring data that is based on the presence of candidate formulae being present in a given formulaConsensus object. Matched precursor formulae yield one point, whereas matched MS/MS fragments yield 0.5 point each.

plotStructure: Plots a structure of a candidate compound using the rcdk package.

plotScores: Plots a barplot with scoring of a candidate compound.

plotSpec: Plots an annotated spectrum for a given candidate compound of a feature group.

consensus: Generates a consensus from multiple compounds objects. The compound results will be merged and optionally compounds will be filtered by their occurrence throughout the specified compounds objects.

Slots

compounds

Lists of all generated compounds. Use the compounds method for access.

algorithm

The algorithm that was used for generation of compounds. Use the algorithm method for access.

References

Guha, R. (2007). 'Chemical Informatics Functionality in R'. Journal of Statistical Software 6(18)