Compound lists class

Contains data of generated chemical compounds for given feature groups.

# S4 method for compounds
compoundTable(obj)

# S4 method for compounds
algorithm(obj)

# S4 method for compounds
groupNames(obj)

# S4 method for compounds
length(x)

# S4 method for compounds
show(object)

# S4 method for compounds,ANY,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for compounds,ANY,missing
[[(x, i, j)

# S4 method for compounds
$(x, name)

# S4 method for compounds
as.data.table(
  x,
  fGroups = NULL,
  fragments = FALSE,
  normalizeScores = "none",
  excludeNormScores = c("score", "individualMoNAScore")
)

# S4 method for compounds
identifiers(compounds)

# S4 method for compounds
filter(
  obj,
  minExplainedPeaks = NULL,
  minScore = NULL,
  minFragScore = NULL,
  minFormulaScore = NULL,
  scoreLimits = NULL,
  elements = NULL,
  fragElements = NULL,
  lossElements = NULL,
  topMost = NULL,
  negate = FALSE
)

# S4 method for compounds
addFormulaScoring(
  compounds,
  formulas,
  updateScore = FALSE,
  formulaScoreWeight = 1
)

# S4 method for compounds
getMCS(obj, index, groupName)

# S4 method for compounds
plotStructure(
  obj,
  index,
  groupName,
  width = 500,
  height = 500,
  useGGPlot2 = FALSE
)

# S4 method for compounds
plotScores(
  obj,
  index,
  groupName,
  normalizeScores = "max",
  excludeNormScores = c("score", "individualMoNAScore"),
  onlyUsed = TRUE,
  useGGPlot2 = FALSE
)

# S4 method for compounds
annotatedPeakList(
  obj,
  index,
  groupName,
  MSPeakLists,
  formulas = NULL,
  onlyAnnotated = FALSE
)

# S4 method for compounds
plotSpec(
  obj,
  index,
  groupName,
  MSPeakLists,
  formulas = NULL,
  plotStruct = TRUE,
  title = NULL,
  useGGPlot2 = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

# S4 method for compounds
plotVenn(obj, ..., labels = NULL, vennArgs = NULL)

# S4 method for compounds
plotUpSet(
  obj,
  ...,
  labels = NULL,
  nsets = length(list(...)) + 1,
  nintersects = NA,
  upsetArgs = NULL
)

# S4 method for compounds
consensus(
  obj,
  ...,
  absMinAbundance = NULL,
  relMinAbundance = NULL,
  uniqueFrom = NULL,
  uniqueOuter = FALSE,
  minMaxNormalization = FALSE,
  rankWeights = 1,
  labels = NULL
)

Arguments

obj, object, x, compounds
The compound object.

i
A numeric or character value which is used to select feature groups by their index or name, respectively (for the order/names see groupNames()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all feature groups are selected.

For [[: should be a scalar value.

...
For plotSpec: Further arguments passed to plot.

Others: Any further (and unique) compounds objects.

drop, j
ignored.

name
The feature group name (partially matched).

fGroups
The featureGroups object that was used to generate this object. If not NULL it is used to add feature group information (retention and m/z values).

fragments
If TRUE then information on annotated fragments will be included.

normalizeScores
A character that specifies how normalization of compound scorings occurs. Either "none" (no normalization), "max" (normalize to max value) or "minmax" (perform min-max normalization). Note that normalization of negative scores (e.g. output by SIRIUS) is always performed as min-max. Furthermore, currently normalization takes the original scoring min/max values into account when candidates were generated. Thus, normalization is not affected when candidate results were removed after they were generated (e.g. by use of filter).

excludeNormScores
A character vector specifying any compound scoring names that should not be normalized. Set to NULL to normalize all scorings. Note that whether any normalization occurs is set by the excludeNormScores argument.

For compounds: By default score and individualMoNAScore are set to mimic the behaviour of the MetFrag web interface.

minExplainedPeaks, minScore, minFragScore, minFormulaScore
Minimum number of explained peaks, overall score, in-silico fragmentation score and formula score, respectively. Set to NULL to ignore. The scoreLimits argument allows for more advanced score filtering.

scoreLimits
Filter results by their scores. Should be a named list that contains two-sized numeric vectors with the minimum/maximum value of a score (use -Inf/Inf for no limits). The names of each element should follow the values returned by compoundScorings()$name. For instance, scoreLimits=list(numberPatents=c(10, Inf)) specifies that numberPatents should be at least 10. For more details of scorings see compoundScorings. Note that a result without a specified scoring is never removed. Set to NULL to skip this filter.

elements
Only retain candidate formulae (neutral form) that match a given elemental restriction. The format of elements is a character string with elements that should be present where each element is followed by a valid amount or a range thereof. If no number is specified then 1 is assumed. For instance, elements="C1-10H2-20O0-2P", specifies that 1-10, 2-20, 0-2 and 1 carbon, hydrogen, oxygen and phosphorus atoms should be present, respectively. When length(elements)>1 formulas are tested to follow at least one of the given elemental restrictions. For instance, elements=c("P", "S") specifies that either one phosphorus or one sulphur atom should be present. Set to NULL to ignore this filter.

fragElements, lossElements
Specifies elemental restrictions for fragment or neutral loss formulae (charged form). Candidates are retained if at least one of the fragment formulae follow (or not follow if negate=TRUE) the given restrictions. See elements for the used format.

topMost
Only keep a maximum of topMost candidates with highest score (or least highest if negate=TRUE). Set to NULL to ignore.

negate
If TRUE then filters are applied in opposite manner.

formulas
The formulas object that should be used for scoring/annotation. For plotSpec: set to NULL to ignore.

updateScore
If set to TRUE then the score column is updated by adding the normalized formulaScore (weighted by formulaScoreWeight). Currently, this only makes sense for MetFrag results!

formulaScoreWeight
Weight used to update scoring (see updateScore parameter).

index
The numeric index of the candidate structure. Multiple indices (i.e. vector with length >=2) may be specified for plotStructure and are mandatory for getMCS. Alternatively, -1 may be specified to these methods to select all candidates. When multiple indices are specified for plotStructure, their maximum common substructure will be drawn.

groupName
The name of the feature group to which the candidate belongs.

width, height
The dimensions (in pixels) of the raster image that should be plotted.

useGGPlot2
If TRUE then ggplot2 is used for plotting, otherwise base plot used. For plotSpec, ggplot2 allows nicely repelled text for annotation. However, base plot is generally faster.

onlyUsed
If TRUE then only scorings are plotted that actually have been used to rank data (see the scoreTypes argument to generateCompoundsMetfrag for more details).

MSPeakLists
The MSPeakLists object that was used to generate the candidate

onlyAnnotated
Set to TRUE to filter out any peaks that could not be annotated.

plotStruct
If TRUE then the candidate structure is drawn in the spectrum.

title
The title of the plot. If NULL a title will be automatically made.

xlim, ylim
Sets the plot size limits used by plot. Set to NULL for automatic plot sizing.

labels
A character with names to use for labelling. If NULL labels are automatically generated.

vennArgs
A list with further arguments passed to VennDiagram plotting functions. Set to NULL to ignore.

nsets
See upset.

nintersects
See upset.

upsetArgs
A list with any further arguments to be passed to upset. Set to NULL to ignore.

absMinAbundance, relMinAbundance
Minimum absolute or relative (0-1) abundance across objects for a result to be kept. For instance, relMinAbundance=0.5 means that a result should be present in at least half of the number of compared objects. Set to NULL to ignore and keep all results. Limits cannot be set when uniqueFrom is not NULL.

uniqueFrom
Set this argument to only retain compounds that are unique within one or more of the objects for which the consensus is made. Selection is done by setting the value of uniqueFrom to a logical (values are recycled), numeric (select by index) or a character (as obtained with algorithm(obj)). For logical and numeric values the order corresponds to the order of the objects given for the consensus. Set to NULL to ignore.

uniqueOuter
If uniqueFrom is not NULL and if uniqueOuter=TRUE: only retain data that are also unique between objects specified in uniqueFrom.

minMaxNormalization
Set to TRUE to apply min-max normalization of (merged) scoring columns. FALSE will apply normalization to the maximum value. Scorings with negative values will always be min-max normalized.

rankWeights
A numeric vector with weights of to calulcate the mean ranking score for each candidate. The value will be re-cycled if necessary, hence, the default value of 1 means equal weights for all considered objects.

Value

plotSpec and plotStructure will return a ggplot object if useGGPlot2 is TRUE.

compoundTable returns a list containing for each feature group a data.table with an overview of all candidate compounds and other data such as candidate scoring, matched MS/MS fragments, etc.

filter returns a filtered compounds object.

addFormulaScoring returns a compounds object updated with formula scoring.

getMCS returns an rcdk molecule object (IAtomContainer).

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g. draw.pairwise.venn and draw.triple.venn).

consensus returns a compounds object that is produced by merging multiple specified compounds objects.

Details

compounds objects are obtained from compound generators.

Methods (by generic)

compoundTable: Accessor method to obtain generated compounds.

algorithm: Accessor method for the algorithm (a character string) used to generate compounds.

groupNames: returns a character vector with the names of the feature groups for which data is present in this object.

length: Obtain total number of candidate compounds.

show: Show summary information for this object.

[: Subset on feature groups.

[[: Extract a compound table for a feature group.

$: Extract a compound table for a feature group.

as.data.table: Returns all MS peak list data in a table.

identifiers: Returns a list containing for each feature group a character vector with database identifiers for all candidate compounds. The list is named by feature group names, and is typically used with the identifiers option of generateCompoundsMetfrag.

filter: Provides rule based filtering for generated compounds. Useful to eliminate unlikely candidates and speed up further processing.

addFormulaScoring: Adds formula ranking data from a formulas object as an extra compound candidate scoring (formulaScore column). The formula score for each compound candidate is between 0-1, where zero means no match with any formula candidates, and one means that the compound candidate's formula is the highest ranked.

getMCS: Calculates the maximum common substructure (MCS) for two or more candidate structures for a feature group. This method uses the get.mcs function from rcdk.

plotStructure: Plots a structure of a candidate compound using the rcdk package. If multiple candidates are specified (i.e. by specifying a vector for index) then the maximum common substructure (MCS) of the selected candidates is drawn.

plotScores: Plots a barplot with scoring of a candidate compound.

annotatedPeakList: Returns an MS/MS peak list annotated with data from a given candidate compound for a feature group.

plotSpec: Plots an annotated spectrum for a given candidate compound for a feature group.

plotVenn: plots a Venn diagram (using VennDiagram) outlining unique and shared compound candidates of up to five different compounds objects. Comparison is made on InChIKey1.

plotUpSet: plots an UpSet diagram (using the upset function) outlining unique and shared compound candidates between different compounds objects. Comparison is made on InChIKey1.

consensus: Generates a consensus of results from multiple objects. In order to rank the consensus candidates, first each of the candidates are scored based on their original ranking (the scores are normalized and the highest ranked candidate gets value 1). The (weighted) mean is then calculated for all scorings of each candidate to derive the final ranking (if an object lacks the candidate its score will be 0). The original rankings for each object is stored in the rank columns.

Slots

compounds

Lists of all generated compounds. Use the compounds method for access.

scoreTypes

A character with all the score types that were used when generating the compounds.

scoreRanges

The original min/max values of all scorings when candidate results were generated. This is used for normalization.

S4 class hierarchy

workflowStep

compounds

compoundsConsensus

compoundsMF

Source

Subscripting of formulae for plots generated by plotSpec is based on the chemistry2expression function from the ReSOLUTION package.

References

Guha, R. (2007). 'Chemical Informatics Functionality in R'. Journal of Statistical Software 6(18)

Conway JR, Lex A, Gehlenborg N (2017). “UpSetR: an R package for the visualization of intersecting sets and their properties.” Bioinformatics, 33(18), 2938-2940. doi: 10.1093/bioinformatics/btx364 , http://dx.doi.org/10.1093/bioinformatics/btx364.

Lex A, Gehlenborg N, Strobelt H, Vuillemot R, Pfister H (2014). “UpSet: Visualization of Intersecting Sets.” IEEE Transactions on Visualization and Computer Graphics, 20(12), 1983--1992. doi: 10.1109/tvcg.2014.2346248 , https://doi.org/10.1109/tvcg.2014.2346248.