Compounds cluster class

Objects from this class are used to store hierarchical clustering data of candidate structures within compounds objects.

# S4 method for compoundsCluster
clusters(obj)

# S4 method for compoundsCluster
cutClusters(obj)

# S4 method for compoundsCluster
clusterProperties(obj)

# S4 method for compoundsCluster
length(x)

# S4 method for compoundsCluster
lengths(x, use.names = TRUE)

# S4 method for compoundsCluster
show(object)

# S4 method for compoundsCluster
treeCut(obj, k = NULL, h = NULL, groupName)

# S4 method for compoundsCluster
treeCutDynamic(obj, maxTreeHeight, deepSplit,
  minModuleSize, groupName)

# S4 method for compoundsCluster,ANY
plot(x, groupName, pal = "Paired",
  colourBranches = length(x) < 50, showLegend = length(x) < 20, ...)

# S4 method for compoundsCluster
getMCS(obj, groupName, cluster)

# S4 method for compoundsCluster
plotStructure(obj, groupName, cluster,
  width = 500, height = 500, withTitle = TRUE)

Arguments

obj, x, object
A compoundsCluster object.

use.names
A logical value specifying whether the returned vector should be named with the feature group names.

k, h
Desired number of clusters or tree height to be used for cutting the dendrogram, respecitively. One or the other must be specified. Analogous to cutree.

groupName
A character specifying the feature group name.

maxTreeHeight, deepSplit, minModuleSize
Arguments used by cutreeDynamicTree.

pal
Colour palette to be used from RColorBrewer.

colourBranches
Whether branches from cut clusters (and their labels) should be coloured. Might be slow with large numbers of clusters, hence, the default is only TRUE when this is not the case.

showLegend
If TRUE and colourBranches is also TRUE then a legend will be shown which outlines cluster numbers and their colours. By default TRUE for small amount of clusters to avoid overflowing the plot.

…
Any arguments directly given to plot.dendrogram.

cluster
A numeric value specifying the cluster.

width, height
The dimensions (in pixels) of the raster image that should be plotted.

withTitle
A logical value specifying whether a title should be added.

Value

cutTree and cutTreeDynamic return the modified compoundsCluster object.

getMCS returns an rcdk molecule object (IAtomContainer).

Details

Objects from this type are returned by the compounds method for makeHCluster.

Methods (by generic)

clusters: Accessor method to the clusters slot. Returns a list that contains for each feature group an object as returned by hclust.

cutClusters: Accessor method to the cutClusters slot. Returns a list that contains for each feature group a vector with cluster membership for each candidate (format as cutree).

clusterProperties: Returns a list with properties on how the clustering was performed.

length: Returns the total number of clusters.

lengths: Returns a vector with the number of clusters per feature group.

show: Show summary information for this object.

treeCut: Manually (re-)cut a dendrogram that was generated for a feature group.

treeCutDynamic: Automatically (re-)cut a dendrogram that was generated for a feature group using the cutreeDynamicTree function from dynamicTreeCut.

plot: Plot the dendrogram for clustered compounds of a feature group. Clusters are highlighted using dendextend.

getMCS: Calculates the maximum common substructure (MCS) for all candidate structures within a specified cluster. This method uses the get.mcs function from rcdk.

plotStructure: Plots the maximum common substructure (MCS) for all candidate structures within a specified cluster.

Slots

clusters

A list with hclust objects for each feature group.

SMILES

A list containing a vector with SMILES for all candidate structures per feature group.

cutClusters

A list with assigned clusters for all candidates per feature group (same format as what cutree returns).

properties

A list containing general properties and parameters used for clustering.