Compounds list class for MetFrag results.

This class is derived from compounds and contains additional specific MetFrag data.

# S4 method for compoundsMF
settings(compoundsMF)

Arguments

compoundsMF
A compoundsMF object.

Details

Objects from this class are generated by generateCompoundsMetFrag

Methods (by generic)

settings: Accessor method for the settings slot.

Slots

settings

A list with all general configuration settings passed to MetFrag. Feature specific items (e.g. spectra and precursor masses) are not contained in this list.

References

Schollee JE, Bourgin M, von Gunten U, McArdell CS and Hollender J (2018). “Non-target screening to trace ozonation transformation products in a wastewater treatment train including different post-treatments.” Water Research, 142, pp. 267--278. doi: 10.1016/j.watres.2018.05.045 .

Ruttkies C, Schymanski EL, Wolf S, Hollender J and Neumann S (2016). “MetFrag relaunched: incorporating strategies beyond in silico fragmentation.” Journal of Cheminformatics, 8(1). doi: 10.1186/s13321-016-0115-9 .

See also