Filtering of grouped features

Basic rule based filtering of feature groups.

# S4 method for featureGroups
filter(obj, intensityThreshold = NULL,
  relAbundance = NULL, absAbundance = NULL,
  interRelRGroupAbundance = NULL, interAbsRGroupAbundance = NULL,
  intraRGroupAbundance = NULL, minBlankThreshold = NULL,
  retentionRange = NULL, mzRange = NULL, chromWidthRange = NULL,
  rGroups = NULL, removeRefAnalyses = FALSE, repetitions = 2,
  negate = FALSE)

# S4 method for featureGroups
replicateGroupSubtract(fGroups, rGroups,
  threshold = 0)

Arguments

intensityThreshold
Minimum intensity. Set to 0 or NULL to skip this step.

relAbundance, absAbundance, interRelRGroupAbundance, interAbsRGroupAbundance
Minimal overall relative/absolute abundance for a feature group to be present within all analyses/replicate groups.

intraRGroupAbundance
Minimum relative abundance (0-1) that a feature group should be present within a replicate group. Set to NULL to skip this step.

minBlankThreshold
Feature groups that are also present in reference analyses (see analysis info) are filtered out unless their relative intensity is above this threshold. Set to NULL to skip this step.

retentionRange, mzRange, chromWidthRange
Range of retention time (in seconds), m/z or chromatographic peak width (in seconds), respectively. Features outside this range will be removed. Should be a numeric vector with length of two containing the min/max values. If the max value is set to a value below 0 then no maximum is assumed. Set to NULL to skip this step.

rGroups
A character vector of replicate groups that should be kept (filter) or subtracted from (replicateGroupSubtract).

removeRefAnalyses
Set to TRUE to remove all analyses that belong to replicate groups that are specified as a reference in the analysis-information. This is useful to simplify the analyses in the specified featureGroups object after blank subtraction. Consequently, when both minBlankThreshold and this argument are set, blank subtraction is performed prior to removing any analyses.

repetitions
Sometimes more feature groups may be removed by repeating filtering steps after another. Usually a value of two is enough to filter the maximum amount of feature groups.

negate
If set to TRUE then filtering operations are performed in opposite manner.

fGroups, obj
featureGroups object to which the filter is applied.

threshold
Minimal relative threshold for a feature group to be not removed. When 0 a feature group is always removed when present in the given replicate groups.

Value

A filtered featureGroups object. Feature groups that are filtered away have their intensity set to zero. In case a feature group is not present in any of the analyses anymore it will be removed completely.

Details

filter performs common rule based filtering of feature groups such as blank subtraction, minimum intensity and minimal replicate abundance.

replicateGroupSubtract removes feature groups present in a given set of replicate groups (unless intensities are above a given threshold).

See also