Filtering of grouped features

Basic rule based filtering of feature groups.

# S4 method for featureGroups
filter(obj, absMinIntensity = NULL,
  relMinIntensity = NULL, preAbsMinIntensity = NULL,
  preRelMinIntensity = NULL, absMinAnalyses = NULL,
  relMinAnalyses = NULL, absMinReplicates = NULL,
  relMinReplicates = NULL, absMinFeatures = NULL,
  relMinFeatures = NULL, absMinReplicateAbundance = NULL,
  relMinReplicateAbundance = NULL, maxReplicateIntRSD = NULL,
  blankThreshold = NULL, retentionRange = NULL, mzRange = NULL,
  mzDefectRange = NULL, chromWidthRange = NULL, rGroups = NULL,
  removeBlanks = FALSE, negate = FALSE)

# S4 method for featureGroups
replicateGroupSubtract(fGroups, rGroups,
  threshold = 0)

Arguments

absMinIntensity, relMinIntensity
Minimum absolute/relative intensity for features to be kept. The relative intensity is determined from the feature with highest intensity (of all features from all groups). Set to 0 or NULL to skip this step.

preAbsMinIntensity, preRelMinIntensity
As absMinIntensity/relMinIntensity, but applied before any other filters. This is typically used to speed-up subsequent filter steps. However, care must be taken that a sufficiently low value is choosen that is not expected to affect subsequent filtering steps. See below why this may be important.

absMinAnalyses, relMinAnalyses
Feature groups are only kept when they contain data for at least this (absolute or relative) amount of analyses. Set to NULL to ignore.

absMinReplicates, relMinReplicates
Feature groups are only kept when they contain data for at least this (absolute or relative) amount of replicates. Set to NULL to ignore.

absMinFeatures, relMinFeatures
Analyses are only kept when they contain at least this (absolute or relative) amount of features. Set to NULL to ignore.

absMinReplicateAbundance, relMinReplicateAbundance
Minimum absolute/relative abundance that a grouped feature should be present within a replicate group. If this mimimum is not met all features within the replicate group are removed. Set to NULL to skip this step.

maxReplicateIntRSD
Maximum relative standard deviation (RSD) of intensity values for features within a replicate group. If the RSD is above this value all features within the replicate group are removed. Set to NULL to ignore.

blankThreshold
Feature groups that are also present in blank analyses (see analysis info) are filtered out unless their relative intensity is above this threshold. For instance, a value of 5 means that only features with an intensity five times higher than that of the blank are kept. The relative intensity values between blanks and non-blanks are determined from the mean of all non-zero blank intensities. Set to NULL to skip this step.

retentionRange, mzRange, mzDefectRange, chromWidthRange
Range of retention time (in seconds), m/z, mass defect (defined as the decimal part of m/z values) or chromatographic peak width (in seconds), respectively. Features outside this range will be removed. Should be a numeric vector with length of two containing the min/max values. The maximum can be Inf to specify no maximum range. Set to NULL to skip this step.

rGroups
A character vector of replicate groups that should be kept (filter) or subtracted from (replicateGroupSubtract).

removeBlanks
Set to TRUE to remove all analyses that belong to replicate groups that are specified as a blank in the analysis-information. This is useful to simplify the analyses in the specified featureGroups object after blank subtraction. When both blankThreshold and this argument are set, blank subtraction is performed prior to removing any analyses.

negate
If set to TRUE then filtering operations are performed in opposite manner.

fGroups, obj
featureGroups object to which the filter is applied.

threshold
Minimum relative threshold (compared to mean intensity of replicate group being subtracted) for a feature group to be not removed. When 0 a feature group is always removed when present in the given replicate groups.

Value

A filtered featureGroups object. Feature groups that are filtered away have their intensity set to zero. In case a feature group is not present in any of the analyses anymore it will be removed completely.

Details

filter performs common rule based filtering of feature groups such as blank subtraction, minimum intensity and minimum replicate abundance. Removing of features occurs by zeroing their intensity values. Furthermore, feature groups that are left completely empty (i.e. all intensities are zero) will be automatically removed.

replicateGroupSubtract removes feature groups present in a given set of replicate groups (unless intensities are above a given threshold). The replicate groups that are subtracted will be removed.

Filter order

When multiple arguments are specified to filter, multiple filters are applied in sequence. Since some of these filters may affect each other, choosing their order correctly may be important for effective data filtering. For instance, when an intensity filter removes features from blank analyses, a subsequent blank filter may not adequately perform blank subtraction. Similarly, when intensity and blank filters are executed after the replicate abundance filter it may be necessary to ensure minimum replicate abundance again as the intensity and blank filters may have removed some features within a replicate group.

With this in mind, filters (if specified) occurs in the following order:

Pre-Intensity filters (i.e. preAbsMinIntensity and preRelMinIntensity).

Chromatography and mass filters (i.e retentionRange, mzRange, mzDefectRange and chromWidthRange).

Replicate abundance filters (i.e. absMinReplicateAbundance, relMinReplicateAbundance and maxReplicateIntRSD).

Blank filter (i.e. blankThreshold).

Intensity filters (i.e. absMinIntensity and relMinIntensity).

Replicate abundance filters (2nd time, only if previous filters affected results).

General abundance filters (i.e. absMinAnalyses, relMinAnalyses, absMinReplicates, relMinReplicates, absMinFeatures and relMinFeatures).

Replicate group filter (i.e. rGroups) and blank analyses removal (i.e. if removeBlanks=TRUE).

If another filtering order is desired then filter should be called multiple times with only one filter argument at a time.

See also