Finding features

Functions and classes for collection of features.

findFeatures(analysisInfo, algorithm, ..., verbose = TRUE)

importFeatures(analysisInfo, type, ...)

findFeaturesBruker(
  analysisInfo,
  doFMF = "auto",
  startRange = 0,
  endRange = 0,
  save = TRUE,
  close = save,
  verbose = TRUE
)

findFeaturesEnviPick(analysisInfo, ..., verbose = TRUE)

importFeaturesEnviMass(analysisInfo, enviProjPath)

findFeaturesOpenMS(
  analysisInfo,
  noiseThrInt = 1000,
  chromSNR = 3,
  chromFWHM = 5,
  mzPPM = 10,
  reEstimateMTSD = TRUE,
  traceTermCriterion = "sample_rate",
  traceTermOutliers = 5,
  minSampleRate = 0.5,
  minTraceLength = 3,
  maxTraceLength = -1,
  widthFiltering = "fixed",
  minFWHM = 3,
  maxFWHM = 60,
  traceSNRFiltering = FALSE,
  localRTRange = 10,
  localMZRange = 6.5,
  isotopeFilteringModel = "metabolites (5% RMS)",
  MZScoring13C = FALSE,
  useSmoothedInts = TRUE,
  extraOpts = NULL,
  intSearchRTWindow = 3,
  logPath = file.path("log", "openms"),
  maxProcAmount = getOption("patRoon.maxProcAmount"),
  verbose = TRUE
)

findFeaturesXCMS(analysisInfo, method = "centWave", ..., verbose = TRUE)

importFeaturesXCMS(xs, analysisInfo)

findFeaturesXCMS3(analysisInfo, param = xcms::CentWaveParam(), verbose = TRUE)

importFeaturesXCMS3(xdata, analysisInfo)

Arguments

analysisInfo
Analysis info table.

algorithm
A character string describing the algorithm that should be used: "bruker", "openms", "xcms", "xcms3", "envipick"

...
further parameters passed to xcmsSet (findFeaturesXCMS), enviPickwrap (featurefinderEnviPick) or to selected feature finding or importing algorithms (findFeatures and importFeatures).

verbose
If set to FALSE then no text output is shown.

type
What type of data should be imported: "xcms", "xcms3" or "envimass".

doFMF
Run the 'Find Molecular Features' algorithm before loading compounds. Valid options are: "auto" (run FMF automatically if current results indicate it is necessary) and "force" (run FMF always, even if cached results exist). Note that checks done if doFMF="auto" are fairly simplistic, hence set doFMF="force" if feature data needs to be updated.

startRange, endRange
Start/End retention range (seconds) from which to collect features. A 0 (zero) for endRange marks the end of the analysis.

close, save
If TRUE then Bruker files are closed and saved after processing with DataAnalysis, respectively. Setting close=TRUE prevents that many analyses might be opened simultaneously in DataAnalysis, which otherwise may use excessive memory or become slow. By default save is TRUE when close is TRUE, which is likely what you want as otherwise any processed data is lost.

enviProjPath
The path of the enviMass project.

noiseThrInt
Noise intensity threshold. Sets algorithm:common:noise_threshold_int option.

chromSNR
Minimum S/N of a mass trace. Sets algorithm:common:chrom_peak_snr option.

chromFWHM
Expected chromatographic peak width (in seconds). Sets algorithm:common:chrom_fwhm option.

mzPPM
Allowed mass deviation (ppm) for trace detection. Sets algorithm:mtd:mass_error_ppm.

reEstimateMTSD
If TRUE then enables dynamic re-estimation of m/z variance during mass trace collection stage. Sets algorithm:mtd:reestimate_mt_sd.

traceTermCriterion, traceTermOutliers, minSampleRate
Termination criterion for the extension of mass traces. See FeatureFinderMetabo. Sets the algorithm:mtd:trace_termination_criterion, algorithm:mtd:trace_termination_outliers and algorithm:mtd:min_sample_rate options, respectively.

minTraceLength, maxTraceLength
Minimum/Maximum length of mass trace (seconds). Set negative value for maxlength to disable maximum. Sets algorithm:mtd:min_trace_length and algorithm:mtd:min_trace_length, respectively.

widthFiltering, minFWHM, maxFWHM
Enable filtering of unlikely peak widths. See FeatureFinderMetabo. Sets algorithm:epd:width_filtering, algorithm:epd:min_fwhm and algorithm:epd:max_fwhm, respectively.

traceSNRFiltering
If TRUE then apply post-filtering by signal-to-noise ratio after smoothing. Sets the algorithm:epd:masstrace_snr_filtering option.

localRTRange, localMZRange
Retention/MZ range where to look for coeluting/isotopic mass traces. Sets the algorithm:ffm:local_rt_range and algorithm:ffm:local_mz_range options, respectively.

isotopeFilteringModel
Remove/score candidate assemblies based on isotope intensities. See FeatureFinderMetabo. Sets the algorithm:ffm:isotope_filtering_model option.

MZScoring13C
Use the 13C isotope as the expected shift for isotope mass traces. See FeatureFinderMetabo. Sets algorithm:ffm:mz_scoring_13C.

useSmoothedInts
If TRUE then use LOWESS intensities instead of raw intensities. Sets the algorithm:ffm:use_smoothed_intensities option.

extraOpts
Named list containing extra options that will be passed to FeatureFinderMetabo. Any options specified here will override any of the above. Example: extraOpts=list("-algorithm:common:noise_threshold_int"=1000) (corresponds to setting noiseThrInt=1000). Set to NULL to ignore.

intSearchRTWindow
Retention time window (in seconds, +/- feature retention time) that is used to find the closest data point to the retention time to obtain the intensity of a feature (this is needed since OpenMS does not provide this data).

logPath
Destination directory for log files with output from executed commands. Will be created if non-existant. Set to NULL to disable logging.

maxProcAmount
Maximum number of processes to run for parallelization. Usually a number close to the amount of physical cores yields most efficient results.

method
The method setting used by XCMS peak finding, see xcms::findPeaks

xs
An xcmsSet object.

param
The method parameters used by XCMS peak finding, see xcms::findChromPeaks

xdata
An XCMSnExp object.

Value

An object of a class which is derived from features.

Details

Several functions exist to collect features (i.e. retention and MS information that represent potential compounds) from a set of analyses. All 'feature finders' return an object derived from the features base class. The next step in a general workflow is to group and align these features across analyses by feature groupers. Note that some feature finders have a plethora of options which sometimes may have a large effect on the quality of results. Fine-tuning parameters is therefore important, and the optimum is largely dependent upon applied analysis methodology and instrumentation.

findFeatures is a generic function that will find features using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as findFeaturesOpenMS and findFeaturesBruker. While these functions may be called directly, findFeatures provides a generic interface and is therefore usually preferred.

importFeatures is a generic function to import feature groups produced by other software. The actual functionality is provided by specific functions such as importFeaturesXCMS and importFeaturesEnviMass.

findFeaturesBruker uses the 'Find Molecular Features' (FMF) algorithm of Bruker DataAnalysis vendor software to find features. The resulting 'compounds' are then transferred from DataAnalysis and stored as features.

findFeaturesEnviPick uses the enviPickwrap. function from the enviPick R package to extract features.

importFeaturesEnviMass imports features from a project generated by the enviMass package. NOTE: this functionality has only been tested with older versions of enviMass.

findFeaturesOpenMS uses the FeatureFinderMetabo TOPP tool (see http://www.openms.de).

findFeaturesXCMS uses the xcmsSet function from the xcms package to find features.

importFeaturesXCMS converts features from an existing xcmsSet object (obtained with the xcms package) to a new features object.

findFeaturesXCMS3 uses the new xcms3 interface from the xcms package to find features.

importFeaturesXCMS3 converts features from an existing XCMSnExp object (obtained with the xcms package) to a new features object.

Note

The file format of analyses for findFeaturesXCMS and findFeaturesXCMS3 must be mzML or mzXML.

findFeaturesBruker only works with Bruker data files (.d extension) and requires Bruker DataAnalysis and the RDCOMClient package to be installed. Furthermore, DataAnalysis combines multiple related masses in a feature (e.g. isotopes, adducts) but does not report the actual (monoisotopic) mass of the feature. Therefore, it is simply assumed that the feature mass equals that of the highest intensity mass peak.

If any errors related to DCOM appear it might be necessary to terminate DataAnalysis (note that DataAnalysis might still be running as a background process). The ProcessCleaner application installed with DataAnalayis can be used for this.

findFeaturesEnviPick Requires analysis files to be in the mzXML format.

The file format of analyses for findFeaturesOpenMS must be mzML. This functionality has been tested with OpenMS version >= 2.0. Please make sure it is installed and its binaries are added to the PATH environment variable or the patRoon.path.OpenMS option is set.

References

Rost HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmstrom L, Aebersold R, Reinert K, Kohlbacher O (2016-sep). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13, 741--748. doi: 10.1038/nmeth.3959 .

pugixml (via Rcpp) is used to process OpenMS XML output.

Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18. URL http://www.jstatsoft.org/v40/i08/.

Eddelbuettel, Dirk (2013) Seamless R and C++ Integration with Rcpp. Springer, New York. ISBN 978-1-4614-6867-7.

Dirk Eddelbuettel and James Joseph Balamuta (2017). Extending R with C++: A Brief Introduction to Rcpp. PeerJ Preprints 5:e3188v1. URL https://doi.org/10.7287/peerj.preprints.3188v1.

Smith, C.A. and Want, E.J. and O'Maille, G. and Abagyan,R. and Siuzdak, G.: XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 78:779-787 (2006)

Ralf Tautenhahn, Christoph Boettcher, Steffen Neumann: Highly sensitive feature detection for high resolution LC/MS BMC Bioinformatics, 9:504 (2008)

H. Paul Benton, Elizabeth J. Want and Timothy M. D. Ebbels Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data Bioinformatics, 26:2488 (2010)

Smith, C.A. and Want, E.J. and O'Maille, G. and Abagyan,R. and Siuzdak, G.: XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 78:779-787 (2006)

Ralf Tautenhahn, Christoph Boettcher, Steffen Neumann: Highly sensitive feature detection for high resolution LC/MS BMC Bioinformatics, 9:504 (2008)

H. Paul Benton, Elizabeth J. Want and Timothy M. D. Ebbels Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data Bioinformatics, 26:2488 (2010)

See also