Grouping of features

Functions and classes for grouping of features across analyses.

# S4 method for features
groupFeatures(obj, algorithm, ..., verbose = TRUE)

# S4 method for data.frame
groupFeatures(obj, algorithm, ..., verbose = TRUE)

importFeatureGroups(path, type, ...)

# S4 method for featuresSet
groupFeatures(obj, algorithm, ..., verbose = TRUE)

# S4 method for featureGroups
makeSet(
  obj,
  ...,
  groupAlgo,
  groupArgs = NULL,
  verbose = TRUE,
  adducts = NULL,
  labels = NULL
)

# S4 method for featureGroupsSet
makeSet(obj, ...)

importFeatureGroupsBrukerPA(
  path,
  feat,
  rtWindow = 12,
  mzWindow = 0.005,
  intWindow = 5,
  warn = TRUE
)

importFeatureGroupsEnviMass(path, feat, positive)

# S4 method for features
groupFeaturesKPIC2(
  feat,
  rtalign = TRUE,
  loadRawData = TRUE,
  groupArgs = list(tolerance = c(0.005, 12)),
  alignArgs = list(),
  verbose = TRUE
)

# S4 method for featuresSet
groupFeaturesKPIC2(
  feat,
  groupArgs = list(tolerance = c(0.005, 12)),
  verbose = TRUE
)

importFeatureGroupsKPIC2(picsSetGrouped, analysisInfo)

# S4 method for featureGroupsKPIC2
delete(obj, ...)

# S4 method for features
groupFeaturesOpenMS(
  feat,
  rtalign = TRUE,
  QT = FALSE,
  maxAlignRT = 30,
  maxAlignMZ = 0.005,
  maxGroupRT = 12,
  maxGroupMZ = 0.005,
  extraOptsRT = NULL,
  extraOptsGroup = NULL,
  verbose = TRUE
)

groupFeaturesSIRIUS(analysisInfo, verbose = TRUE)

# S4 method for features
groupFeaturesXCMS(
  feat,
  rtalign = TRUE,
  loadRawData = TRUE,
  groupArgs = list(mzwid = 0.015),
  retcorArgs = list(method = "obiwarp"),
  verbose = TRUE
)

# S4 method for featuresSet
groupFeaturesXCMS(feat, groupArgs = list(mzwid = 0.015), verbose = TRUE)

importFeatureGroupsXCMS(xs, analysisInfo)

# S4 method for featureGroupsXCMS
delete(obj, ...)

# S4 method for features
groupFeaturesXCMS3(
  feat,
  rtalign = TRUE,
  loadRawData = TRUE,
  groupParam = xcms::PeakDensityParam(sampleGroups = analysisInfo(feat)$group),
  preGroupParam = groupParam,
  retAlignParam = xcms::ObiwarpParam(),
  verbose = TRUE
)

# S4 method for featuresSet
groupFeaturesXCMS3(
  feat,
  groupParam = xcms::PeakDensityParam(sampleGroups = analysisInfo(feat)$group),
  verbose = TRUE
)

importFeatureGroupsXCMS3(xdata, analysisInfo)

# S4 method for featureGroupsXCMS3
delete(obj, ...)

Arguments

obj
Either a features object to be grouped, or a data.frame with analysis info to be passed to groupFeaturesSIRIUS or a featureGroups objects passed to makeSet.

algorithm
A character that specifies the algorithm to be used: either "openms", "xcms", "xcms3" or "kpic2" (features method), or "sirius" (data.frame method).

...
Any parameters to be passed to the selected grouping/importing algorithm.

For makeSet: further featureGroups objects that should be used for the sets workflow.

verbose
if FALSE then no text output will be shown.

path
The path that should be used for importing. For importFeatureGroupsBrukerPA an exported 'bucket table' .txt file from Bruker ProfileAnalysis, for importFeatureGroupsBrukerTASQ an exported global result table (converted to .csv) and for importFeatureGroupsEnviMass the path of the enviMass project.

type
Which file type should be imported or exported: "brukerpa" (Bruker ProfileAnalysis), "brukertasq" (Bruker TASQ), envimass (enviMass, only import) or "mzmine" (MZMine, only export).

groupAlgo
groupAlgo The name of the feature grouping algorithm. See the algorithm argument description.

groupArgs
A list with arguments directly passed to groupFeatures (can be named). Example: groupArgs=list(maxAlignMZ=0.002).

adducts
The adduct assignments to each set. Should either be a list with adduct objects or a character vector (e.g. "[M+H]+"). The order should follow that of the objects given to the obj and ... arguments. if NULL then adduct annotations are used.

labels
The labels, or set names, for each set to be created. The order should follow that of the objects given to the obj and ... arguments. If NULL, then labels are automatically generated from the polarity of the specified adducts argument (e.g. "positive", "negative").

feat
The features to be grouped. importFeatureGroupsBrukerPA and importFeatureGroupsEnviMass only support features generated by findFeaturesBruker and importFeaturesEnviMass, respectively.

rtWindow, mzWindow, intWindow
Search window values for retention time (seconds), m/z (Da) and intensity used to find back features within feature groups from PA (+/- the retention/mass/intensity value of a feature).

warn
Warn about missing or duplicate features when relating them back from grouped features.

positive
Whether data from positive (TRUE) or negative (FALSE) should be loaded.

rtalign
Enable retention time alignment.

loadRawData
Set to TRUE if analyses are available as mzXML or mzML files. Otherwise MS data is not loaded, and some dummy data (e.g. file paths) is used in the returned object.

alignArgs
named character vector that may contain extra parameters to be used by KPIC::PICset.align.

picsSetGrouped
A grouped PIC set object (e.g. as returned by KPIC::PICset.group).

analysisInfo
A data.frame with analysis info.

QT
If enabled, use FeatureLinkerUnlabeledQT instead of FeatureLinkerUnlabeled for feature grouping.

maxAlignRT, maxAlignMZ
Used for retention alignment. Maximum retention time or m/z difference (seconds/Dalton) for feature pairing. Sets -algorithm:pairfinder:distance_RT:max_difference and -algorithm:pairfinder:distance_MZ:max_difference otpions, respectively.

maxGroupRT, maxGroupMZ
as maxAlignRT and maxAlignMZ, but for grouping of features. Sets -algorithm:distance_RT:max_difference and -algorithm:distance_MZ:max_difference options, respectively.

extraOptsRT, extraOptsGroup
Named list containing extra options that will be passed to MapAlignerPoseClustering or FeatureLinkerUnlabeledQT/FeatureLinkerUnlabeled, respectively. Any options specified here will override any of the above. Example: extraOptsGroup=list("-algorithm:distance_RT:max_difference"=12) (corresponds to setting maxGroupRT=12). Set to NULL to ignore.

retcorArgs
named character vector that may contain extra parameters to be used by xcms::retcor.

xs
An xcmsSet object.

groupParam, retAlignParam
parameter object that is directly passed to xcms::groupChromPeaks and xcms::adjustRtime, respectively.

preGroupParam
grouping parameters applied when features are grouped prior to alignment (only with peak groups alignment).

xdata
An XCMSnExp object.

Value

An object of a class which is derived from featureGroups.

makeSet and the featuresSet method of groupFeatures return a featureGroupsSet object.

Details

After features have been found the logical next step is to align and group them across analyses. This process is necessary to allow comparison of features between multiple analyses, which otherwise would be difficult due to small deviations in retention and mass data. Thus, algorithms of 'feature groupers' are used to collect features with similar retention and mass data. In addition, advanced retention time alignment algorithms exist to enhance grouping of features even with relative large retention time deviations (e.g. possibly observed from analyses collected over a long period). Like finding of features, various algorithms are supported which may have many parameters that can be fine-tuned. This fine-tuning is likely to be necessary, since optimal settings often depend on applied methodology and instrumentation.

groupFeatures is a generic function that will group features using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as groupFeaturesOpenMS and groupFeaturesXCMS3. While these functions may be called directly, groupFeatures provides a generic interface and is therefore usually preferred.

The data.frame method for groupFeatures is a special case that currently only supports the "sirius" algorithm.

importFeatureGroups is a generic function to import feature groups produced by other software. The actual functionality is provided by specific functions such as importFeatureGroupsBrukerPA and importFeatureGroupsEnviMass.

makeSet is used to initiate a sets workflow. See the Sets workflows section for more details.

importFeatureGroupsBrukerPA imports grouped features generated with Bruker ProfileAnalysis (PA). To do so, a 'bucket table' should be generated using PA and exported as .txt file. Please note that this function only supports features generated by findFeaturesBruker and it is crucial that DataAnalysis files remain unchanged when features are collected and the bucket table is generated. Furthermore, please note that PA does not retain information about originating features for generated buckets. For this reason, this function tries to find back the original features and care must be taken to correctly specify search parameters (rtWindow, mzWindow, intWindow).

importFeatureGroupsEnviMass imports grouped features ('profiles') generated with enviMass. Note that this function only imports 'raw' profiles, not any results from further componentization steps performed in enviMass. Furthermore, this functionality has only been tested with older versions of enviMass. Finally, please note that this function only supports features imported by importFeaturesEnviMass (obviously, the same project should be used for both importing functions).

groupFeaturesKPIC2 uses the the KPIC2 R package for grouping of features. Grouping of features and alignment of their retention times are performed with the KPIC::PICset.group and KPIC::PICset.align functions, respectively.

importFeatureGroupsKPIC2 imports grouped features from an KPIC object.

groupFeaturesOpenMS uses the OpenMS tools for grouping of features (see http://www.openms.de). Retention times may be aligned by the MapAlignerPoseClustering TOPP tool. Grouping is achieved by either the FeatureLinkerUnlabeled or FeatureLinkerUnlabeledQT TOPP tools.

groupFeaturesSIRIUS uses SIRIUS to find and group features. This is done by running the lcms-align command on every analyses at once. Note that grouping feature data from other algorithms than SIRIUS are therefore not supported.

groupFeaturesXCMS uses the xcms package for grouping of features. Grouping of features and alignment of their retention times are performed with the xcms::group and xcms::retcor functions, respectively. Both functions have an extensive list of parameters to modify their behaviour and may therefore be used to potentially optimize results.

importFeatureGroupsXCMS converts grouped features from an xcmsSet object (from the xcms package).

groupFeaturesXCMS3 uses the new interface from the xcms package for grouping of features. Grouping of features and alignment of their retention times are performed with the xcms::groupChromPeaks and xcms::adjustRtime functions, respectively. Both of these functions support an extensive amount of parameters that modify their behaviour and may therefore require optimization.

importFeatureGroupsXCMS3 converts grouped features from an XCMSnExp object (from the xcms package).

Note

In most cases it will be necessary to centroid your MS input files. The only exception is Bruker, however, you will still need centroided mzXML/mzML files for e.g. plotting chromatograms. In this case the centroided MS files should be stored in the same directory as the raw Bruker .d files. The convertMSFiles function can be used to centroid data.

makeSet Currently does not support making sets from featureGroupsSet objects.

Sets workflows

With a sets workflow the features are first neutralized, i.e. their m/z values is replaced by their neutral monoisotopic mass. The latter is calculated from adduct annotations, specified either directly, or already present in the input objects. The features are then grouped by any of the groupFeatures algorithms. The neutralization step ensures that features measured with e.g. different ionization polarities can be grouped since their neutral mass will be the same. The sets workflow is initiated in either of the following two ways:

The groupFeatures method can be called with a featuresSet object (generated with the features method of makeSet).

By using the makeSet method for featureGroups. This uses the features method internally to first create a featuresSet object from all the feature data inside the input featureGroups objects. The (neutralized) features are then (re-)grouped with groupFeatures.

The advantage of the second method is that it preserves adduct annotations present in the input featureGroups objects (e.g. as set by selectIons or adducts<-). Furthermore, the second approach allows more advanced workflows where the input featureGroups are first pre-treated with e.g. filter before the sets object is made. On the other hand, the first approach is easier as it doesn't require intermediate feature grouping steps, and is often sufficient as adduct annotations can be made afterwards with selectIons/adducts<-, and most filter operations do not need to be done per individual set.

With both approaches the resulting featureGroupsSet object will contain adduct annotations, either based on those already present from input featureGroups objects or as specified by the adducts argument to makeSet.

NOTE the loadRawData and arguments related to retention time alignment (e.g. rtalign, retcorArgs) are currently not supported for sets workflows when using algorithms from KPIC2 or XCMS.

References

Ji H, Zeng F, Xu Y, Lu H, Zhang Z (2017). “KPIC2: An Effective Framework for Mass Spectrometry-Based Metabolomics Using Pure Ion Chromatograms.” Analytical Chemistry, 89(14), 7631--7640. doi: 10.1021/acs.analchem.7b01547 .

Rost HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmstrom L, Aebersold R, Reinert K, Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), 741--748. doi: 10.1038/nmeth.3959 .

pugixml (via Rcpp) is used to process OpenMS XML output.

Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18, .

Eddelbuettel, Dirk (2013) Seamless R and C++ Integration with Rcpp, Springer, New York, ISBN 978-1-4614-6867-7, .

Dirk Eddelbuettel and James Joseph Balamuta (2018). Extending R with C++: A Brief Introduction to Rcpp. The American Statistician. 72(1), .

Duhrkop K, Fleischauer M, Ludwig M, Aksenov AA, Melnik AV, Meusel M, Dorrestein PC, Rousu J, Bocker S (2019). “SIRIUS 4: a rapid tool for turning tandem mass spectra into metabolite structure information.” Nature Methods, 16(4), 299--302. doi: 10.1038/s41592-019-0344-8 .

Smith, C.A. and Want, E.J. and O'Maille, G. and Abagyan,R. and Siuzdak, G.: XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 78:779-787 (2006)

Ralf Tautenhahn, Christoph Boettcher, Steffen Neumann: Highly sensitive feature detection for high resolution LC/MS BMC Bioinformatics, 9:504 (2008)

H. Paul Benton, Elizabeth J. Want and Timothy M. D. Ebbels Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data Bioinformatics, 26:2488 (2010)

Smith, C.A. and Want, E.J. and O'Maille, G. and Abagyan,R. and Siuzdak, G.: XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 78:779-787 (2006)

Ralf Tautenhahn, Christoph Boettcher, Steffen Neumann: Highly sensitive feature detection for high resolution LC/MS BMC Bioinformatics, 9:504 (2008)

H. Paul Benton, Elizabeth J. Want and Timothy M. D. Ebbels Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data Bioinformatics, 26:2488 (2010)

See also