Grouping of features

Functions and classes for grouping of features across analyses.

# S4 method for features
groupFeatures(feat, algorithm, ...)

importFeatureGroups(path, type, ...)

importFeatureGroupsBrukerPA(path, feat, maxRtDev = 12, maxMzDev = 0.005,
  maxIntDev = 5, warn = TRUE)

importFeatureGroupsEnviMass(path, feat, positive)

groupFeaturesOpenMS(feat, rtalign = TRUE, QT = FALSE, maxAlignRT = 30,
  maxAlignMZ = 0.005, maxGroupRT = 12, maxGroupMZ = 0.005)

groupFeaturesXCMS(feat, rtalign = TRUE, exportedData = TRUE,
  groupArgs = list(mzwid = 0.015), retcorArgs = list(method = "obiwarp"))

importFeatureGroupsXCMS(xs, analysisInfo)

Arguments

feat
The features to be grouped. importFeatureGroupsBrukerPA and importFeatureGroupsEnviMass only support features generated by findFeaturesBruker and importFeaturesEnviMass, respectively.

algorithm
A character string describing the algorithm that should be used: "openms", "xcms"

…
Any parameters to be passed to the selected grouping/importing algorithm.

path
The path that should be used for importing. For importFeatureGroupsBrukerPA an exported 'bucket table' .txt file from Bruker ProfileAnalysis, for importFeatureGroupsBrukerTASQ an exported global result table (converted to .csv) and for importFeatureGroupsEnviMass the path of the enviMass project.

type
The type of data to be imported: brukerpa for Bruker ProfileAnalysis, brukertasq for Bruker TASQ or envimass for enviMass.

maxRtDev, maxMzDev, maxIntDev
Maximum allowance in difference of retention time (seconds), m/z (Da) and intensity that are used to relate features with feature groups (buckets) from PA (see details).

warn
Warn about missing or duplicate features when relating them back from grouped features.

positive
Whether data from positive (TRUE) or negative (FALSE) should be loaded.

rtalign
Enable retention time alignment.

QT
If enabled, use FeatureLinkerUnlabeledQT instead of FeatureLinkerUnlabeled for feature grouping.

maxAlignRT, maxAlignMZ
Used for retention alignment. Maximum retention time or m/z difference (seconds/Dalton) for feature pairing. Sets -algorithm:pairfinder:distance_RT:max_difference and -algorithm:pairfinder:distance_MZ:max_difference otpions, respectively.

maxGroupRT, maxGroupMZ
as maxAlignRT and maxAlignMZ, but for grouping of features. Sets -algorithm:distance_RT:max_difference and -algorithm:distance_MZ:max_difference options, respectively.

exportedData
Set to TRUE if analyses were exported as mzXML or mzML files.

groupArgs, retcorArgs
named character vector that can contain extra parameters to be used by xcms::group and xcms::retcor, respectively.

xs
An xcmsSet object.

analysisInfo
A data.frame with analysis info.

Value

An object of a class which is derived from featureGroups.

Details

After features have been found the logical next step is to align and group them across analyses. This process is necessary to allow comparison of features between multiple analyses, which otherwise would be difficult due to small deviations in retention and mass data. Thus, algorithms of 'feature groupers' are used to collect features with similar retention and mass data. In addition, advanced retention time alignment algorithms exist to enhance grouping of features even with relative large retention time deviations (e.g. possibly observed from analyses collected over a long period). Like finding of features, various algorithms are supported which may have many parameters that can be fine-tuned. This fine-tuning is likely to be necessary, since optimal settings often depend on applied methodology and instrumentation. groupFeatures is a generic function that will group features using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as groupFeaturesOpenMS and groupFeaturesXCMS. While these functions may be called directly, groupFeatures provides a generic interface and is therefore usually preferred. importFeatureGroups is a generic function to import feature groups produced by other software. The actual functionality is provided by specific functions such as importFeatureGroupsBrukerPA and importFeatureGroupsEnviMass. importFeatureGroupsBrukerPA imports grouped features generated with Bruker ProfileAnalysis (PA). To do so, a 'bucket table' should be generated using PA and exported as .txt file. Please note that this function only supports features generated by findFeaturesBruker and it is crucial that DataAnalysis files remain unchanged when features are collected and the bucket table is generated. Furthermore, please note that PA does not retain information about originating features for generated buckets. For this reason, this function tries to find back the original features and care must be taken to correctly specify search parameters (maxRtDev, maxMzDev, maxIntDev). importFeatureGroupsEnviMass imports grouped features ('profiles') generated with enviMass. Please note that this function only supports features imported by importFeaturesEnviMass (obviously, the same project should be used for both importing functions). groupFeaturesOpenMS uses the OpenMS tools for grouping of features (see http://www.openms.de). Retention times may be aligned by the MapAlignerPoseClustering TOPP tool. Grouping is achieved by either the FeatureLinkerUnlabeled or FeatureLinkerUnlabeledQT TOPP tools. groupFeaturesXCMS uses the xcms package for grouping of features. Grouping of features and alignment of their retention times are performed with the xcms::group and xcms::retcor functions, respectively. Both functions have an extensive list of parameters to modify their behaviour and may therefore be used to potentially optimize results. importFeatureGroupsXCMS converts grouped features from an xcmsSet object (from the xcms package).

References

Röst HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmström L, Aebersold R, Reinert K and Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), pp. 741--748. doi: 10.1038/nmeth.3959 . Smith, C.A. and Want, E.J. and O'Maille, G. and Abagyan,R. and Siuzdak, G.: XCMS: Processing mass spectrometry data for metabolite profiling using nonlinear peak alignment, matching and identification, Analytical Chemistry, 78:779-787 (2006) Ralf Tautenhahn, Christoph Boettcher, Steffen Neumann: Highly sensitive feature detection for high resolution LC/MS BMC Bioinformatics, 9:504 (2008) H. Paul Benton, Elizabeth J. Want and Timothy M. D. Ebbels Correction of mass calibration gaps in liquid chromatography-mass spectrometry metabolomics data Bioinformatics, 26:2488 (2010)

See also

featureGroups-class