Plotting of grouped features

Various plotting functions for feature group data.

# S4 method for featureGroups,missing
plot(
  x,
  colourBy = c("none", "rGroups", "fGroups"),
  onlyUnique = FALSE,
  retMin = FALSE,
  showLegend = TRUE,
  col = NULL,
  pch = NULL,
  ...
)

# S4 method for featureGroups
plotInt(
  obj,
  average = FALSE,
  normFunc = NULL,
  xnames = TRUE,
  showLegend = FALSE,
  pch = 20,
  type = "b",
  lty = 3,
  col = NULL,
  ...
)

# S4 method for featureGroupsSet
plotInt(
  obj,
  average = FALSE,
  normFunc = NULL,
  xnames = !sets,
  showLegend = sets,
  pch = 20,
  type = "b",
  lty = 3,
  col = NULL,
  ...,
  sets = FALSE
)

# S4 method for featureGroups
plotChord(
  obj,
  addSelfLinks = FALSE,
  addRetMzPlots = TRUE,
  average = FALSE,
  outerGroups = NULL,
  addIntraOuterGroupLinks = FALSE,
  ...
)

# S4 method for featureGroups
plotChroms(
  obj,
  rtWindow = 30,
  mzExpWindow = 0.001,
  retMin = FALSE,
  topMost = NULL,
  topMostByRGroup = FALSE,
  EICs = NULL,
  showPeakArea = FALSE,
  showFGroupRect = TRUE,
  title = NULL,
  colourBy = c("none", "rGroups", "fGroups"),
  showLegend = TRUE,
  onlyPresent = TRUE,
  annotate = c("none", "ret", "mz"),
  showProgress = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

# S4 method for featureGroups
plotVenn(obj, which = NULL, ...)

# S4 method for featureGroupsSet
plotVenn(obj, which = NULL, ..., sets = FALSE)

# S4 method for featureGroups
plotUpSet(obj, which = NULL, nsets = length(which), nintersects = NA, ...)

# S4 method for featureGroups
plotVolcano(
  obj,
  FCParams,
  showLegend = TRUE,
  averageFunc = mean,
  col = NULL,
  pch = 19,
  ...
)

Arguments

x

featureGroups object to be accessed.

colourBy

Sets the automatic colour selection: "none" for a single colour or "rGroups"/"fGroups" for a distinct colour per replicate/feature group.

onlyUnique

If TRUE and colourBy="rGroups" then only feature groups that are unique to a replicate group are plotted.

retMin

Plot retention time in minutes (instead of seconds).

showLegend

Plot a legend if TRUE.

col

Colour(s) used. If col=NULL then colours are automatically generated.

pch, type, lty

Common plotting parameters passed to e.g. plot. For plot: if pch=NULL then values are automatically assigned.

...

passed to plot (plot and plotChroms), lines (plotInt), VennDiagram plotting functions (plotVenn), chordDiagram (plotChord) or upset (plotUpSet).

obj

featureGroups object to be accessed.

average

If TRUE then data within replicate groups are averaged.

For as.data.table: if features=TRUE other feature properties are also averaged.

normFunc

Function that should be used for normalization of data. The function is called for all intensities/areas of a feature group and these quantities are divided by the result of the function call. For example, when max is used normalized intensities will be between zero and one. If all quantities are zero then the function will not be called. Set to NULL to perform no normalization.

xnames

Plot analysis (or replicate group if average=TRUE) names on the x axis.

sets

(sets workflow) For plotInt: if TRUE then feature intensities are plot per set (order follows the analysis information).

For plotVenn: If TRUE then the which argument changes its meaning and is used to specify the names of the sets to be compared.

addSelfLinks

If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots

Set to TRUE to enable m/z vs retention time scatter plots.

outerGroups

Character vector of names to be used as outer groups. The values in the specified vector should be named by analysis names (average set to FALSE) or replicate group names (average set to TRUE), for instance: c(analysis1 = "group1", analysis2 = "group1", analysis3 = "group2"). Set to NULL to disable outer groups.

addIntraOuterGroupLinks

If TRUE then links will be added within outer groups.

rtWindow

Retention time (in seconds) that will be subtracted/added to respectively the minimum and maximum retention time of the plotted feature groups. Thus, setting this value to a positive value will 'zoom out' on the retention time axis.

mzExpWindow

In case the m/z window to plot an EIC for a particular analysis is not known (i.e. no feature was detected of the feature group to be plot and onlyPresent=FALSE) then the EIC m/z range is estimated from the range for the complete feature group and expanded by the offset defined by mzExpWindow.

topMost

Only plot EICs from features within this number of top most intense analyses. If NULL then all analyses are used for plotted.

topMostByRGroup

If set to TRUE and topMost is set: only plot EICs for the top most features in each replicate group. For instance, when topMost=1 and topMostByRGroup=TRUE, then EICs will be plotted for the most intense feature of each replicate group.

EICs

Internal parameter for now and should be kept at NULL (default).

showPeakArea

Set to TRUE to display integrated chromatographic peak ranges by filling (shading) their areas.

showFGroupRect

Set to TRUE to mark the full retention/intensity range of all features within a feature group by drawing a rectangle around it.

title

Character string used for title of the plot. If NULL a title will be automatically generated.

onlyPresent

If TRUE then EICs will only be generated for analyses in which a particular feature group was detected. Disabling this option might be useful to see if any features were 'missed'.

annotate

If set to "ret" and/or "mz" then retention and/or m/z values will be drawn for each plotted feature group.

showProgress

if set to TRUE then a text progressbar will be displayed when all EICs are being plot. Set to "none" to disable any annotation.

xlim, ylim

Sets the plot size limits used by plot. Set to NULL for automatic plot sizing.

which

A character vector with replicate groups used for comparison. Set to NULL to ignore.

For plotVenn: alternatively a named list containing elements of character vectors with replicate groups to compare. For instance, which=list(infl = c("influent-A", "influent-B"), effl = c("effluent-A", "effluent-B")), will compare the features in replicate groups "influent-A/B" against those in "effluent-A/B". The names of the list are used for labelling in the plot, and will be made automatically if not specified.

nsets, nintersects

See upset.

FCParams

A parameter list to calculate Fold change data. See getFCParams for more details.

averageFunc

Function used for averaging.

Value

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g.draw.pairwise.venn and draw.triple.venn).

Details

plot Generates an m/z vs retention time plot for all featue groups. Optionally highlights unique/overlapping presence amongst replicate groups.

plotInt Generates a line plot for the (averaged) intensity of feature groups within all analyses

plotChord Generates a chord diagram which can be used to visualize shared presence of feature groups between analyses or replicate groups. In addition, analyses/replicates sharing similar properties (e.g. location, age, type) may be grouped to enhance visualization between these 'outer groups'.

plotChroms Plots extracted ion chromatograms (EICs) of feature groups.

plotVenn plots a Venn diagram (using VennDiagram) outlining unique and shared feature groups between up to five replicate groups.

plotUpSet plots an UpSet diagram (using the upset function) outlining unique and shared feature groups between given replicate groups.

plotVolcano Plots Fold change data in a 'Volcano plot'.

Sets workflows

The following methods are changed or with new functionality:

plotVenn and plotInt allow to handle data per set. See the sets argument description.

References

Gu, Z. (2014) circlize implements and enhances circular visualization in R. Bioinformatics.

Conway JR, Lex A, Gehlenborg N (2017). “UpSetR: an R package for the visualization of intersecting sets and their properties.” Bioinformatics, 33(18), 2938-2940. doi: 10.1093/bioinformatics/btx364 , http://dx.doi.org/10.1093/bioinformatics/btx364.

Lex A, Gehlenborg N, Strobelt H, Vuillemot R, Pfister H (2014). “UpSet: Visualization of Intersecting Sets.” IEEE Transactions on Visualization and Computer Graphics, 20(12), 1983--1992. doi: 10.1109/tvcg.2014.2346248 .

See also