Base class for grouped features.

This class holds all the information for grouped features.

# S4 method for featureGroups
names(x)

# S4 method for featureGroups
analyses(obj)

# S4 method for featureGroups
replicateGroups(obj)

# S4 method for featureGroups
groupNames(obj)

# S4 method for featureGroups
length(x)

# S4 method for featureGroups
show(object)

# S4 method for featureGroups
groups(object)

# S4 method for featureGroups
analysisInfo(obj)

# S4 method for featureGroups
groupInfo(fGroups)

# S4 method for featureGroups
featureTable(obj)

# S4 method for featureGroups
getFeatures(obj)

# S4 method for featureGroups
groupFeatIndex(fGroups)

# S4 method for featureGroups,ANY,ANY,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for featureGroups,ANY,ANY
[[(x, i, j)

# S4 method for featureGroups
$(x, name)

# S4 method for featureGroups
export(fGroups, type, out)

# S4 method for featureGroups
as.data.table(x, average = FALSE,
  features = FALSE)

# S4 method for featureGroups,ANY
plot(x, retMin = FALSE, col = NULL,
  pch = 16, ...)

# S4 method for featureGroups
plotInt(obj, average = FALSE, pch = 20,
  type = "b", lty = 3, col = NULL, ...)

# S4 method for featureGroups
plotChord(obj, addSelfLinks = FALSE,
  addRetMzPlots = TRUE, average = FALSE, outerGroups = NULL,
  addIntraOuterGroupLinks = FALSE, ...)

# S4 method for featureGroups
plotEIC(obj, rtWindow = 30, mzWindow = 0.005,
  retMin = FALSE, topMost = NULL, EICs = NULL,
  showPeakArea = FALSE, showFGroupRect = TRUE, title = NULL,
  colourBy = c("none", "rGroups", "fGroups"), showLegend = TRUE,
  onlyPresent = TRUE, annotate = c("none", "ret", "mz"),
  showProgress = FALSE, xlim = NULL, ylim = NULL, ...)

# S4 method for featureGroups
plotVenn(obj, which = NULL, ...)

# S4 method for featureGroups
plotUpSet(obj, which = NULL,
  nsets = length(which), nintersects = NA, ...)

# S4 method for featureGroups
unique(x, which, relativeTo = NULL,
  outer = FALSE)

# S4 method for featureGroups
overlap(fGroups, which, exclusive = FALSE)

# S4 method for featureGroups
regression(fGroups)

Arguments

fGroups, obj, x, object
featureGroups object to be accessed.

i, j
A numeric or character value which is used to select analyses/feature groups by their index or name, respectively (for the order/names see analyses()/names()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all analyses/feature groups are selected.

For [[: should be a scalar value. If j is not specified, i selects by feature groups instead.

…
Ignored for "[" operator or passed to plot (plot and plotEIC), lines (plotInt), VennDiagram plotting functions (plotVenn), chordDiagram (plotChord) or upset (plotUpSet).

drop
ignored.

name
The feature group name (partially matched).

type
Which file type should be exported: "brukerpa" (Bruker ProfileAnalysis), "brukertasq" (Bruker TASQ) or "mzmine" (MZMine).

out
The destination file for the exported data.

average
Average data within replicate groups.

features
If TRUE then feature specific data will be added. If average=TRUE this data will be averaged for each feature group.

retMin
Plot retention time in minutes (instead of seconds).

col
Colour(s) used. If col=NULL then colours are automatically generated.

pch, type, lty
Common plotting parameters passed to e.g. plot.

addSelfLinks
If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots
Set to TRUE to enable m/z vs retention time scatter plots.

outerGroups
Character vector of names to be used as outer groups. The values in the specified vector should be named by analysis names (average set to FALSE) or replicate group names (average set to TRUE), for instance: c(analysis1 = "group1", analysis2 = "group1", analysis3 = "group2"). Set to NULL to disable outer groups.

addIntraOuterGroupLinks
If TRUE then links will be added within outer groups.

rtWindow
Retention time (in seconds) that will be subtracted/added to respectively the minimum and maximum retention time of the plotted feature groups. Thus, setting this value to a positive value will 'zoom out' on the retention time axis.

mzWindow
The m/z value (in Da) which will be subtracted/added to a feature group m/z value to determine the width of its EIC.

topMost
Only plot EICs from features within this number of top most intense analyses. If NULL then all analyses are used for plotted.

EICs
Internal parameter for now and should be kept at NULL (default).

showPeakArea
Set to TRUE to display integrated chromatographic peak ranges by filling (shading) their areas.

showFGroupRect
Set to TRUE to mark the full retention/intensity range of all features within a feature group by drawing a rectangle around it.

title
Character string used for title of the plot. If NULL a title will be automatically generated.

colourBy
How to colour different extracted ion chromatograms within the plot: "none" for a single colour or "rGroups"/"fGroups" for a distinct colour per replicate/feature group.

showLegend
If TRUE a legend will be shown with either replicate groups (colourBy == "rGroups") or feature groups (colourBy == "fGroups"). If colourBy is "none" no legend will be shown.

onlyPresent
If TRUE then EICs will only be generated for analyses in which a particular feature group was detected. Disabling this option might be useful to see if any features were 'missed'.

annotate
If set to "ret" and/or "mz" then retention and/or m/z values will be drawn for each plotted feature group.

showProgress
if set to TRUE then a text progressbar will be displayed when all EICs are being plot. Set to "none" to disable any annotation.

xlim, ylim
Sets the plot size limits used by plot. Set to NULL for automatic plot sizing.

which
A character vector with replicate groups used for comparison. For plotting functions: set to NULL for all replicate groups.

nsets, nintersects
See upset.

relativeTo
A character vector with replicate groups that should be used for unique comparison. If NULL then all replicate groups are used for comparison. Replicate groups specified in which are ignored.

outer
If TRUE then only feature groups are kept which do not overlap between the specified replicate groups for the which parameter.

exclusive
If TRUE then all feature groups are removed that are not unique to the given replicate groups.

Value

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g. draw.pairwise.venn and draw.triple.venn).

regression returns a data.frame containing retention, m/z, R squared, intercepts and slopes for all feature groups.

Details

The featureGroup class is the workhorse of patRoon: almost all functionality operate on its instantiated objects. The class holds all information from grouped features (obtained from features). This class itself is virtual, hence, objects are not created directly from it. Instead, 'feature groupers' such as groupFeaturesXCMS return a featureGroups derived object after performing the actual grouping of features across analyses.

Methods (by generic)

names: Obtain feature group names.

analyses: returns a character vector with the names of the analyses for which data is present in this object.

replicateGroups: returns a character vector with the names of the replicate groups for which data is present in this object.

groupNames: Same as names. Provided for consistency to other classes.

length: Obtain number of feature groups.

show: Shows summary information for this object.

groups: Accessor for groups slot.

analysisInfo: Obtain analysisInfo (see analysisInfo slot in features).

groupInfo: Accessor for groupInfo slot.

featureTable: Obtain feature information (see features).

getFeatures: Accessor for features slot.

groupFeatIndex: Accessor for ftindex slot.

[: Subset on analyses/feature groups.

[[: Extract intensity values.

$: Extract intensity values for a feature group.

export: Exports feature groups to a .csv file that is readable to Bruker ProfileAnalysis (a 'bucket table'), Bruker TASQ (an analyte database) or that is suitable as input for the Targeted peak detection functionality of MZmine.

as.data.table: Obtain a summary table (a data.table) with retention, m/z, intensity and optionally other feature data.

plot: Generates an m/z vs retention time plot for all featue groups.

plotInt: Generates a line plot for the (averaged) intensity of feature groups within all analyses

plotChord: Generates a chord diagram which can be used to visualize shared presence of feature groups between analyses or replicate groups. In addition, analyses/replicates sharing similar properties (e.g. location, age, type) may be grouped to enhance visualization between these 'outer groups'.

plotEIC: Plots extracted ion chromatograms (EICs) of feature groups.

plotVenn: plots a Venn diagram (using VennDiagram) outlining unique and shared feature groups between up to five replicate groups.

plotUpSet: plots an UpSet diagram (using the upset function) outlining unique and shared feature groups between given replicate groups.

unique: Obtain a subset with unique feature groups present in one or more specified replicate group(s).

overlap: Obtain a subset with feature groups that overlap between a set of specified replicate group(s).

regression: Performs linear regression for all detected feature groups within specified analyses. The analyses to be used (e.g. a set of standards) must have the conc column set in the analysis-information that was used to generate the feature groups.

Slots

groups

Matrix (data.table) with intensities for each feature group (columns) per analysis (rows). Access with groups method.

analysisInfo,features

Analysis info and features class associated with this object. Access with analysisInfo and featureTable methods, respectively.

groupInfo

data.frame with retention time (rts column, in seconds) and m/z (mzs column) for each feature group. Access with groupInfo method.

ftindex

Matrix (data.table) with feature indices for each feature group (columns) per analysis (rows). Each index corresponds to the row within the feature table of the analysis (see featureTable).

S4 class hierarchy

workflowStep

featureGroups

featureGroupsBruker

featureGroupsConsensus

featureGroupsEnviMass

featureGroupsOpenMS

featureGroupsScreening

featureGroupsBrukerTASQ

featureGroupsXCMS

References

Gu, Z. (2014) circlize implements and enhances circular visualization in R. Bioinformatics.

Conway JR, Lex A, Gehlenborg N (2017). “UpSetR: an R package for the visualization of intersecting sets and their properties.” Bioinformatics, 33(18), 2938-2940. doi: 10.1093/bioinformatics/btx364 , http://dx.doi.org/10.1093/bioinformatics/btx364.

Lex A, Gehlenborg N, Strobelt H, Vuillemot R, Pfister H (2014). “UpSet: Visualization of Intersecting Sets.” IEEE Transactions on Visualization and Computer Graphics, 20(12), 1983--1992. doi: 10.1109/tvcg.2014.2346248 , https://doi.org/10.1109/tvcg.2014.2346248.