Base class for grouped features.

This class holds all the information for grouped features.

# S4 method for featureGroups
names(x)

# S4 method for featureGroups
length(x)

# S4 method for featureGroups
show(object)

# S4 method for featureGroups
groups(object)

# S4 method for featureGroups
analysisInfo(obj)

# S4 method for featureGroups
groupInfo(fGroups)

# S4 method for featureGroups
featureTable(obj)

# S4 method for featureGroups
getFeatures(fGroups)

# S4 method for featureGroups
groupFeatIndex(fGroups)

# S4 method for featureGroups,ANY,ANY,ANY
[(x, i, j, ..., drop = TRUE)

# S4 method for featureGroups
export(fGroups, type, out)

# S4 method for featureGroups
groupTable(fGroups, average = FALSE)

# S4 method for featureGroups,ANY
plot(x, retMin = TRUE, ...)

# S4 method for featureGroups
plotInt(obj, average = FALSE, ...)

# S4 method for featureGroups
plotChord(obj, addSelfLinks = FALSE,
  addRetMzPlots = TRUE, average = FALSE, outerGroups = NULL,
  addIntraOuterGroupLinks = FALSE, ...)

# S4 method for featureGroups
plotEIC(obj, rtWindow = 30, mzWindow = 0.005,
  retMin = FALSE, topMost = NULL, EICs = NULL, showPeakArea = FALSE,
  showFGroupRect = TRUE, title = NULL, colourBy = c("none", "rGroups",
  "fGroups"), showLegend = TRUE, onlyPresent = TRUE, annotate = c("none",
  "ret", "mz"), showProgress = FALSE, ...)

# S4 method for featureGroups
plotVenn(obj, which = NULL, ...)

# S4 method for featureGroups
unique(x, which, relativeTo = NULL, outer = FALSE)

# S4 method for featureGroups
overlap(fGroups, which, exclusive = FALSE)

# S4 method for featureGroups
regression(fGroups)

Arguments

fGroups, obj, x, object
featureGroups object to be accessed.

i
Either a numeric, character or logical vector that is used to select analyses by their index, name (for index order/names see analysisInfo()) and logical selection, respectively. If missing all analyses are selected.

j
Either a numeric, character or logical vector that is used to select feature groups by their index, name (for index order/names use groupInfo) and logical selection, respectively. If missing all feature groups are selected.

…
Ignored for "[" operator or passed to plot (plot, plotInt and plotEIC), VennDiagram plotting functions (plotVenn) or chordDiagram (plotChord).

drop
Remove feature groups that are not present anymore in any of the analyses after subsetting.

type
Which file type should be exported: "brukerpa" (Bruker ProfileAnalysis), "brukertasq" (Bruker TASQ) or "mzmine" (MZMine).

out
The destination file for the exported data.

average
Average data within replicate groups.

retMin
Plot retention time in minutes (instead of seconds).

addSelfLinks
If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots
Set to TRUE to enable m/z vs retention time scatter plots.

outerGroups
Character vector of names to be used as outer groups. The values in the specified vector should be named by analysis names (average set to FALSE) or replicate group names (average set to TRUE), for instance: c(analysis1 = "group1", analysis2 = "group1", analysis3 = "group2"). Set to NULL to disable outer groups.

addIntraOuterGroupLinks
If TRUE then links will be added within outer groups.

rtWindow
Retention time (in seconds) that will be subtracted/added to respectively the minimum and maximum retention time of the plotted feature groups. Thus, setting this value to a positive value will 'zoom out' on the retention time axis.

mzWindow
The m/z value (in Da) which will be subtracted/added to a feature group m/z value to determine the width of its EIC.

topMost
Only plot EICs from features within this number of top most intense analyses. If NULL then all analyses are used for plotted.

EICs
Internal parameter for now and should be kept at NULL (default).

showPeakArea
Set to TRUE to display integrated chromatographic peak ranges by filling (shading) their areas.

showFGroupRect
Set to TRUE to mark the full retention/intensity range of all features within a feature group by drawing a rectangle around it.

title
Character string used for title of the plot. If NULL a title will be automatically generated.

colourBy
How to colour different extracted ion chromatograms within the plot: "none" for a single colour or "rGroups"/"fGroups" for a distinct colour per replicate/feature group.

showLegend
If TRUE a legend will be shown with either replicate groups (colourBy == "rGroups") or feature groups (colourBy == "fGroups"). If colourBy is "none" no legend will be shown.

onlyPresent
If TRUE then EICs will only be generated for analyses in which a particular feature group was detected. Disabling this option might be useful to see if any features were 'missed'.

annotate
If set to "ret" and/or "mz" then retention and/or m/z values will be drawn for each plotted feature group.

showProgress
if set to TRUE then a text progressbar will be displayed when all EICs are being plot. Set to "none" to disable any annotation.

which
A character vector with replicate groups used for comparison.

relativeTo
A character vector with replicate groups that should be used for unique comparison. If NULL then all replicate groups are used for comparison. Replicate groups specified in which are ignored.

outer
If TRUE then only feature groups are kept which do not overlap between the specified replicate groups for the which parameter.

exclusive
If TRUE then all feature groups are removed that are not unique to the given replicate groups.

Value

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area (i.e. feature group count) for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g. draw.pairwise.venn and draw.triple.venn).

A data.frame containing retention, m/z, R squared, intercepts and slopes for all feature groups.

Details

The featureGroup class is the workhorse of patRoon: almost all functionality operate on its instantiated objects. The class holds all information from grouped features (obtained from features). This class itself is virtual, hence, objects are not created directly from it. Instead, 'feature groupers' such as groupFeaturesXCMS return a featureGroups derived object after performing the actual grouping of features across analyses.

Methods (by generic)

names: Obtain feature group names.

length: Obtain number of feature groups.

show: Shows summary information for this object.

groups: Accessor for groups slot.

analysisInfo: Obtain analysisInfo (see analysisInfo slot in features).

groupInfo: Accessor for groupInfo slot.

featureTable: Obtain feature information (see features).

getFeatures: Accessor for features slot.

groupFeatIndex: Accessor for ftindex slot.

[: Operator to subset on analyses and/or feature groups.

export: Exports feature groups to a .csv file that is readable to Bruker ProfileAnalysis (a 'bucket table'), Bruker TASQ (an analyte database) or that is suitable as input for the Targeted peak detection functionality of MZmine.

groupTable: Obtain summary table (a data.table) with retention, m/z and intensity information.

plot: Generates an m/z vs retention time plot for all featue groups.

plotInt: Generates a line plot for the (averaged) intensity of feature groups within all analyses

plotChord: Generates a chord diagram which can be used to visualize shared presence of feature groups between analyses or replicate groups. In addition, analyses/replicates sharing similar properties (e.g. location, age, type) may be grouped to enhance visualization between these 'outer groups'.

plotEIC: Plots extracted ion chromatograms (EICs) of feature groups. This function uses the xcms package for loading EIC data.

plotVenn: plots a Venn diagram (using VennDiagram) outlining unique and shared feature groups between up to five replicate groups.

unique: Obtain a subset with unique feature groups present in one or more specified replicate group(s).

overlap: Obtain a subset with feature groups that overlap between a set of specified replicate group(s).

regression: Performs linear regression for all detected feature groups within specified analyses. The analyses to be used (e.g. a set of standards) must have the conc column set in the analysis-information that was used to generate the feature groups.

Slots

groups

Matrix (data.table) with intensities for each feature group (columns) per analysis (rows). Access with groups method.

analysisInfo,features

Analysis info and features class associated with this object. Access with analysisInfo and featureTable methods, respectively.

groupInfo

data.frame with retention time (rts column, in seconds) and m/z (mzs column) for each feature group. Access with groupInfo method.

ftindex

Matrix (data.table) with feature indices for each feature group (columns) per analysis (rows). Each index corresponds to the row within the feature table of the analysis (see featureTable).

References

Gu, Z. (2014) circlize implements and enhances circular visualization in R. Bioinformatics.