Base class for grouped features.

This class holds all the information for grouped features.

# S4 method for featureGroups
names(x)

# S4 method for featureGroups
analyses(obj)

# S4 method for featureGroups
replicateGroups(obj)

# S4 method for featureGroups
groupNames(obj)

# S4 method for featureGroups
length(x)

# S4 method for featureGroups
show(object)

# S4 method for featureGroups
groupTable(object, areas = FALSE)

# S4 method for featureGroups
analysisInfo(obj)

# S4 method for featureGroups
groupInfo(fGroups)

# S4 method for featureGroups
featureTable(obj)

# S4 method for featureGroups
getFeatures(obj)

# S4 method for featureGroups
groupFeatIndex(fGroups)

# S4 method for featureGroups
groupQualities(fGroups)

# S4 method for featureGroups
groupScores(fGroups)

# S4 method for featureGroups
annotations(obj)

# S4 method for featureGroups
adducts(obj)

# S4 method for featureGroups
adducts(obj) <- value

# S4 method for featureGroups,ANY,ANY,missing
[(x, i, j, ..., rGroups, results, drop = TRUE)

# S4 method for featureGroups,ANY,ANY
[[(x, i, j)

# S4 method for featureGroups
$(x, name)

# S4 method for featureGroups
delete(obj, i = NULL, j = NULL, ...)

# S4 method for featureGroups
export(obj, type, out)

getFCParams(rGroups, ...)

# S4 method for featureGroups
as.data.table(
  x,
  average = FALSE,
  areas = FALSE,
  features = FALSE,
  qualities = FALSE,
  regression = FALSE,
  averageFunc = mean,
  normFunc = NULL,
  FCParams = NULL
)

# S4 method for featureGroups,missing
plot(
  x,
  colourBy = c("none", "rGroups", "fGroups"),
  onlyUnique = FALSE,
  retMin = FALSE,
  showLegend = TRUE,
  col = NULL,
  pch = NULL,
  ...
)

# S4 method for featureGroups
plotInt(
  obj,
  average = FALSE,
  normFunc = NULL,
  xnames = TRUE,
  showLegend = FALSE,
  pch = 20,
  type = "b",
  lty = 3,
  col = NULL,
  ...
)

# S4 method for featureGroups
plotChord(
  obj,
  addSelfLinks = FALSE,
  addRetMzPlots = TRUE,
  average = FALSE,
  outerGroups = NULL,
  addIntraOuterGroupLinks = FALSE,
  ...
)

# S4 method for featureGroups
plotChroms(
  obj,
  rtWindow = 30,
  mzExpWindow = 0.001,
  retMin = FALSE,
  topMost = NULL,
  topMostByRGroup = FALSE,
  EICs = NULL,
  showPeakArea = FALSE,
  showFGroupRect = TRUE,
  title = NULL,
  colourBy = c("none", "rGroups", "fGroups"),
  showLegend = TRUE,
  onlyPresent = TRUE,
  annotate = c("none", "ret", "mz"),
  showProgress = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

# S4 method for featureGroups
plotVenn(obj, which = NULL, ...)

# S4 method for featureGroups
plotUpSet(obj, which = NULL, nsets = length(which), nintersects = NA, ...)

# S4 method for featureGroups
plotVolcano(
  obj,
  FCParams,
  showLegend = TRUE,
  averageFunc = mean,
  col = NULL,
  pch = 19,
  ...
)

# S4 method for featureGroups
unique(x, which, relativeTo = NULL, outer = FALSE)

# S4 method for featureGroups
overlap(fGroups, which, exclusive)

# S4 method for featureGroups
calculatePeakQualities(
  obj,
  weights,
  flatnessFactor,
  avgFunc = mean,
  parallel = TRUE
)

# S4 method for featureGroups
selectIons(
  fGroups,
  components,
  prefAdduct,
  onlyMonoIso = TRUE,
  chargeMismatch = "adduct"
)

# S4 method for featureGroupsSet
sets(obj)

# S4 method for featureGroupsSet
adducts(obj, set, ...)

# S4 method for featureGroupsSet
adducts(obj, set, reGroup = TRUE) <- value

# S4 method for featureGroupsSet
delete(obj, i = NULL, j = NULL, ...)

# S4 method for featureGroupsSet
show(object)

# S4 method for featureGroupsSet
featureTable(obj)

# S4 method for featureGroupsSet,ANY,ANY,missing
[(x, i, j, ..., rGroups, sets = NULL, drop = TRUE)

# S4 method for featureGroupsSet
export(obj, type, out, set)

# S4 method for featureGroupsSet
as.data.table(
  x,
  average = FALSE,
  areas = FALSE,
  features = FALSE,
  qualities = FALSE,
  regression = FALSE,
  averageFunc = mean,
  normFunc = NULL,
  FCParams = NULL
)

# S4 method for featureGroupsSet
filter(
  obj,
  ...,
  negate = FALSE,
  sets = NULL,
  absMinSets = NULL,
  relMinSets = NULL
)

# S4 method for featureGroupsSet
plotInt(
  obj,
  average = FALSE,
  normFunc = NULL,
  xnames = !sets,
  showLegend = sets,
  pch = 20,
  type = "b",
  lty = 3,
  col = NULL,
  ...,
  sets = FALSE
)

# S4 method for featureGroupsSet
plotVenn(obj, which = NULL, ..., sets = FALSE)

# S4 method for featureGroupsSet
unique(x, which, ..., sets = FALSE)

# S4 method for featureGroupsSet
overlap(fGroups, which, exclusive, sets = FALSE)

# S4 method for featureGroupsSet
selectIons(fGroups, components, prefAdduct, ...)

# S4 method for featureGroupsSet
unset(obj, set)

Arguments

areas
If set to TRUE then areas are considered instead of peak intensities.

For as.data.table: ignored if features=TRUE, as areas of features are always reported.

fGroups, obj, x, object
featureGroups object to be accessed.

value
For adducts<-: A character with adduct annotations assigned to each feature group. The length should equal the number of feature groups. Can be named with feature group names to customize the assignment order.

i, j
For [/[[: A numeric or character value which is used to select analyses/feature groups by their index or name, respectively (for the order/names see analyses()/names()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all analyses/feature groups are selected.

For [[: should be a scalar value. If j is not specified, i selects by feature groups instead.

For delete: The data to remove from. i are the analyses as numeric index, logical or character, j the feature groups as numeric index, logical or character. If either is NULL then data for all is removed. j may also be a function: it will be called for each feature group, with a vector of the group intensities as first argument, the group name as second argument, and any other arguments passed as ... to delete. The return value of this function specifies the analyses of the features in the group to be removed (same format as i).

...
For the "[" operator: ignored.

For delete: passed to the function specified as j.

For getFCParams: optional named arguments that override defaults.

For sets workflow methods: further arguments passed to the base featureGroups method.

Otherwise passed to plot (plot and plotChroms), lines (plotInt), VennDiagram plotting functions (plotVenn), chordDiagram (plotChord) or upset (plotUpSet).

rGroups
For [: An optional character vector: if specified only keep results for the given replicate groups (equivalent to the rGroups argument to filter).

For getFCParams: a character vector with the names of the two replicate groups to be compared.

results
Optional argument. If specified only feature groups with results in the specified object are kept. The class of results should be featureAnnotations or components. Multiple objects can be specified in a list: in this case a feature group is kept if it has a result in any of the objects (equivalent to the results argument to filter).

drop
ignored.

name
The feature group name (partially matched).

out
The destination file for the exported data.

average
If TRUE then data within replicate groups are averaged.

For as.data.table: if features=TRUE other feature properties are also averaged.

features
If TRUE then feature specific data will be added. If average=TRUE this data will be averaged for each feature group.

qualities
Adds feature (group) qualities (qualities="quality"), scores (qualities="score") or both (qualities="both"), if this data is available (i.e. from calculatePeakQualities). If qualities=FALSE then nothing is reported.

regression
Set to TRUE to add regression data for each feature group. For this a linear model is created (intensity/area [depending on areas argument] vs concentration). The model concentrations (e.g. of a set of standards) is derived from the conc column of the analysis information. From this model the intercept, slope and R2 is added to the output. In addition, when features=TRUE, concentrations for each feature are added. Note that no regression information is added when no conc column is present in the analysis information or when less than two concentrations are specified (i.e. the minimum amount).

averageFunc
Function used for averaging.

for as.data.table: only used when average=TRUE or FCParams != NULL.

normFunc
Function that should be used for normalization of data. The function is called for all intensities/areas of a feature group and these quantities are divided by the result of the function call. For example, when max is used normalized intensities will be between zero and one. If all quantities are zero then the function will not be called. Set to NULL to perform no normalization.

FCParams
A parameter list to calculate Fold change data (see getFCParams and the Fold change calculation section).

For as.data.table: set to NULL to not perform FC calculations.

colourBy
Sets the automatic colour selection: "none" for a single colour or "rGroups"/"fGroups" for a distinct colour per replicate/feature group.

onlyUnique
If TRUE and colourBy="rGroups" then only feature groups that are unique to a replicate group are plotted.

retMin
Plot retention time in minutes (instead of seconds).

showLegend
Plot a legend if TRUE.

col
Colour(s) used. If col=NULL then colours are automatically generated.

pch, type, lty
Common plotting parameters passed to e.g. plot. For plot: if pch=NULL then values are automatically assigned.

xnames
Plot analysis (or replicate group if average=TRUE) names on the x axis.

addSelfLinks
If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots
Set to TRUE to enable m/z vs retention time scatter plots.

outerGroups
Character vector of names to be used as outer groups. The values in the specified vector should be named by analysis names (average set to FALSE) or replicate group names (average set to TRUE), for instance: c(analysis1 = "group1", analysis2 = "group1", analysis3 = "group2"). Set to NULL to disable outer groups.

addIntraOuterGroupLinks
If TRUE then links will be added within outer groups.

rtWindow
Retention time (in seconds) that will be subtracted/added to respectively the minimum and maximum retention time of the plotted feature groups. Thus, setting this value to a positive value will 'zoom out' on the retention time axis.

mzExpWindow
In case the m/z window to plot an EIC for a particular analysis is not known (i.e. no feature was detected of the feature group to be plot and onlyPresent=FALSE) then the EIC m/z range is estimated from the range for the complete feature group and expanded by the offset defined by mzExpWindow.

topMost
Only plot EICs from features within this number of top most intense analyses. If NULL then all analyses are used for plotted.

topMostByRGroup
If set to TRUE and topMost is set: only plot EICs for the top most features in each replicate group. For instance, when topMost=1 and topMostByRGroup=TRUE, then EICs will be plotted for the most intense feature of each replicate group.

EICs
Internal parameter for now and should be kept at NULL (default).

showPeakArea
Set to TRUE to display integrated chromatographic peak ranges by filling (shading) their areas.

showFGroupRect
Set to TRUE to mark the full retention/intensity range of all features within a feature group by drawing a rectangle around it.

title
Character string used for title of the plot. If NULL a title will be automatically generated.

onlyPresent
If TRUE then EICs will only be generated for analyses in which a particular feature group was detected. Disabling this option might be useful to see if any features were 'missed'.

annotate
If set to "ret" and/or "mz" then retention and/or m/z values will be drawn for each plotted feature group.

showProgress
if set to TRUE then a text progressbar will be displayed when all EICs are being plot. Set to "none" to disable any annotation.

xlim, ylim
Sets the plot size limits used by plot. Set to NULL for automatic plot sizing.

which
A character vector with replicate groups used for comparison.

For plotting functions: set to NULL for all replicate groups.

For plotVenn: alternatively a named list containing elements of character vectors with replicate groups to compare. For instance, which=list(infl = c("influent-A", "influent-B"), effl = c("effluent-A", "effluent-B")), will compare the features in replicate groups "influent-A/B" against those in "effluent-A/B". The names of the list are used for labelling in the plot, and will be made automatically if not specified.

nsets, nintersects
See upset.

relativeTo
A character vector with replicate groups that should be used for unique comparison. If NULL then all replicate groups are used for comparison. Replicate groups specified in which are ignored.

outer
If TRUE then only feature groups are kept which do not overlap between the specified replicate groups for the which parameter.

exclusive
If TRUE then all feature groups are removed that are not unique to the given replicate groups.

weights
A named numeric vector that defines the weight for each score to calculate the totalScore. The names of the vector follow the score names. Unspecified weights are defaulted to 1. Example: weights=c(ApexBoundaryRatioScore=0.5, GaussianSimilarityScore=2).

flatnessFactor
Passed to MetaClean as the flatness.factor argument to calculateJaggedness and calculateModality.

avgFunc
The function used to average the peak qualities and scores for each feature group.

parallel
If set to TRUE then code is executed in parallel through the futures package. Please see the parallelization section in the handbook for more details.

components
The components object that was generated for the given featureGroups object. Obviously, the components must be created with algorithms that support adduct/isotope annotations, such as those from RAMClustR and cliqueMS.

prefAdduct
The 'preferred adduct' (see method description). This is often "[M+H]+" or "[M-H]-".

onlyMonoIso
Set to TRUE to only keep feature groups that were annotated as monoisotopic. Feature groups are never removed by this setting if no isotope annotations are available.

chargeMismatch
Specifies how to deal with a mismatch in charge between adduct and isotope annotations. Valid values are: "adduct" (ignore isotope annotation), "isotope" (ignore adduct annotation), "none" (ignore both annotations) and "ignore" (don't check for charge mismatches). Important: when OpenMS is used to find features, it already removes any detected non-monoisotopic features by default. Hence, in such case setting chargeMismatch="adduct" is more appropriate.

set
(sets workflow) The name of the set.

reGroup
Set to TRUE to re-group the features after the adduct annotations are changed. See the Sets workflow section for more details.

sets
(sets workflow) For [ and filter: a character with name(s) of the sets to keep (or remove if negate=TRUE).

For plotInt: if TRUE then feature intensities are plot per set (order follows the analysis information).

For plotVenn, overlap and unique: If TRUE then the which argument changes its meaning and is used to specify the names of the sets to be compared.

negate
If set to TRUE then filtering operations are performed in opposite manner.

absMinSets, relMinSets
(sets workflow) Feature groups are only kept when they contain data for at least this (absolute or relative) amount of sets. Set to NULL to ignore.

Value

delete returns the object for which the specified data was removed.

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g. draw.pairwise.venn and draw.triple.venn).

calculatePeakQualities returns a modified object amended with peak qualities and scores.

selectIons returns a featureGroups object with only the selected feature groups and amended with adduct annotations.

Details

The featureGroup class is the workhorse of patRoon: almost all functionality operate on its instantiated objects. The class holds all information from grouped features (obtained from features). This class itself is virtual, hence, objects are not created directly from it. Instead, 'feature groupers' such as groupFeaturesXCMS return a featureGroups derived object after performing the actual grouping of features across analyses.

getFCParams creates a parameter list to calculate Fold changes (see the Fold change calculation section).

Methods (by generic)

names: Obtain feature group names.

analyses: returns a character vector with the names of the analyses for which data is present in this object.

replicateGroups: returns a character vector with the names of the replicate groups for which data is present in this object.

groupNames: Same as names. Provided for consistency to other classes.

length: Obtain number of feature groups.

show: Shows summary information for this object.

groupTable: Accessor for groups slot.

analysisInfo: Obtain analysisInfo (see analysisInfo slot in features).

groupInfo: Accessor for groupInfo slot.

featureTable: Obtain feature information (see features).

getFeatures: Accessor for features slot.

groupFeatIndex: Accessor for ftindex slot.

groupQualities: Accessor for groupQualities slot.

groupScores: Accessor for groupScores slot.

annotations: Accessor for annotations slot.

adducts: Returns a named character with adduct annotations assigned to each feature group (if available).

adducts<-: Sets adduct annotations for feature groups.

[: Subset on analyses/feature groups.

[[: Extract intensity values.

$: Extract intensity values for a feature group.

delete: Completely deletes specified feature groups.

export: Exports feature groups to a .csv file that is readable to Bruker ProfileAnalysis (a 'bucket table'), Bruker TASQ (an analyte database) or that is suitable as input for the Targeted peak detection functionality of MZmine.

as.data.table: Obtain a summary table (a data.table) with retention, m/z, intensity and optionally other feature data.

plot: Generates an m/z vs retention time plot for all featue groups. Optionally highlights unique/overlapping presence amongst replicate groups.

plotInt: Generates a line plot for the (averaged) intensity of feature groups within all analyses

plotChord: Generates a chord diagram which can be used to visualize shared presence of feature groups between analyses or replicate groups. In addition, analyses/replicates sharing similar properties (e.g. location, age, type) may be grouped to enhance visualization between these 'outer groups'.

plotChroms: Plots extracted ion chromatograms (EICs) of feature groups.

plotVenn: plots a Venn diagram (using VennDiagram) outlining unique and shared feature groups between up to five replicate groups.

plotUpSet: plots an UpSet diagram (using the upset function) outlining unique and shared feature groups between given replicate groups.

plotVolcano: Plots Fold change data in a 'Volcano plot'.

unique: Obtain a subset with unique feature groups present in one or more specified replicate group(s).

overlap: Obtain a subset with feature groups that overlap between a set of specified replicate group(s).

calculatePeakQualities: Calculates peak and group qualities for all features and feature groups. The peak qualities (and scores) are calculated with the features method of this function, and subsequently averaged per feature group. Then, MetaClean is used to calculate the Elution Shift and Retention Time Consistency group quality metrics (see the MetaClean publication cited below for more details). Similarly to the features method, these metrics are scored by normalizing qualities among all groups and scaling them from 0 (worst) to 1 (best). The totalScore for each group is then calculated as the weighted sum from all feature (group) scores. The getMCTrainData and predictCheckFeaturesSession functions can be used to train and apply Pass/Fail ML models from MetaClean.

selectIons: uses componentization results to select feature groups with preferred adduct ion and/or isotope annotation. Typically, this means that only feature groups are kept if they are (de-)protonated adducts and are monoisotopic. The adduct annotation assignments for the selected feature groups are copied from the components to the annotations slot. If the adduct for a feature group is unknown, its annotation is defaulted to the 'preferred' adduct, and hence, the feature group will never be removed. Furthermore, if a component does not contain an annotation with the preferred adduct, the most intense feature group is selected instead. Similarly, if no isotope annotation is available, the feature group is assumed to be monoisotopic and thus not removed. An important advantage of selectIons is that it may considerably simplify your dataset. Furthermore, the adduct assignments allow formula/compound annotation steps later in the workflow to improve their annotation accuracy. On the other hand, it is important the componentization results are reliable. Hence, it is highly recommended that, prior to calling selectIons, the settings to generateComponents are optimized and its results are reviewed with checkComponents. Finally, the adducts<- method can be used to manually correct adduct assignments afterwards if necessary.

Slots

groups

Matrix (data.table) with intensities for each feature group (columns) per analysis (rows). Access with groups method.

analysisInfo,features

Analysis info and features class associated with this object. Access with analysisInfo and featureTable methods, respectively.

groupInfo

data.frame with retention time (rts column, in seconds) and m/z (mzs column) for each feature group. Access with groupInfo method.

ftindex

Matrix (data.table) with feature indices for each feature group (columns) per analysis (rows). Each index corresponds to the row within the feature table of the analysis (see featureTable).

groupQualities,groupScores

A data.table with qualities/scores for each feature group (see the calculatePeakQualities method).

annotations

A data.table with adduct annotations for each group (see the selectIons method).

groupAlgo,groupArgs,groupVerbose

(sets workflow) Grouping parameters that were used when this object was created. Used by adducts<- and selectIons when these methods perform a re-grouping of features.

annotations

(sets workflow) As the featureGroups slot, but contains the annotation data per set.

Fold change calculation

the as.data.table and plotVolcano methods can be used to calculate/plot (log2) Fold changes (FC) between two replicate groups to easily identify significant changes. The calculation process is configured through a pramater list, which can be constructed with the getFCParams function. The parameter list has the following entries:

rGroups the name of the two replicate groups to compare (taken from the rGroups argument to getFCParams).

thresholdFC: the threshold log FC for a feature group to be classified as increasing/decreasing.

thresholdPV: the threshold log P for a feature group to be significantly different.

zeroMethod,zeroValue: how to handle zero values when calculating the FC: add adds an offset to zero values, "fixed" sets zero values to a fixed number and "omit" removes zero data. The number that is added/set by the former two options is defined by zeroValue.

PVTestFunc: a function that is used to calculate P values (usually using t.test).

PVAdjFunc: a function that is used to adjust P values (usually using p.adjust)

S4 class hierarchy

workflowStep

featureGroups

featureGroupsSet

featureGroupsScreeningSet

featureGroupsUnset

featureGroupsScreening

featureGroupsSetScreeningUnset

featureGroupsBruker

featureGroupsConsensus

featureGroupsEnviMass

featureGroupsKPIC2

featureGroupsOpenMS

featureGroupsSIRIUS

featureGroupsBrukerTASQ

featureGroupsXCMS

featureGroupsXCMS3

Sets workflows

The featureGroupsSet class is applicable for sets workflows. This class is derived from featureGroups and therefore largely follows the same user interface.

The following methods are specifically defined for sets workflows:

sets Returns the set names for this object.

unset Converts the object data for a specified set into a 'non-set' object (featureGroupsUnset), which allows it to be used in 'regular' workflows. The adduct annotations for the selected set are used to convert all feature (group) masses to ionic m/z values. The annotations persist in the converted object.

The following methods are changed or with new functionality:

adducts, adducts<- require the set argument. The order of the data that is returned/changed follows that of the annotations slot. Furthermore, adducts<- will perform a re-grouping of features when its reGroup parameter is set to TRUE. The implications for this are discussed below.

filter and the subset operator ([) have specific arguments to choose/filter by (feature presence in) sets. See the argument descriptions.

as.data.table: normalization of intensities is performed per set.

export Only allows to export data from one set. The unset method is used prior to exporting the data.

overlap, unique, plotVenn, plotInt allow to handle data per set. See the sets argument description.

selectIons Will perform a re-grouping of features. The implications of this are discussed below.

A re-grouping of features occurs if selectIons is called or adducts<- is used with reGroup=TRUE. Afterwards, it is very likely that feature group names are changed. Since data generated later in the workflow (e.g. annotation steps) rely on feature group names, these objects are not valid anymore, and must be re-generated.

References

Gu, Z. (2014) circlize implements and enhances circular visualization in R. Bioinformatics.

Conway JR, Lex A, Gehlenborg N (2017). “UpSetR: an R package for the visualization of intersecting sets and their properties.” Bioinformatics, 33(18), 2938-2940. doi: 10.1093/bioinformatics/btx364 , http://dx.doi.org/10.1093/bioinformatics/btx364.

Lex A, Gehlenborg N, Strobelt H, Vuillemot R, Pfister H (2014). “UpSet: Visualization of Intersecting Sets.” IEEE Transactions on Visualization and Computer Graphics, 20(12), 1983--1992. doi: 10.1109/tvcg.2014.2346248 .

Chetnik K, Petrick L, Pandey G (2020). “MetaClean: a machine learning-based classifier for reduced false positive peak detection in untargeted LC-MS metabolomics data.” Metabolomics, 16(11). doi: 10.1007/s11306-020-01738-3 .

Author

The code to calculate and plot Fold change data was created by Bas van de Velde.