Comparing feature groups

Functionality to compare feature groups and make a consensus.

# S4 method for featureGroups
comparison(..., groupAlgo,
  groupArgs = list(rtalign = FALSE))

# S4 method for featureGroupsComparison,ANY
plot(x, retMin = TRUE, ...)

# S4 method for featureGroupsComparison
plotVenn(obj, which = NULL, ...)

# S4 method for featureGroupsComparison
plotUpSet(obj, which = names(obj),
  ...)

# S4 method for featureGroupsComparison
plotChord(obj, addSelfLinks = FALSE,
  addRetMzPlots = TRUE, ...)

# S4 method for featureGroupsComparison
consensus(obj, relAbundance = 0,
  absAbundance = 0)

# S4 method for featureGroupsComparison
unique(x, which, ...)

Arguments

…
For comparison: featureGroups objects that should be compared. Objects may be labeled by naming the parameters, otherwise labels are automatically determined. For plot, plotVenn, plotChord, unique: further options passed to plot, VennDiagram plotting functions (e.g. draw.pairwise.venn), chordDiagram and unique (featureGroups signature), respectively. For plotUpSet: any further arguments passed to the plotUpSet method defined for featureGroups.

groupAlgo
The feature grouping algorithm that should be used for grouping pseudo features (see details). Valid values are: "xcms" or "openms".

groupArgs
A list containing further parameters for feature grouping.

x, obj
The featureGroupsComparison object.

retMin
If TRUE retention times are plotted as minutes (seconds otherwise).

which
A character vector specifying one or more labels of compared feature groups. For plotVenn: if NULL then all compared groups are used.

addSelfLinks
If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots
Set to TRUE to enable m/z vs retention time scatter plots.

relAbundance, absAbundance
Feature groups with relative/absolute abundance below this number will be removed from the consensus. Set to 0 for no limits.

Value

comparison returns a featureGroupsComparison object.

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area (i.e. feature group count) for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g. draw.pairwise.venn and draw.triple.venn).

consensus returns a featureGroups object with a consensus from the compared feature groups.

Details

Feature groups objects originating from differing feature finding and/or grouping algorithms (or their parameters) may be compared to assess their output and generate a consensus.

The comparison method generates a featureGroupsComparison object from given feature groups objects, which in turn may be used for (visually) comparing presence of feature groups and generating a consensus. Internally, this function will collapse each feature groups object to pseudo features objects by averaging their retention times, m/z values and intensities, where each original feature groups object becomes an 'analysis'. All pseudo features are then grouped using regular feature grouping algorithms so that a comparison can be made.

plot generates an m/z vs retention time plot.

plotVenn plots a Venn diagram outlining unique and shared feature groups between up to five compared feature groups.

plotUpSet plots an UpSet diagram outlining unique and shared feature groups.

plotChord plots a chord diagram to visualize the distribution of feature groups.

consensus combines all compared feature groups and averages their retention, m/z and intensity data.

unique returns unique feature groups from one or more compared feature groups.