Feature groups comparison class

This class is used for comparing different featureGroups objects.

# S4 method for featureGroupsComparison
names(x)

# S4 method for featureGroupsComparison
length(x)

# S4 method for featureGroupsComparison,ANY,ANY,ANY
[(x, i, j = "missing",
  drop = "missing")

Arguments

x
A featureGroupsComparison object.

i
Character, numeric or logical vector used for subsetting (analogous to a list).

j, drop
Not used. Should be ignored.

Details

Objects from this class are returned by comparison.

Methods (by generic)

names: Obtain the labels that were given to each compared feature group.

length: Number of feature groups objects that were compared.

[: Creates a subset of the compared feature groups.

Slots

fGroupsList

A list of featureGroups object that were compared

comparedFGroups

A pseudo featureGroups object containing grouped feature groups.