Base features class

Holds information for all features present within a set of analysis.

# S4 method for features
length(x)

# S4 method for features
show(object)

# S4 method for features
featureTable(obj)

# S4 method for features
analysisInfo(obj)

Arguments

obj, x, object
features object to be accessed

Value

featureTable: A list containing a data.table for each analysis with feature data

analysisInfo: A data.frame containing a column with analysis name (analysis), its path (path), and other columns such as replicate group name (group) and blank reference (ref).

Details

This class provides a way to store intensity, retention times, m/z and other data for all features in a set of analyses. The class is virtual and derived objects are created by 'feature finders' such as findFeaturesOpenMS, findFeaturesXCMS and findFeaturesBruker.

Methods (by generic)

length: Obtain total number of features.

show: Shows summary information for this object.

featureTable: Get table with feature information

analysisInfo: Get analysis information

Slots

features

List of features per analysis file. Use the featureTable method for access.

analysisInfo

Analysis group information. Use the analysisInfo method for access.

See also

feature-finding