Find features using KPIC2

Uses the KPIC2 R package to extract features.

findFeaturesKPIC2(
  analysisInfo,
  kmeans = TRUE,
  level = 1000,
  ...,
  parallel = TRUE,
  verbose = TRUE
)

Arguments

analysisInfo

A data.frame with Analysis information.

kmeans

If TRUE then getPIC.kmeans is used to obtain PICs, otherwise it is getPIC.

level

Passed to getPIC or getPIC.kmeans

...

Further parameters passed to getPIC/getPIC.kmeans

parallel

If set to TRUE then code is executed in parallel through the futures package. Please see the parallelization section in the handbook for more details.

verbose

If set to FALSE then no text output is shown.

Value

An object of a class which is derived from features.

Details

This function uses KPIC2 to automatically find features. This function is called when calling findFeatures with algorithm="kpic2".

The MS files should be in the mzML or mzXML format.

The input MS data files need to be centroided. The convertMSFiles function can be used to centroid data.

References

Ji H, Zeng F, Xu Y, Lu H, Zhang Z (2017). “KPIC2: An Effective Framework for Mass Spectrometry-Based Metabolomics Using Pure Ion Chromatograms.” Analytical Chemistry, 89(14), 7631--7640. doi:10.1021/acs.analchem.7b01547 .

See also