Find features using SAFD

Uses SAFD to obtain features. This functionality is still experimental. Please see the details below.

findFeaturesSAFD(
  analysisInfo,
  profPath = NULL,
  mzRange = c(0, 400),
  maxNumbIter = 1000,
  maxTPeakW = 300,
  resolution = 30000,
  minMSW = 0.02,
  RThreshold = 0.75,
  minInt = 2000,
  sigIncThreshold = 5,
  S2N = 2,
  minPeakWS = 3,
  verbose = TRUE
)

Arguments

analysisInfo

A data.frame with Analysis information.

profPath

A character vector with paths to the profile MS data for each analysis (will be re-cycled if necessary). See the Using SAFD section for more details.

mzRange

The m/z window to be imported (passed to the import_files_MS1 function).

maxNumbIter, maxTPeakW, resolution, minMSW, RThreshold, minInt, sigIncThreshold, S2N, minPeakWS

Parameters directly passed to the safd_s3D function.

verbose

If set to FALSE then no text output is shown.

Value

An object of a class which is derived from features.

Details

This function uses SAFD to automatically find features. This function is called when calling findFeatures with algorithm="safd".

The support for SAFD is still experimental, and its interface might change in the future.

In order to use SAFD, please make sure that its julia packages are installed and you have verified that everything works, e.g. by running the test data.

This algorithm supports profile and centroided MS data. If the use of profile data is desired, centroided data must still be available for other functionality of patRoon. The centroided data is specified through the 'regular' analysis info mechanism. The location to any profile data is specified through the profPath argument (NULL for no profile data). The base file names (i.e. the file name without path and extension) of both centroid and profile data must be the same. Furthermore, the format of the profile data must be mzXML.

Parallelization

findFeaturesSAFD uses multiprocessing to parallelize computations. Please see the parallelization section in the handbook for more details and patRoon options for configuration options.

Note that for caching purposes, the analyses files must always exist on the local host computer, even if it is not participating in computations.

References

Samanipour S, OBrien JW, Reid MJ, Thomas KV (2019). “Self Adjusting Algorithm for the Nontargeted Feature Detection of High Resolution Mass Spectrometry Coupled with Liquid Chromatography Profile Data.” Analytical Chemistry, 91(16), 10800--10807. doi: 10.1021/acs.analchem.9b02422 .

See also