Find features using SIRIUS

Uses SIRIUS to find features.

findFeaturesSIRIUS(analysisInfo, verbose = TRUE)

Arguments

analysisInfo

A data.frame with Analysis information.

verbose

If set to FALSE then no text output is shown.

Value

An object of a class which is derived from features.

Details

This function uses SIRIUS to automatically find features. This function is called when calling findFeatures with algorithm="sirius".

The features are collected by running the lcms-align SIRIUS command for every analysis.

The MS files should be in the mzML or mzXML format. Furthermore, this algorithms requires the presence of (data-dependent) MS/MS data.

The input MS data files need to be centroided. The convertMSFiles function can be used to centroid data.

Parallelization

findFeaturesSIRIUS uses multiprocessing to parallelize computations. Please see the parallelization section in the handbook for more details and patRoon options for configuration options.

Note that for caching purposes, the analyses files must always exist on the local host computer, even if it is not participating in computations.

References

Duhrkop K, Fleischauer M, Ludwig M, Aksenov AA, Melnik AV, Meusel M, Dorrestein PC, Rousu J, Bocker S (2019). “SIRIUS 4: a rapid tool for turning tandem mass spectra into metabolite structure information.” Nature Methods, 16(4), 299--302. doi:10.1038/s41592-019-0344-8 .

See also