Automatic chemical formula generation

Functionality to automatically calculate chemical formulae for all feature groups.

# S4 method for featureGroups
generateFormulas(fGroups, algorithm, ...)

generateFormulasDA(
  fGroups,
  precursorMzSearchWindow = 0.002,
  MSMode = "both",
  adduct,
  featThreshold = 0.75,
  save = TRUE,
  close = save
)

generateFormulasGenForm(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = "[M+H]+",
  elements = "CHNOP",
  hetero = TRUE,
  oc = FALSE,
  extraOpts = NULL,
  calculateFeatures = TRUE,
  featThreshold = 0.75,
  MSMode = "both",
  isolatePrec = TRUE,
  timeout = 120,
  topMost = 50,
  maxProcAmount = getOption("patRoon.maxProcAmount"),
  batchSize = 8
)

generateFormulasSirius(
  fGroups,
  MSPeakLists,
  relMzDev = 5,
  adduct = "[M+H]+",
  elements = "CHNOP",
  profile = "qtof",
  database = NULL,
  noise = NULL,
  topMost = 100,
  extraOpts = NULL,
  calculateFeatures = TRUE,
  featThreshold = 0.75,
  logPath = file.path("log", "sirius"),
  maxProcAmount = getOption("patRoon.maxProcAmount")
)

formulaScorings()

Arguments

fGroups
featureGroups object for which formulae should be generated. This should be the same or a subset of the object that was used to create the specified MSPeakLists (only relevant for algorithms using MSPeakLists). In the case of a subset only the remaining feature groups in the subset are considered.

algorithm
A character string describing the algorithm that should be used: "bruker", "genform", "sirius"

...
Any parameters to be passed to the selected formula generation algorithm.

precursorMzSearchWindow
Search window for m/z values (+/- the feature m/z) used to find back feature data of precursor/parent ions from MS/MS spectra (this data is not readily available from SmartFormula3D results).

MSMode
Whether formulae should be generated only from MS data ("ms"), MS/MS data ("msms") or both ("both").

For GenForm selecting "both" will fall back to formula calculation with only MS data in case no MS/MS data is available.

For calulation with Bruker DataAnalysis selecting "both" will calculate formulae from MS data and MS/MS data and combines the results (duplicated formulae are removed). This is useful when poor MS/MS data would exclude proper candidates.

adduct
An adduct object (or something that can be converted to it with as.adduct). Examples: "[M-H]-", "[M+Na]+".

featThreshold
If calculateFeatures=TRUE: minimum presence (0-1) of a formula in all features before it is considered as a candidate for a feature group. For instance, featThreshold dictates that a formula should be present in at least 75 feature group.

close, save
If TRUE then Bruker files are closed and saved after processing with DataAnalysis, respectively. Setting close=TRUE prevents that many analyses might be opened simultaneously in DataAnalysis, which otherwise may use excessive memory or become slow. By default save is TRUE when close is TRUE, which is likely what you want as otherwise any processed data is lost.

MSPeakLists
An MSPeakLists object that was generated for the supplied fGroups.

relMzDev
Maximum relative deviation between the measured and candidate formula m/z values (in ppm). Sets the ppm and --ppm-max commandline options for GenForm and SIRIUS, respectively.

elements
Elements to be considered for formulae calculation. This will heavily affects the number of candidates! Always try to work with a minimal set by excluding elements you don't expect. For generateFormulasSirius, the minimum/maximum number of elements can also be specified, for example: a value of "C[5]H[10-15]O" will only consider formulae with up to five carbon atoms, between ten and fifteen hydrogen atoms and any amount of oxygen atoms. Sets the el and --elements commandline options for GenForm and SIRIUS, respectively.

hetero
Only consider formulae with at least one hetero atom. Sets the het commandline option.

oc
Only consider organic formulae (i.e. with at least one carbon atom). Sets the oc commandline option.

extraOpts
An optional character vector with any other commandline options that will be passed to GenForm or SIRIUS. See the GenForm options section/SIRIUS manual for all available commandline options.

calculateFeatures
If TRUE fomulae are first calculated for all features prior to feature group assignment (see details).

isolatePrec
Settings used for isolation of precursor mass peaks and their isotopes. This isolation is highly important for accurate isotope scoring of candidates, as non-relevant mass peaks will dramatically decrease the score. The value of isolatePrec should either be a list with parameters (see the filter method for MSPeakLists for more details), TRUE for default parameters (the z parameter is automatically deduced from the adduct argument) or FALSE for no isolation (e.g. when you already performed isolation with the filter method).

timeout
Maximum time (in seconds) that a GenForm command is allowed to execute. If this time is exceeded a warning is emitted and the command is terminated. See the notes section for more information on the need of timeouts.

topMost
Only keep this number of candidates (per feature group) with highest score. For SIRIUS: Sets the --candidates commandline option.

maxProcAmount
Maximum number of processes to run for parallelization. Usually a number close to the amount of physical cores yields most efficient results.

batchSize
Maximum number of GenForm commands that should be run sequentially in each parallel process. Combining commands with short runtimes (such as GenForm) can significantly increase parallel performance. For more information see executeMultiProcess.

profile
Name of the configuration profile, for example: "qtof", "orbitrap", "fticr". Sets the --profile commandline option.

database
If not NULL, use a database for retrieval of formula candidates. Possible values are: "pubchem", "bio", "kegg", "hmdb". Sets the --database commandline option.

noise
Median intensity of the noise (NULL ignores this parameter). Sets the --noise commandline option.

logPath
Destination directory for log files with output from executed commands. Will be created if non-existant. Set to NULL to disable logging.

Value

A formulas object containing all generated formulae.

Details

Several algorithms are provided to automatically generate formulae for given feature groups. All tools use the accurate mass of a feature to back-calculate candidate formulae. Depending on the algorithm and data availability, other data such as isotopic pattern and MS/MS fragments may be used to further improve formula assignment and ranking.

When DataAnalysis is used for formula generation or calculateFeatures=TRUE formulae are first calculated for each feature. The results are then combined for final assignment of candidate formulae for each feature group. If a formula was found in multiple features within the group, the reported scorings and mass errors are averaged and other numeric values are those from the feature in the analysis of the "analysis" column. The calculation of formulae on 'feature level' might result in a more thorough formula search and better removal of outliers (controlled by featThreshold argument). In contrast, when calculations occur on 'feature group level' (i.e. calculateFeatures=FALSE), formulae are directly assigned to each feature group (by using group averaged peak MS lists), which significantly reduces processing time is, especially with many analyses. Note that in both situations subsequent algorithms that use formula data (e.g. addFormulaScoring and reporting functions) only use formula data that was eventually assigned to feature groups. Furthermore, please note that calculation of formulae with DataAnalysis always occurs on 'feature level'.

generateFormulas is a generic function that will generate formulae using one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateFormulasDA and generateFormulasGenForm. While these functions may be called directly, generateFormulas provides a generic interface and is therefore usually preferred.

generateFormulasDA uses Bruker DataAnalysis to generate chemical formulae. This method supports scoring based on overlap between measured and theoretical isotopic patterns (both MS and MS/MS data) and the presence of 'fitting' MS/MS fragments. The method will iterate through all features (or "Compounds" in DataAnalysis terms) and call SmartFormula (and SmartFormula3D if MS/MS data is available) to generate all formulae. Parameters affecting formula calculation have to be set in advance within the DataAnalysis method for each analysis (e.g. by setDAMethod). This method requires that features were obtained with findFeaturesBruker. Unlike other algorithms, there is no prior need to generate MS peak lists.

generateFormulasGenForm uses GenForm to generate chemical formulae. When MS/MS data is available it will be used to score candidate formulae by presence of 'fitting' fragments.

generateFormulasSirius uses SIRIUS to generate chemical formulae. Similarity of measured and theoretical isotopic patterns will be used for scoring candidates. Note that SIRIUS requires availability of MS/MS data.

formulaScorings returns a data.frame with information on which scoring terms are used and what their algorithm specific name is.

Note

If any errors related to DCOM appear it might be necessary to terminate DataAnalysis (note that DataAnalysis might still be running as a background process). The ProcessCleaner application installed with DataAnalayis can be used for this.

generateFormulasGenForm always sets the exist and oei GenForm commandline options.

Formula calculation with GenForm may produce an excessive number of candidates for high m/z values (e.g. above 600) and/or many elemental combinations (set by elements). In this scenario formula calculation may need a very long time. Timeouts are used to avoid excessive computational times by terminating long running commands (set by the timeout argument).

Scorings

Each algorithm implements their own scoring system. Their names have been harmonized where possible. An overview is obtained with the formulaScorings function:

name	genform	sirius	bruker	description

combMatch	comb_match	-	-	MS and MS/MS combined match value

frag_mSigma	-	-	mSigma (SmartFormula3D)	Deviation of isotopic pattern of fragment

frag_score	-	-	Score (SmartFormula3D)	MS/MS fragment score

isoScore	MS_match	isoScore	-	How well the isotopic pattern matches

mSigma	-	-	mSigma	Deviation of the isotopic pattern

MSMSScore	MSMS_match	treeScore	-	How well MS/MS data matches

score	-	score	Score	Overall MS formula score

GenForm options

Below is a list of options (generated by running GenForm without commandline options) which can be set by the extraOpts parameter.

Formula calculation from MS and MS/MS data as described in
Meringer et al (2011) MATCH Commun Math Comput Chem 65: 259-290
Usage: GenForm ms=<filename> [msms=<filename>] [out=<filename>]
        [exist[=mv]] [m=<number>] [ion=-e|+e|-H|+H|+Na] [cha=<number>]
        [ppm=<number>] [msmv=ndp|nsse|nsae] [acc=<number>] [rej=<number>]
        [thms=<number>] [thmsms=<number>] [thcomb=<number>]
        [sort[=ppm|msmv|msmsmv|combmv]] [el=<elements> [oc]] [ff=<fuzzy formula>]
        [vsp[=<even|odd>]] [vsm2mv[=<value>]] [vsm2ap2[=<value>]] [hcf]
        [wm[=lin|sqrt|log]] [wi[=lin|sqrt|log]] [exp=<number>] [oei]
        [dbeexc=<number>] [ivsm2mv=<number>] [vsm2ap2=<number>]
        [oms[=<filename>]] [omsms[=<filename>]] [oclean[=<filename>]]
        [analyze [loss] [intens]] [dbe] [cm] [pc] [sc]
Explanation:
        ms      : filename of MS data (*.txt)
        msms    : filename of MS/MS data (*.txt)
        out     : output generated formulas
        exist   : allow only molecular formulas for that at least one
                  structural formula exists;overrides vsp, vsm2mv, vsm2ap2;
                  argument mv enables multiple valencies for P and S
        m       : experimental molecular mass (default: mass of MS basepeak)
        ion     : type of ion measured (default: M+H)
        ppm     : accuracy of measurement in parts per million (default: 5)
        msmv    : MS match value based on normalized dot product, normalized
                  sum of squared or absolute errors (default: nsae)
        acc     : allowed deviation for full acceptance of MS/MS peak in ppm
                  (default: 2)
        rej     : allowed deviation for total rejection of MS/MS peak in ppm
                  (default: 4)
        thms    : threshold for the MS match value
        thmsms  : threshold for the MS/MS match value
        thcomb  : threshold for the combined match value
        sort    : sort generated formulas according to mass deviation in ppm,
                  MS match value, MS/MS match value or combined match value
        el      : used chemical elements (default: CHBrClFINOPSSi)
        oc      : only organic compounds, i.e. with at least one C atom
        ff      : overwrites el and oc and uses fuzzy formula for limits of
                  element multiplicities
        het     : formulas must have at least one hetero atom
        vsp     : valency sum parity (even for graphical formulas)
        vsm2mv  : lower bound for valency sum - 2 * maximum valency
                  (>=0 for graphical formulas)
        vsm2ap2 : lower bound for valency sum - 2 * number of atoms + 2
                  (>=0 for graphical connected formulas)
        hcf     : apply Heuerding-Clerc filter
        wm      : m/z weighting for MS/MS match value
        wi      : intensity weighting for MS/MS match value
        exp     : exponent used, when wi is set to log
        oei     : allow odd electron ions for explaining MS/MS peaks
        dbeexc  : excess of double bond equivalent for ions
        ivsm2mv : lower bound for valency sum - 2 * maximum valency
                  for fragment ions
        ivsm2ap2: lower bound for valency sum - 2 * number of atoms + 2
                  for fragment ions
        oms     : write scaled MS peaks to output
        omsms   : write weighted MS/MS peaks to output
        oclean  : write explained MS/MS peaks to output
        analyze : write explanations for MS/MS peaks to output
        loss    : for analyzing MS/MS peaks write losses instead of fragments
        intens  : write intensities of MS/MS peaks to output
        dbe     : write double bond equivalents to output
        cm      : write calculated ion masses to output
        pc      : output match values in percent
        sc      : strip calculated isotope distributions
        noref   : hide the reference information

References

Meringer M, Reinker S, Zhang J, Muller A (2011). “MS/MS Data Improves Automated Determination of Molecular Formulas by Mass Spectrometry.” MATCH Commun. Math. Comput. Chem., 65(2), 259--290.

Duhrkop K, Bocker S (2015). “Fragmentation Trees Reloaded.” In Przytycka TM (ed.), Research in Computational Molecular Biology, 65--79. ISBN 978-3-319-16706-0.

Bocker S, Letzel MC, Liptak Z, Pervukhin A (2008). “SIRIUS: decomposing isotope patterns for metabolite identification.” Bioinformatics, 25(2), 218--224. doi: 10.1093/bioinformatics/btn603 .

See also