Formula lists class

Contains data of generated chemical formulae for given feature groups.

# S4 method for formulas
formulaTable(obj, features = FALSE)

# S4 method for formulas
algorithm(obj)

# S4 method for formulas
analyses(obj)

# S4 method for formulas
groupNames(obj)

# S4 method for formulas
length(x)

# S4 method for formulas
show(object)

# S4 method for formulas,ANY,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for formulas,ANY,ANY
[[(x, i, j)

# S4 method for formulas
$(x, name)

# S4 method for formulas
makeTable(obj, fGroups = NULL, average = FALSE,
  countElements = NULL, countFragElements = NULL, OM = FALSE,
  maxFormulas = NULL, maxFragFormulas = NULL)

# S4 method for formulas
filter(obj, minExplainedFragPeaks = NULL,
  elements = NULL, fragElements = NULL, lossElements = NULL,
  topMost = NULL, scoreLimits = NULL, OM = NULL)

# S4 method for formulas
plotSpec(obj, precursor, groupName, analysis = NULL,
  MSPeakLists, useGGPlot2 = FALSE)

# S4 method for formulas
consensus(obj, ..., formThreshold = 0)

Arguments

obj, x, object, formulas
The formulas object.

features
If TRUE returns formula data for features, otherwise for feature groups.

i, j
A numeric or character value which is used to select analyses/feature groups by their index or name, respectively (for the order/names see analyses()/groupNames()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all analyses/feature groups are selected.

For [[: should be a skalar value. If j is not specified, i selects by feature groups instead.

…
One or more formulas objects that should be used to generate the consensus.

drop
ignored.

fGroups
The featureGroups object that was used to generate this formulas object. If not NULL it is used to add feature group information (retention and m/z values).

average
If set to TRUE an 'average formula' is generated for each feature group by combining all elements from all candidates and averaging their amounts. This obviously leads to non-existing formulae, however, this data may be useful to deal with multiple candidate formulae per feature group when performing elemental characterization.

countElements, countFragElements
A character vector with elements that should be counted for each MS(/MS) formula candidate. For instance, c("C", "H") adds columns for both carbon and hydrogen amounts of each formula. Note that the neutral formula (neutral_formula column) is used to count elements of non-fragmented formulae, whereas the charged formula of fragments (frag_formula column) is used for fragments. Set to NULL to not count any elements.

OM
For makeTable: if set to TRUE several columns with information relevant for organic matter (OM) characterization will be added (e.g. elemental ratios, classification). This will also make sure that countElements contains at least C, H, N, O, P and S.

For filter: If TRUE then several filters are applied to exclude unlikely formula candidates present in organic matter (OM). See Source section for details.

maxFormulas, maxFragFormulas
Maximum amount of unique candidate formulae (or fragment formulae) per feature group. Set to NULL to ignore.

minExplainedFragPeaks
Minimum number of fragment peaks that are explained. Setting this to 1 will remove any MS only formula results. Set to NULL to ignore.

elements
Only retain candidate formulae that match a given elemental restriction. The format of elements is a character string with elements that should be present where each element is followed by a valid amount or a range thereof. If no number is specified then 1 is assumed. For instance, elements="C1-10H2-20O0-2P", specifies that 1-10, 2-20, 0-2 and 1 carbon, hydrogen, oxygen and phosphorus atoms should be present, respectively. When length(elements)>1 formulas are tested to follow at least one of the given elemental restrictions. For instance, elements=c("P", "S") specifies that either one phosphorus or sulphur atom should be present. Set to NULL to ignore this filter.

fragElements, lossElements
Specifies elemental restrictions for fragment or neutral loss formulas. Candidates are retained if at least one of the fragment/neutral loss formulae follow the given restrictions. See elements for the used format.

topMost
Only retain no more than this amount of best ranked candidates for each feature group.

scoreLimits
Filter results by their scores. Should be a named list that contains two-sized numeric vectors with the minimum/maximum value of a score (use -Inf/Inf for no limits). The names of each element should follow the values returned by formulaScorings()$name. For instance, scoreLimits=list(isoScore=c(0.5, Inf)) specifies that the isotopic match score should be at least 0.5. More details of scorings can be obtained with formulaScorings. Set to NULL to skip this filter.

precursor
The formula of the precursor (in ionic form, i.e. as detected by the MS).

groupName
The name of the feature group to which the candidate belongs.

analysis
A character specifying the analysis for which the annotated spectrum should be plotted. If NULL then annotation results for the complete feature group will be plotted.

MSPeakLists
The MSPeakLists object that was used to generate the candidate

useGGPlot2
If TRUE then ggplot2 is used for plotting, otherwise base plot used. For plotSpec, ggplot2 allows nicely repelled text for annotation. However, base plot is generally faster.

formThreshold
Fractional minimum amount (0-1) of which a formula candidate should be present within all objects. For instance, a value of 0.5 means that a particular formula should be present in at least 50% of all objects.

Value

formulaTable returns a list containing for each feature group (or feature if features=TRUE) a data.table with an overview of all generated formulae and other data such as candidate scoring and MS/MS fragments.

makeTable returns a data.table.

filter returns a filtered formulas object.

plotSpec will return a ggplot object if useGGPlot2 is TRUE.

consensus returns a formulas object that is produced by merging results from multiple formulas objects.

Details

formulas objects are obtained from formula generators.

Methods (by generic)

formulaTable: Accessor method to obtain generated formulae.

algorithm: Accessor method for the algorithm (a character string) used to generate formulae.

analyses: returns a character vector with the names of the analyses for which data is present in this object.

groupNames: returns a character vector with the names of the feature groups for which data is present in this object.

length: Obtain total number of formulae entries.

show: Show summary information for this object.

[: Subset on feature groups.

[[: Extract a formula table. If both arguments (i and j) are specified, the feature specific formula table belonging to the analysis (i)/feature group (j) is returned. Otherwise the formula table for the feature group specified by j is returned.

$: Extract a formula table for a feature group.

makeTable: Generates a table with all candidate formulae for each feature group and other information such as element counts.

filter: Performs rule based filtering on formula results.

plotSpec: Plots an annotated spectrum for a given candidate formula of a feature or feature group.

consensus: Generates a consensus of results from multiple formulas objects.

Slots

formulas,featureFormulas

Lists of all generated formulae. Use the formulaTable method for access.

algorithm

The algorithm that was used for generation of formulae. Use the algorithm method for access.

Note

filter does not modify any formula results for features (if present).

Source

Calculation of the aromaticity index (AI) and related double bond equivalents (DBE_AI) is performed as described in Koch 2015. Formula classification is performed by the rules described in Abdulla 2013. Filtering of OM related molecules is performed as described in Koch 2006 and Kujawinski 2006. (see references).

References

Koch BP, Dittmar T (2015). “From mass to structure: an aromaticity index for high-resolution mass data of natural organic matter.” Rapid Communications in Mass Spectrometry, 30(1), 250--250. doi: 10.1002/rcm.7433 , https://doi.org/10.1002/rcm.7433.

Abdulla HA, Sleighter RL, Hatcher PG (2013). “Two Dimensional Correlation Analysis of Fourier Transform Ion Cyclotron Resonance Mass Spectra of Dissolved Organic Matter: A New Graphical Analysis of Trends.” Analytical Chemistry, 85(8), 3895--3902. doi: 10.1021/ac303221j , https://doi.org/10.1021/ac303221j.

Koch BP, Dittmar T (2006). “From mass to structure: an aromaticity index for high-resolution mass data of natural organic matter.” Rapid Communications in Mass Spectrometry, 20(5), 926--932. doi: 10.1002/rcm.2386 , https://doi.org/10.1002/rcm.2386.

Kujawinski EB, Behn MD (2006). “Automated Analysis of Electrospray Ionization Fourier Transform Ion Cyclotron Resonance Mass Spectra of Natural Organic Matter.” Analytical Chemistry, 78(13), 4363--4373. doi: 10.1021/ac0600306 , https://doi.org/10.1021/ac0600306.