Formula lists class

Contains data of generated chemical formulae for given feature groups.

# S4 method for formulas
formulaTable(obj)

# S4 method for formulas
algorithm(obj)

# S4 method for formulas
length(x)

# S4 method for formulas
show(object)

# S4 method for formulas
consensus(obj, ..., fGroups, formAnaThreshold = 0.75,
  formListThreshold = 0, maxFormulas = 10, maxFragFormulas = 10,
  minIntensity = NULL, maxIntensity = NULL, minPreferredFormulas = 1,
  minPreferredIntensity = NULL, elements = c(), fragElements = c())

Arguments

obj, x, object, formulas
The formulas object.

…
Any further formulas objects on top of the object given for the obj parameter to which a consensus should be generated. Using the … parameter formulae from various formula generators can be merged.

fGroups
The featureGroups object that was used when formulae were generated.

formAnaThreshold, formListThreshold
Fractional minimum amount (0-1) to which a generated formula should be present in all the analyses (formAnaThreshold) or in all given formula lists (formListThreshold) containing the related feature group. For instance, a value of 0.5 for formAnaThreshold means that a particular formula should be present in at least 50% of all analyses containing the feature group that was used to generate the formula.

maxFormulas, maxFragFormulas
Maximum amount of candidate formulae (or fragment formulae) per feature group.

minIntensity
Results from features below this intensity will be excluded.

maxIntensity
Results from features above this intensity will be excluded.

minPreferredFormulas
Preferred minimum amount of candidate formulae per feature group. See minPreferredIntensity.

minPreferredIntensity
Minimum preferred intensity of formulae candidates per feature group. Results from features below this intensity will be removed unless the total count of formulae of the related feature group is below minPreferredFormulas. When a higher minIntensity is given this parameter has no effect.

elements
An optional character vector containing elements for which the total count present in each formula should be calculated. For instance, c("C", "H") will count all carbon and hydrogen atoms present in each formula.

fragElements
Same as elements parameter, but for MS/MS formulae.

Value

formulas returns a list containing for each analysis and each feature group a data.table with an overview of all generated formulae and other data such as candidate scoring and MS/MS fragments.

consensus returns a formulaConsensus object.

Details

formulas objects are obtained from formula generators.

Methods (by generic)

formulaTable: Accessor method to obtain generated formulae.

algorithm: Accessor method for the algorithm (a character string) used to generate formulae.

length: Obtain total number of formulae entries.

show: Show summary information for this object.

consensus: Generates a consensus and other summarizing data for formulae of feature groups that are present in multiple analyses.

Slots

formulas

Lists of all generated formulae. Use the formulaTable method for access.

algorithm

The algorithm that was used for generation of formulae. Use the algorithm method for access.