Formula lists class

Contains data of generated chemical formulae for given feature groups.

# S4 method for formulas
formulaTable(obj, features = FALSE)

# S4 method for formulas
algorithm(obj)

# S4 method for formulas
analyses(obj)

# S4 method for formulas
groupNames(obj)

# S4 method for formulas
length(x)

# S4 method for formulas
show(object)

# S4 method for formulas,ANY,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for formulas,ANY,ANY
[[(x, i, j)

# S4 method for formulas
$(x, name)

# S4 method for formulas
as.data.table(
  x,
  fGroups = NULL,
  average = FALSE,
  countElements = NULL,
  countFragElements = NULL,
  OM = FALSE,
  maxFormulas = NULL,
  maxFragFormulas = NULL,
  normalizeScores = "none",
  excludeNormScores = NULL
)

# S4 method for formulas
filter(
  obj,
  minExplainedPeaks = NULL,
  elements = NULL,
  fragElements = NULL,
  lossElements = NULL,
  topMost = NULL,
  scoreLimits = NULL,
  OM = FALSE,
  negate = FALSE
)

# S4 method for formulas
annotatedPeakList(
  obj,
  precursor,
  groupName,
  analysis = NULL,
  MSPeakLists,
  onlyAnnotated = FALSE
)

# S4 method for formulas
plotScores(
  obj,
  precursor,
  groupName,
  analysis = NULL,
  normalizeScores = "max",
  excludeNormScores = NULL,
  useGGPlot2 = FALSE
)

# S4 method for formulas
plotSpectrum(
  obj,
  precursor,
  groupName,
  analysis = NULL,
  MSPeakLists,
  title = NULL,
  useGGPlot2 = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

# S4 method for formulas
plotVenn(obj, ..., labels = NULL, vennArgs = NULL)

# S4 method for formulas
plotUpSet(
  obj,
  ...,
  labels = NULL,
  nsets = length(list(...)) + 1,
  nintersects = NA,
  upsetArgs = NULL
)

# S4 method for formulas
consensus(
  obj,
  ...,
  absMinAbundance = NULL,
  relMinAbundance = NULL,
  uniqueFrom = NULL,
  uniqueOuter = FALSE,
  rankWeights = 1,
  labels = NULL
)

Arguments

obj, x, object, formulas
The formulas object.

features
If TRUE returns formula data for features, otherwise for feature groups.

i, j
A numeric or character value which is used to select analyses/feature groups by their index or name, respectively (for the order/names see analyses()/groupNames()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all analyses/feature groups are selected.

For [[: should be a scalar value. If j is not specified, i selects by feature groups instead.

...
For plotSpectrum: Further arguments passed to plot.

Others: Any further (and unique) formulas objects.

drop
ignored.

name
The feature group name (partially matched).

fGroups
The featureGroups object that was used to generate this object. If not NULL it is used to add feature group information (retention and m/z values).

average
If set to TRUE an 'average formula' is generated for each feature group by combining all elements from all candidates and averaging their amounts. This obviously leads to non-existing formulae, however, this data may be useful to deal with multiple candidate formulae per feature group when performing elemental characterization.

countElements, countFragElements
A character vector with elements that should be counted for each MS(/MS) formula candidate. For instance, c("C", "H") adds columns for both carbon and hydrogen amounts of each formula. Note that the neutral formula (neutral_formula column) is used to count elements of non-fragmented formulae, whereas the charged formula of fragments (frag_formula column) is used for fragments. Set to NULL to not count any elements.

OM
For as.data.table: if set to TRUE several columns with information relevant for organic matter (OM) characterization will be added (e.g. elemental ratios, classification). This will also make sure that countElements contains at least C, H, N, O, P and S.

For filter: If TRUE then several filters are applied to exclude unlikely formula candidates present in organic matter (OM). See Source section for details.

maxFormulas, maxFragFormulas
Maximum amount of unique candidate formulae (or fragment formulae) per feature group. Set to NULL to ignore.

normalizeScores
A character that specifies how normalization of compound scorings occurs. Either "none" (no normalization), "max" (normalize to max value) or "minmax" (perform min-max normalization). Note that normalization of negative scores (e.g. output by SIRIUS) is always performed as min-max. Furthermore, currently normalization for compounds takes the original min/max scoring values into account when candidates were generated. Thus, for compounds scoring, normalization is not affected when candidate results were removed after they were generated (e.g. by use of filter).

excludeNormScores
A character vector specifying any compound scoring names that should not be normalized. Set to NULL to normalize all scorings. Note that whether any normalization occurs is set by the excludeNormScores argument.

For compounds: By default score and individualMoNAScore are set to mimic the behavior of the MetFrag web interface.

minExplainedPeaks
Minimum number of fragment peaks that are explained. Setting this to 1 will remove any MS only formula results. Set to NULL to ignore.

elements
Only retain candidate formulae (neutral form) that match a given elemental restriction. The format of elements is a character string with elements that should be present where each element is followed by a valid amount or a range thereof. If no number is specified then 1 is assumed. For instance, elements="C1-10H2-20O0-2P", specifies that 1-10, 2-20, 0-2 and 1 carbon, hydrogen, oxygen and phosphorus atoms should be present, respectively. When length(elements)>1 formulas are tested to follow at least one of the given elemental restrictions. For instance, elements=c("P", "S") specifies that either one phosphorus or one sulphur atom should be present. Set to NULL to ignore this filter.

fragElements, lossElements
Specifies elemental restrictions for fragment or neutral loss formulae (charged form). Candidates are retained if at least one of the fragment formulae follow (or not follow if negate=TRUE) the given restrictions. See elements for the used format.

topMost
Retain no more than this amount of best ranked (or worst ranked if negate=TRUE) candidates for each feature group.

scoreLimits
Filter results by their scores. Should be a named list that contains two-sized numeric vectors with the minimum/maximum value of a score (use -Inf/Inf for no limits). The names of each element should follow the values returned by formulaScorings()$name. For instance, scoreLimits=list(isoScore=c(0.5, Inf)) specifies that the isotopic match score should be at least 0.5. More details of scorings can be obtained with formulaScorings. Note that a result without a specified scoring is never removed. Set to NULL to skip this filter.

negate
If TRUE then filters are applied in opposite manner.

precursor
The formula of the precursor (in ionic form, i.e. as detected by the MS).

groupName
The name of the feature group to which the candidate belongs.

analysis
A character specifying the analysis for which the annotated spectrum should be plotted. If NULL then annotation results for the complete feature group will be plotted.

MSPeakLists
The MSPeakLists object that was used to generate the candidate

onlyAnnotated
Set to TRUE to filter out any peaks that could not be annotated.

useGGPlot2
If TRUE then ggplot2 is used for plotting, otherwise base plot used. For plotSpectrum, ggplot2 allows nicely repelled text for annotation. However, base plot is generally faster.

title
The title of the plot. Set to NULL for an automatically generated title.

xlim, ylim
Sets the plot size limits used by plot. Set to NULL for automatic plot sizing.

labels
A character with names to use for labelling. If NULL labels are automatically generated.

vennArgs
A list with further arguments passed to VennDiagram plotting functions. Set to NULL to ignore.

nsets, nintersects
See upset.

upsetArgs
A list with any further arguments to be passed to upset. Set to NULL to ignore.

absMinAbundance, relMinAbundance
Minimum absolute or relative (0-1) abundance across objects for a result to be kept. For instance, relMinAbundance=0.5 means that a result should be present in at least half of the number of compared objects. Set to NULL to ignore and keep all results. Limits cannot be set when uniqueFrom is not NULL.

uniqueFrom
Set this argument to only retain formulas that are unique within one or more of the objects for which the consensus is made. Selection is done by setting the value of uniqueFrom to a logical (values are recycled), numeric (select by index) or a character (as obtained with algorithm(obj)). For logical and numeric values the order corresponds to the order of the objects given for the consensus. Set to NULL to ignore.

uniqueOuter
If uniqueFrom is not NULL and if uniqueOuter=TRUE: only retain data that are also unique between objects specified in uniqueFrom.

rankWeights
A numeric vector with weights of to calulcate the mean ranking score for each candidate. The value will be re-cycled if necessary, hence, the default value of 1 means equal weights for all considered objects.

Value

formulaTable returns a list containing for each feature group (or feature if features=TRUE) a data.table with an overview of all generated formulae and other data such as candidate scoring and MS/MS fragments.

as.data.table returns a data.table.

filter returns a filtered formulas object.

plotSpectrum will return a ggplot object if useGGPlot2 is TRUE.

plotVenn (invisibly) returns a list with the following fields:

gList the gList object that was returned by the utilized VennDiagram plotting function.

areas The total area for each plotted group.

intersectionCounts The number of intersections between groups.

The order for the areas and intersectionCounts fields is the same as the parameter order from the used plotting function (see e.g. draw.pairwise.venn and draw.triple.venn).

consensus returns a formulas object that is produced by merging results from multiple formulas objects.

Details

formulas objects are obtained from formula generators.

Methods (by generic)

formulaTable: Accessor method to obtain generated formulae.

algorithm: Accessor method for the algorithm (a character string) used to generate formulae.

analyses: returns a character vector with the names of the analyses for which data is present in this object.

groupNames: returns a character vector with the names of the feature groups for which data is present in this object.

length: Obtain total number of formulae entries.

show: Show summary information for this object.

[: Subset on feature groups.

[[: Extract a formula table. If both arguments (i and j) are specified, the feature specific formula table belonging to the analysis (i)/feature group (j) is returned. Otherwise the formula table for the feature group specified by j is returned.

$: Extract a formula table for a feature group.

as.data.table: Generates a table with all candidate formulae for each feature group and other information such as element counts.

filter: Performs rule based filtering on formula results.

annotatedPeakList: Returns an MS/MS peak list annotated with data from a given candidate formula.

plotScores: Plots a barplot with scoring of a candidate compound.

plotSpectrum: Plots an annotated spectrum for a given candidate formula of a feature or feature group.

plotVenn: plots a Venn diagram (using VennDiagram) outlining unique and shared formula candidates of up to five different formulas objects.

plotUpSet: plots an UpSet diagram (using the upset function) outlining unique and shared formula candidates between different formula objects.

consensus: Generates a consensus of results from multiple objects. In order to rank the consensus candidates, first each of the candidates are scored based on their original ranking (the scores are normalized and the highest ranked candidate gets value 1). The (weighted) mean is then calculated for all scorings of each candidate to derive the final ranking (if an object lacks the candidate its score will be 0). The original rankings for each object is stored in the rank columns.

Slots

formulas,featureFormulas

Lists of all generated formulae. Use the formulaTable method for access.

scoreRanges

The original min/max values of all scorings when candidate results were generated. This is used for normalization.

Note

filter does not modify any formula results for features (if present).

Source

Calculation of the aromaticity index (AI) and related double bond equivalents (DBE_AI) is performed as described in Koch 2015. Formula classification is performed by the rules described in Abdulla 2013. Filtering of OM related molecules is performed as described in Koch 2006 and Kujawinski 2006. (see references).

Subscripting of formulae for plots generated by plotSpectrum is based on the chemistry2expression function from the ReSOLUTION package.

S4 class hierarchy

workflowStep

formulas

References

Koch BP, Dittmar T (2015). “From mass to structure: an aromaticity index for high-resolution mass data of natural organic matter.” Rapid Communications in Mass Spectrometry, 30(1), 250--250. doi: 10.1002/rcm.7433 , https://doi.org/10.1002/rcm.7433.

Abdulla HA, Sleighter RL, Hatcher PG (2013). “Two Dimensional Correlation Analysis of Fourier Transform Ion Cyclotron Resonance Mass Spectra of Dissolved Organic Matter: A New Graphical Analysis of Trends.” Analytical Chemistry, 85(8), 3895--3902. doi: 10.1021/ac303221j , https://doi.org/10.1021/ac303221j.

Koch BP, Dittmar T (2006). “From mass to structure: an aromaticity index for high-resolution mass data of natural organic matter.” Rapid Communications in Mass Spectrometry, 20(5), 926--932. doi: 10.1002/rcm.2386 , https://doi.org/10.1002/rcm.2386.

Kujawinski EB, Behn MD (2006). “Automated Analysis of Electrospray Ionization Fourier Transform Ion Cyclotron Resonance Mass Spectra of Natural Organic Matter.” Analytical Chemistry, 78(13), 4363--4373. doi: 10.1021/ac0600306 , https://doi.org/10.1021/ac0600306.

Conway JR, Lex A, Gehlenborg N (2017). “UpSetR: an R package for the visualization of intersecting sets and their properties.” Bioinformatics, 33(18), 2938--2940. doi: 10.1093/bioinformatics/btx364 , https://doi.org/10.1093/bioinformatics/btx364.

Lex A, Gehlenborg N, Strobelt H, Vuillemot R, Pfister H (2014). “UpSet: Visualization of Intersecting Sets.” IEEE Transactions on Visualization and Computer Graphics, 20(12), 1983--1992. doi: 10.1109/tvcg.2014.2346248 , https://doi.org/10.1109/tvcg.2014.2346248.