Componentization of adducts, isotopes etc. with CAMERA

Interfaces with CAMERA to generate components from known adducts, isotopes and in-source fragments.

# S4 method for featureGroups
generateComponentsCAMERA(
  fGroups,
  ionization = NULL,
  onlyIsotopes = FALSE,
  minSize = 2,
  relMinReplicates = 0.5,
  extraOpts = NULL
)

# S4 method for featureGroupsSet
generateComponentsCAMERA(fGroups, ionization = NULL, ...)

Arguments

fGroups

featureGroups object for which components should be generated.

ionization

Which ionization polarity was used to generate the data: should be "positive" or "negative". If the featureGroups object has adduct annotations, and ionization=NULL, the ionization will be detected automatically.

(sets workflow) This parameter is not supported for sets workflows, as the ionization will always be detected automatically.

onlyIsotopes

Logical value. If TRUE only isotopes are considered when generating components (faster). Corresponds to quick argument of CAMERA::annotate.

minSize

The minimum size of a component. Smaller components than this size will be removed. See note below.

relMinReplicates

Feature groups within a component are only kept when they contain data for at least this (relative) amount of replicate analyses. For instance, 0.5 means that at least half of the replicates should contain data for a particular feature group in a component. In this calculation replicates that are fully absent within a component are not taken in to account. See note below.

extraOpts

Named character vector with extra arguments directly passed to CAMERA::annotate. Set to NULL to ignore.

...

(sets workflow) Further arguments passed to the non-sets workflow method.

Value

A components (derived) object containing all generated components.

Details

This function uses CAMERA to generate components. This function is called when calling generateComponents with algorithm="camera".

The specified featureGroups object is automatically converted to an xcmsSet object using getXCMSSet.

Note

The default value for minSize and relMinReplicates results in extra filtering, hence, the final results may be different than what the algorithm normally would return.

Sets workflows

In a sets workflow the componentization is first performed for each set independently. The resulting components are then all combined in a componentsSet object. Note that the components themselves are never merged. The components are renamed to include the set name from which they were generated (e.g. "CMP1" becomes "CMP1-positive").

References

Kuhl, C., Tautenhahn, R., Boettcher, C., Larson, T. R. and Neumann, S. CAMERA: an integrated strategy for compound spectra extraction and annotation of liquid chromatography/mass spectrometry data sets. Analytical Chemistry, 84:283-289 (2012)

See also