Generate components based on intensity profiles

Generates components based on intensity profiles of feature groups.

generateComponentsIntClust(fGroups, ...)

# S4 method for class 'featureGroups'
generateComponentsIntClust(
  fGroups,
  method = "complete",
  metric = "euclidean",
  normalized = TRUE,
  average = TRUE,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)

Arguments

fGroups

featureGroups object for which components should be generated.

...

Any parameters to be passed to the selected component generation algorithm.

method

Clustering method that should be applied (passed to fastcluster::hclust).

metric

Distance metric used to calculate the distance matrix (passed to daisy).

normalized, average

Passed to as.data.table to perform normalization and averaging of data.

maxTreeHeight, deepSplit, minModuleSize

Arguments used by cutreeDynamicTree.

Value

The components are stored in objects derived from componentsIntClust.

Details

This function uses hierarchical clustering of intensity profiles to generate components. This function is called when calling generateComponents with algorithm="intclust".

Hierarchical clustering is performed on normalized (and optionally replicate averaged) intensity data and the resulting dendrogram is automatically cut with cutreeDynamicTree. The distance matrix is calculated with daisy and clustering is performed with fastcluster::hclust. The clustering of the resulting components can be further visualized and modified using the methods defined for componentsIntClust.

Sets workflows

In a sets workflow normalization of feature intensities occur per set.

References

fastcluster1

Scholle2018patRoon

See also