Generate components based on intensity profiles

Generates components based on intensity profiles of feature groups.

# S4 method for featureGroups
generateComponentsIntClust(
  fGroups,
  method = "complete",
  metric = "euclidean",
  normFunc = max,
  average = TRUE,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)

Arguments

fGroups

featureGroups object for which components should be generated.

method

Clustering method that should be applied (passed to fastcluster::hclust).

metric

Distance metric used to calculate the distance matrix (passed to daisy).

normFunc, average

Passed to as.data.table to perform normalization and averaging of data.

maxTreeHeight, deepSplit, minModuleSize

Arguments used by cutreeDynamicTree.

Value

The components are stored in objects derived from componentsIntClust.

Details

This function uses hierarchical clustering of intensity profiles to generate components. This function is called when calling generateComponents with algorithm="intclust".

Hierarchical clustering is performed on normalized (and optionally replicate averaged) intensity data and the resulting dendrogram is automatically cut with cutreeDynamicTree. The distance matrix is calculated with daisy and clustering is performed with fastcluster::hclust. The clustering of the resulting components can be further visualized and modified using the methods defined for componentsIntClust.

Sets workflows

In a sets workflow normalization of feature intensities occur per set.

References

Daniel Müllner (2013). fastcluster: Fast Hierarchical, Agglomerative Clustering Routines for R and Python. Journal of Statistical Software, 53(9), 1-18. URL https://www.jstatsoft.org/v53/i09/.

Schollee JE, Bourgin M, von Gunten U, McArdell CS, Hollender J (2018). “Non-target screening to trace ozonation transformation products in a wastewater treatment train including different post-treatments.” Water Research, 142, 267--278. doi: 10.1016/j.watres.2018.05.045 .

See also