Generate components based on MS/MS similarity

Generates components based on MS/MS similarity between feature groups.

generateComponentsSpecClust(fGroups, ...)

# S4 method for class 'featureGroups'
generateComponentsSpecClust(
  fGroups,
  MSPeakLists,
  method = "complete",
  specSimParams = getDefSpecSimParams(),
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1
)

Arguments

fGroups

featureGroups object for which components should be generated.

...

Any parameters to be passed to the selected component generation algorithm.

MSPeakLists

The MSPeakLists object for the given feature groups that should be used for MS spectral similarity calculations.

method

Clustering method that should be applied (passed to fastcluster::hclust).

specSimParams

A named list with parameters that influence the calculation of MS spectra similarities. See the spectral similarity parameters documentation for more details.

maxTreeHeight, deepSplit, minModuleSize

Arguments used by cutreeDynamicTree.

Value

The components are stored in objects derived from componentsSpecClust.

Details

This function uses hierarchical clustering of MS/MS spectra to generate components. This function is called when calling generateComponents with algorithm="specclust".

The similarities are converted to a distance matrix and used as input for hierarchical clustering, and the resulting dendrogram is automatically cut with cutreeDynamicTree. The clustering is performed with fastcluster::hclust.

Sets workflows

In a sets workflow the spectral similarities for each set are combined as is described for the spectrumSimilarity method for sets workflows.

References

fastcluster1

See also

Author

Rick Helmus <r.helmus@uva.nl> and Bas van de Velde (major contributions to spectral binning and similarity calculation).