Generation of MS Peak Lists

Functionality to convert MS and MS/MS data into MS peak lists.

# S4 method for featureGroups
generateMSPeakLists(fGroups, algorithm, ...)

Arguments

fGroups

The featureGroups object from which MS peak lists should be extracted.

algorithm

A character string describing the algorithm that should be used: "bruker", "brukerfmf", "mzr"

...

Any parameters to be passed to the selected MS peak lists generation algorithm.

Value

A MSPeakLists object.

Details

Formula calculation and identification tools rely on mass spectra that belong to features of interest. For processing, MS (and MS/MS) spectra are typically reduced to a table with a column containing measured m/z values and a column containing their intensities. These 'MS peak lists' can then be used for formula generation and compound generation.

MS and MS/MS peak lists are first generated for all features (or a subset, if the topMost argument is set). During this step multiple spectra over the feature elution profile are averaged. Subsequently, peak lists will be generated for each feature group by averaging peak lists of the features within the group. Functionality that uses peak lists will either use data from individual features or from group averaged peak lists. For instance, the former may be used by formulae calculation, while compound identification and plotting functionality typically uses group averaged peak lists.

generateMSPeakLists is a generic function that will generateMSPeakLists by one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as generateMSPeakListsMzR and generateMSPeakListsDA. While these functions may be called directly, generateMSPeakLists provides a generic interface and is therefore usually preferred.

Note

In most cases it will be necessary to centroid your MS input files. The only exception is Bruker, however, you will still need centroided mzXML/mzML files for e.g. plotting chromatograms. In this case the centroided MS files should be stored in the same directory as the raw Bruker .d files. The convertMSFiles function can be used to centroid data.

Sets workflows

With a sets workflow, the feature group averaged peak lists are made per set. This is important, because for averaging peak lists cannot be mixed, for instance, when different ionization modes were used to generate the sets. The group averaged peaklists are then simply combined and labelled in the final peak lists. However, please note that annotation and other functionality typically uses only the set specific peak lists, as this functionality cannot work with mixed peak lists.

See also