Generate peak lists with Bruker DataAnalysis from bruker features

Uses 'compounds' that were generated by the Find Molecular Features (FMF) algorithm of Bruker DataAnalysis to extract MS peak lists.

# S4 method for featureGroups
generateMSPeakListsDAFMF(
  fGroups,
  minMSIntensity = 500,
  minMSMSIntensity = 500,
  close = TRUE,
  save = close,
  avgFGroupParams = getDefAvgPListParams()
)

# S4 method for featureGroupsSet
generateMSPeakListsDAFMF(fGroups, ...)

Arguments

fGroups
The featureGroups object from which MS peak lists should be extracted.

minMSIntensity
Minimum intensity for peak lists obtained with DataAnalysis. Highly recommended to set >0 as DA tends to report many very low intensity peaks.

minMSMSIntensity
Minimum intensity for peak lists obtained with DataAnalysis. Highly recommended to set >0 as DA tends to report many very low intensity peaks.

close
If TRUE then Bruker files are closed and saved after processing with DataAnalysis, respectively. Setting close=TRUE prevents that many analyses might be opened simultaneously in DataAnalysis, which otherwise may use excessive memory or become slow. By default save is TRUE when close is TRUE, which is likely what you want as otherwise any processed data is lost.

save
If TRUE then Bruker files are closed and saved after processing with DataAnalysis, respectively. Setting close=TRUE prevents that many analyses might be opened simultaneously in DataAnalysis, which otherwise may use excessive memory or become slow. By default save is TRUE when close is TRUE, which is likely what you want as otherwise any processed data is lost.

avgFGroupParams
A list with parameters used for averaging of peak lists for feature groups. See getDefAvgPListParams for more details.

...
(sets workflow) Further arguments passed to the non-sets workflow method.

Value

A MSPeakLists object.

Details

This function uses Bruker DataAnalysis with FMF to generate MS peak lists. This function is called when calling generateMSPeakLists with algorithm="brukerfmf".

This function is similar to generateMSPeakListsDA, but uses 'compounds' that were generated by the Find Molecular Features (FMF) algorithm to extract MS peak lists. This is generally much faster , however, it only works when features were obtained with the findFeaturesBruker function. Since all MS spectra are generated in advance by Bruker DataAnalysis, only few parameters exist to customize its operation.

Note

If any errors related to DCOM appear it might be necessary to terminate DataAnalysis (note that DataAnalysis might still be running as a background process). The ProcessCleaner application installed with DataAnalayis can be used for this.

See also