Generate peak lists with mzR

Uses the mzR package to read the MS data needed for MS peak lists.

generateMSPeakListsMzR(fGroups, ...)

# S4 method for class 'featureGroups'
generateMSPeakListsMzR(
  fGroups,
  maxMSRtWindow = 5,
  precursorMzWindow = 4,
  topMost = NULL,
  avgFeatParams = getDefAvgPListParams(),
  avgFGroupParams = getDefAvgPListParams()
)

# S4 method for class 'featureGroupsSet'
generateMSPeakListsMzR(fGroups, ...)

Arguments

fGroups

The featureGroups object from which MS peak lists should be extracted.

...

Further arguments passed to the non-sets workflow method.

maxMSRtWindow

Maximum chromatographic peak window used for spectrum averaging (in seconds, +/- retention time). If NULL all spectra from a feature will be taken into account. Lower to decrease processing time.

precursorMzWindow

The m/z window (in Da) to find MS/MS spectra of a precursor. This is typically used for Data-Dependent like MS/MS data and should correspond to the isolation m/z window (i.e. +/- the precursor m/z) that was used to collect the data. For Data-Independent MS/MS experiments, where precursor ions are not isolated prior to fragmentation (e.g. bbCID, MSe, all-ion, ...) the value should be NULL.

topMost

Only extract MS peak lists from a maximum of topMost analyses with highest intensity. If NULL all analyses will be used.

avgFeatParams

Parameters used for averaging MS peak lists of individual features. Analogous to avgFGroupParams.

avgFGroupParams

A list with parameters used for averaging of peak lists for feature groups. See getDefAvgPListParams for more details.

Value

A MSPeakLists object.

Details

This function uses mzR to generate MS peak lists. This function is called when calling generateMSPeakLists with algorithm="mzr".

The MS data files should be either in .mzXML or .mzML format.

The input MS data files need to be centroided. The convertMSFiles function can be used to centroid data.

References

mzR1

mzR2

mzR3

mzR4

mzR5

See also