Miscellaneous generics

Various (S4) generic functions providing a common interface for common tasks such as plotting and filtering data. The actual functionality and function arguments are often specific for the implemented methods, for this reason, please refer to the linked method documentation for each generic.

algorithm(obj)

analysisInfo(obj)

analyses(obj)

annotatedPeakList(obj, ...)

clusterProperties(obj)

clusters(obj)

cutClusters(obj)

consensus(obj, ...)

featureTable(obj)

filter(obj, ...)

getFeatures(obj)

getMCS(obj, ...)

groupNames(obj)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotEIC(obj, ...)

plotInt(obj, ...)

plotScores(obj, ...)

plotSpec(obj, ...)

plotStructure(obj, ...)

plotVenn(obj, ...)

plotUpSet(obj, ...)

replicateGroups(obj)

treeCut(obj, k = NULL, h = NULL, ...)

treeCutDynamic(obj, maxTreeHeight = 1, deepSplit = TRUE,
  minModuleSize = 1, ...)

Arguments

obj
The object the generic should be applied to.

…
Any further method specific arguments. See method documentation for details.

addSelfLinks
If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots
Set to TRUE to enable m/z vs retention time scatter plots.

k, h
Desired numbers of clusters. See cutree.

maxTreeHeight, deepSplit, minModuleSize
Arguments used by cutreeDynamicTree.

Details

algorithm returns the algorithm that was used to generate the object.

Methods are defined for: components; compounds; formulas; optimizationResult; workflowStep.

analysisInfo returns the analysis information from an object.

Methods are defined for: featureGroups; features.

analyses returns a character vector with the analyses for which data is present in this object.

Methods are defined for: featureGroups; features; formulas; MSPeakLists.

annotatedPeakList returns an annotated MS peak list.

Methods are defined for: compounds; formulas.

clusterProperties Obtain a list with properties of the generated cluster(s).

Methods are defined for: componentsIntClust; compoundsCluster.

clusters Obtain clustering object(s).

Methods are defined for: componentsIntClust; compoundsCluster.

cutClusters Returns assigned cluster indices of a cut cluster.

Methods are defined for: componentsIntClust; compoundsCluster.

consensus combines and merges data from various algorithms to generate a consensus.

Methods are defined for: components; compounds; featureGroupsComparison; formulas.

featureTable returns feature information.

Methods are defined for: featureGroups; features.

filter provides various functionality to do post-filtering of data.

Methods are defined for: components; compounds; featureGroups; features; featuresXCMS; formulas; MSPeakLists.

getFeatures returns the object's features object.

Methods are defined for: featureGroups.

getMCS Calculcates the maximum common substructure.

Methods are defined for: compounds; compoundsCluster.

groupNames returns a character vector with the names of the feature groups for which data is present in this object.

Methods are defined for: components; compounds; compoundsCluster; featureGroups; formulas; MSPeakLists.

plotChord plots a Chord diagram to assess overlapping data.

Methods are defined for: featureGroups; featureGroupsComparison.

plotEIC plots extracted ion chromatogram(s).

Methods are defined for: components; featureGroups.

plotInt plots the intensity of all contained features.

Methods are defined for: componentsIntClust; featureGroups.

plotScores plots candidate scorings.

Methods are defined for: compounds; formulas.

plotSpec plots a (annotated) spectrum.

Methods are defined for: components; compounds; formulas; MSPeakLists.

plotStructure plots a chemical structure.

Methods are defined for: compounds; compoundsCluster.

plotVenn plots a Venn diagram to assess unique and overlapping data.

Methods are defined for: compounds; featureGroups; featureGroupsComparison; formulas.

plotUpSet plots an UpSet diagram to assess unique and overlapping data.

Methods are defined for: compounds; featureGroups; featureGroupsComparison; formulas.

replicateGroups returns a character vector with the analyses for which data is present in this object.

Methods are defined for: featureGroups; features.

treeCut Manually cut a cluster.

Methods are defined for: componentsIntClust; compoundsCluster.

treeCutDynamic Automatically cut a cluster.

Methods are defined for: componentsIntClust; compoundsCluster.

Other generics

Below are methods that are defined for existing generics (e.g. defined in base). Please see method specific documentation for more details.

[ Subsets data within an object.

Methods are defined for: components,ANY,ANY,missing; compounds,ANY,missing,missing; compoundsCluster,ANY,missing,missing; featureGroups,ANY,ANY,missing; featureGroupsComparison,ANY,missing,missing; features,ANY,missing,missing; featuresXCMS,ANY,missing,missing; formulas,ANY,missing,missing; MSPeakLists,ANY,ANY,missing.

[[ Extract data from an object.

Methods are defined for: components,ANY,ANY; compounds,ANY,missing; featureGroups,ANY,ANY; featureGroupsComparison,ANY,missing; features,ANY,missing; formulas,ANY,ANY; MSPeakLists,ANY,ANY.

$ Extract data from an object.

Methods are defined for: components; compounds; featureGroups; featureGroupsComparison; features; formulas; MSPeakLists.

as.data.table Converts an object to a table (data.table).

Methods are defined for: components; compounds; featureGroups; features; formulas; MSPeakLists; workflowStep.

as.data.frame Converts an object to a table (data.frame).

Methods are defined for: workflowStep.

length Returns the length of an object.

Methods are defined for: components; compounds; compoundsCluster; featureGroups; featureGroupsComparison; features; formulas; MSPeakLists; optimizationResult.

lengths Returns the lengths of elements within this object.

Methods are defined for: compoundsCluster; optimizationResult.

names Return names for this object.

Methods are defined for: components; featureGroups; featureGroupsComparison.

plot Generates a plot for an object.

Methods are defined for: componentsIntClust,ANY; compoundsCluster,ANY; featureGroups,ANY; featureGroupsComparison,ANY; optimizationResult,ANY.

show Prints information about this object.

Methods are defined for: adduct; components; compounds; compoundsCluster; featureGroups; features; formulas; MSPeakLists; optimizationResult; workflowStep.