Miscellaneous generics

Various (S4) generic functions providing a common interface for common tasks such as plotting and filtering data. The actual functionality and function arguments are often specific for the implemented methods, for this reason, please refer to the linked method documentation for each generic.

adducts(obj, ...)

adducts(obj, ...) <- value

algorithm(obj)

analysisInfo(obj)

analyses(obj)

annotatedPeakList(obj, ...)

annotations(obj, ...)

calculatePeakQualities(obj, weights = NULL, flatnessFactor = 0.05, ...)

clusterProperties(obj)

clusters(obj)

consensus(obj, ...)

convertToMFDB(TPs, out, ...)

convertToSuspects(TPs, ...)

cutClusters(obj)

defaultExclNormScores(obj)

export(obj, type, out, ...)

featureTable(obj, ...)

filter(obj, ...)

getFeatures(obj)

getMCS(obj, ...)

groupNames(obj)

makeSet(obj, ...)

plotChord(obj, addSelfLinks = FALSE, addRetMzPlots = TRUE, ...)

plotChroms(obj, ...)

plotGraph(obj, onlyLinked = TRUE, ...)

plotInt(obj, ...)

plotScores(obj, ...)

plotSilhouettes(obj, kSeq, ...)

plotSpectrum(obj, ...)

plotStructure(obj, ...)

plotVenn(obj, ...)

plotUpSet(obj, ...)

delete(obj, ...)

plotVolcano(obj, ...)

replicateGroups(obj)

setObjects(obj)

sets(obj)

treeCut(obj, k = NULL, h = NULL, ...)

treeCutDynamic(
  obj,
  maxTreeHeight = 1,
  deepSplit = TRUE,
  minModuleSize = 1,
  ...
)

unset(obj, set)

Arguments

obj
The object the generic should be applied to.

...
Any further method specific arguments. See method documentation for details.

value
The replacement value.

weights, flatnessFactor
See method documentation.

TPs
The transformationProducts derived object.

out
Output file.

type
The export type.

addSelfLinks
If TRUE then 'self-links' are added which represent non-shared data.

addRetMzPlots
Set to TRUE to enable m/z vs retention time scatter plots.

onlyLinked
Only plots linked objects if TRUE.

kSeq
An integer vector containing the sequence that should be used for average silhouette width calculation.

k, h
Desired numbers of clusters. See cutree.

maxTreeHeight, deepSplit, minModuleSize
Arguments used by cutreeDynamicTree.

set
The name of the set.

Details

adducts returns assigned adducts of the object.

Methods are defined for: featureGroups; featureGroupsSet.

adducts<- sets adducts of the object.

Methods are defined for: featureGroups; featureGroupsSet.

algorithm returns the algorithm that was used to generate the object.

Methods are defined for: optimizationResult; workflowStep.

analysisInfo returns the analysis information from an object.

Methods are defined for: featureGroups; features; MSPeakListsSet.

analyses returns a character vector with the analyses for which data is present in this object.

Methods are defined for: featureGroups; features; formulas; MSPeakLists.

annotatedPeakList returns an annotated MS peak list.

Methods are defined for: compounds; compoundsSet; formulas; formulasSet.

annotations returns annotations.

Methods are defined for: featureAnnotations; featureGroups; formulas.

calculatePeakQualities calculates chromatographic peak qualities and scores.

Methods are defined for: featureGroups; features.

clusterProperties Obtain a list with properties of the generated cluster(s).

Methods are defined for: componentsClust; compoundsCluster.

clusters Obtain clustering object(s).

Methods are defined for: componentsClust; compoundsCluster.

consensus combines and merges data from various algorithms to generate a consensus.

Methods are defined for: components; componentsSet; compounds; compoundsSet; featureGroupsComparison; featureGroupsComparisonSet; formulas; formulasSet.

convertToMFDB Exports the object to a local database that can be used with MetFrag.

Methods are defined for: .

convertToSuspects Converts an object to a suspect list.

Methods are defined for: .

cutClusters Returns assigned cluster indices of a cut cluster.

Methods are defined for: componentsClust; compoundsCluster.

defaultExclNormScores Returns default scorings that are excluded from normalization.

Methods are defined for: compounds; formulas.

export exports workflow data to a given format.

Methods are defined for: featureGroups; featureGroupsSet.

featureTable returns feature information.

Methods are defined for: featureGroups; featureGroupsSet; features.

filter provides various functionality to do post-filtering of data.

Methods are defined for: components; componentsSet; componentsTPs; compounds; compoundsSet; featureAnnotations; featureGroups; featureGroupsScreening; featureGroupsScreeningSet; featureGroupsSet; features; featuresSet; formulasSet; MSPeakLists; MSPeakListsSet; transformationProductsBT.

getFeatures returns the object's features object.

Methods are defined for: featureGroups.

getMCS Calculcates the maximum common substructure.

Methods are defined for: compounds; compoundsCluster.

groupNames returns a character vector with the names of the feature groups for which data is present in this object.

Methods are defined for: components; compoundsCluster; featureAnnotations; featureGroups; MSPeakLists.

makeSet creates a set from given workflow objects.

Methods are defined for: featureGroups; featureGroupsSet; features; featuresSet.

plotChord plots a Chord diagram to assess overlapping data.

Methods are defined for: featureGroups; featureGroupsComparison.

plotChroms plots extracted ion chromatogram(s).

Methods are defined for: components; featureGroups.

plotGraph Plots an interactive network graph.

Methods are defined for: componentsNT; componentsNTSet; componentsTPs.

plotInt plots the intensity of all contained features.

Methods are defined for: componentsIntClust; featureGroups; featureGroupsSet.

plotScores plots candidate scorings.

Methods are defined for: compounds; formulas.

plotSilhouettes plots silhouette widths to evaluate the desired cluster size.

Methods are defined for: componentsClust; compoundsCluster.

plotSpectrum plots a (annotated) spectrum.

Methods are defined for: components; compounds; compoundsSet; formulas; formulasSet; MSPeakLists; MSPeakListsSet.

plotStructure plots a chemical structure.

Methods are defined for: compounds; compoundsCluster.

plotVenn plots a Venn diagram to assess unique and overlapping data.

Methods are defined for: featureAnnotations; featureGroups; featureGroupsComparison; featureGroupsSet.

plotUpSet plots an UpSet diagram to assess unique and overlapping data.

Methods are defined for: featureAnnotations; featureGroups; featureGroupsComparison.

delete Deletes results.

Methods are defined for: components; componentsClust; componentsSet; compoundsSet; featureAnnotations; featureGroups; featureGroupsKPIC2; featureGroupsScreening; featureGroupsScreeningSet; featureGroupsSet; featureGroupsXCMS; featureGroupsXCMS3; features; featuresKPIC2; featuresXCMS; featuresXCMS3; formulas; formulasSet.

plotVolcano plots a volcano plot.

Methods are defined for: featureGroups.

replicateGroups returns a character vector with the analyses for which data is present in this object.

Methods are defined for: featureGroups; features.

setObjects returns the set objects of this object. See the documentation of workflowStepSet.

Methods are defined for: workflowStepSet.

sets returns the names of the sets inside this object. See the documentation for sets workflows.

Methods are defined for: featureGroupsSet; featuresSet; workflowStepSet.

treeCut Manually cut a cluster.

Methods are defined for: componentsClust; compoundsCluster.

treeCutDynamic Automatically cut a cluster.

Methods are defined for: componentsClust; compoundsCluster.

unset Converts this object to a regular non-set object. See the documentation for sets workflows.

Methods are defined for: componentsNTSet; componentsSet; compoundsConsensusSet; compoundsSet; featureGroupsScreeningSet; featureGroupsSet; featuresSet; formulasConsensusSet; formulasSet; MSPeakListsSet.

Other generics

Below are methods that are defined for existing generics (e.g. defined in base). Please see method specific documentation for more details.

[ Subsets data within an object.

Methods are defined for: components,ANY,ANY,missing; componentsSet,ANY,ANY,missing; compoundsCluster,ANY,missing,missing; compoundsSet,ANY,missing,missing; featureAnnotations,ANY,missing,missing; featureGroups,ANY,ANY,missing; featureGroupsComparison,ANY,missing,missing; featureGroupsScreening,ANY,ANY,missing; featureGroupsScreeningSet,ANY,ANY,missing; featureGroupsSet,ANY,ANY,missing; features,ANY,missing,missing; featuresSet,ANY,missing,missing; formulasSet,ANY,missing,missing; MSPeakLists,ANY,ANY,missing; MSPeakListsSet,ANY,ANY,missing; transformationProducts,ANY,missing,missing.

[[ Extract data from an object.

Methods are defined for: components,ANY,ANY; featureAnnotations,ANY,missing; featureGroups,ANY,ANY; featureGroupsComparison,ANY,missing; features,ANY,missing; formulas,ANY,ANY; MSPeakLists,ANY,ANY; transformationProducts,ANY,missing.

$ Extract data from an object.

Methods are defined for: components; featureAnnotations; featureGroups; featureGroupsComparison; features; MSPeakLists; transformationProducts.

as.data.table Converts an object to a table (data.table).

Methods are defined for: components; componentsTPs; featureAnnotations; featureGroups; featureGroupsScreening; featureGroupsScreeningSet; featureGroupsSet; features; featuresSet; formulas; MSPeakLists; MSPeakListsSet; transformationProducts; workflowStep.

as.data.frame Converts an object to a table (data.frame).

Methods are defined for: workflowStep.

length Returns the length of an object.

Methods are defined for: components; compoundsCluster; featureAnnotations; featureGroups; featureGroupsComparison; features; MSPeakLists; optimizationResult; transformationProducts.

lengths Returns the lengths of elements within this object.

Methods are defined for: compoundsCluster; optimizationResult.

names Return names for this object.

Methods are defined for: components; featureGroups; featureGroupsComparison; transformationProducts.

plot Generates a plot for an object.

Methods are defined for: componentsClust,missing; compoundsCluster,missing; featureGroups,missing; featureGroupsComparison,missing; optimizationResult,missing.

show Prints information about this object.

Methods are defined for: adduct; components; componentsFeatures; componentsSet; compounds; compoundsCluster; compoundsSet; featureGroups; featureGroupsScreening; featureGroupsScreeningSet; featureGroupsSet; features; featuresSet; formulas; formulasSet; MSPeakLists; MSPeakListsSet; optimizationResult; transformationProducts; workflowStep; workflowStepSet.