Group features using OpenMS

Group and align features with OpenMS tools

# S4 method for features
groupFeaturesOpenMS(
  feat,
  rtalign = TRUE,
  QT = FALSE,
  maxAlignRT = 30,
  maxAlignMZ = 0.005,
  maxGroupRT = 12,
  maxGroupMZ = 0.005,
  extraOptsRT = NULL,
  extraOptsGroup = NULL,
  verbose = TRUE
)

Arguments

feat
The features object with the features to be grouped.

rtalign
Set to TRUE to enable retention time alignment.

QT
If enabled, use FeatureLinkerUnlabeledQT instead of FeatureLinkerUnlabeled for feature grouping.

maxAlignRT, maxAlignMZ
Used for retention alignment. Maximum retention time or m/z difference (seconds/Dalton) for feature pairing. Sets -algorithm:pairfinder:distance_RT:max_difference and -algorithm:pairfinder:distance_MZ:max_difference otpions, respectively.

maxGroupRT, maxGroupMZ
as maxAlignRT and maxAlignMZ, but for grouping of features. Sets -algorithm:distance_RT:max_difference and -algorithm:distance_MZ:max_difference options, respectively.

extraOptsRT, extraOptsGroup
Named list containing extra options that will be passed to MapAlignerPoseClustering or FeatureLinkerUnlabeledQT/FeatureLinkerUnlabeled, respectively. Any options specified here will override any of the above. Example: extraOptsGroup=list("-algorithm:distance_RT:max_difference"=12) (corresponds to setting maxGroupRT=12). Set to NULL to ignore.

verbose
if FALSE then no text output will be shown.

Value

An object of a class which is derived from featureGroups.

The featuresSet method (for sets workflows) returns a featureGroupsSet object.

Details

This function uses OpenMS to group features. This function is called when calling groupFeatures with algorithm="openms".

Retention times may be aligned by the MapAlignerPoseClustering TOPP tool. Grouping is achieved by either the FeatureLinkerUnlabeled or FeatureLinkerUnlabeledQT TOPP tools.

References

Rost HL, Sachsenberg T, Aiche S, Bielow C, Weisser H, Aicheler F, Andreotti S, Ehrlich H, Gutenbrunner P, Kenar E, Liang X, Nahnsen S, Nilse L, Pfeuffer J, Rosenberger G, Rurik M, Schmitt U, Veit J, Walzer M, Wojnar D, Wolski WE, Schilling O, Choudhary JS, Malmstrom L, Aebersold R, Reinert K, Kohlbacher O (2016). “OpenMS: a flexible open-source software platform for mass spectrometry data analysis.” Nature Methods, 13(9), 741--748. doi: 10.1038/nmeth.3959 .

pugixml (via Rcpp) is used to process OpenMS XML output.

Dirk Eddelbuettel and Romain Francois (2011). Rcpp: Seamless R and C++ Integration. Journal of Statistical Software, 40(8), 1-18, .

Eddelbuettel, Dirk (2013) Seamless R and C++ Integration with Rcpp, Springer, New York, ISBN 978-1-4614-6867-7, .

Dirk Eddelbuettel and James Joseph Balamuta (2018). Extending R with C++: A Brief Introduction to Rcpp. The American Statistician. 72(1), .

See also