Import features

Generic function to import features produced by other software.

importFeatures(analysisInfo, type, ...)

Arguments

analysisInfo

A data.frame with Analysis information.

type

What type of data should be imported: "xcms", "xcms3", "kpic2" or "envimass".

...

Further arguments passed to the selected import algorithm function.

Value

An object of a class which is derived from features.

Details

importFeatures is a generic function that will import features by one of the supported algorithms. The actual functionality is provided by algorithm specific functions such as importFeaturesXCMS3 and importFeaturesKPIC2. While these functions may be called directly, importFeatures provides a generic interface and is therefore usually preferred.

See also