Class containing optimization results.

Objects from this class contain optimization results resulting from design of experiment (DoE).

# S4 method for optimizationResult
algorithm(obj)

# S4 method for optimizationResult
length(x)

# S4 method for optimizationResult
lengths(x, use.names = FALSE)

# S4 method for optimizationResult
show(object)

# S4 method for optimizationResult,ANY
plot(
  x,
  paramSet,
  DoEIteration,
  paramsToPlot = NULL,
  maxCols = NULL,
  type = "contour",
  image = TRUE,
  contours = "colors",
  ...
)

# S4 method for optimizationResult
optimizedParameters(object, paramSet = NULL, DoEIteration = NULL)

# S4 method for optimizationResult
optimizedObject(object, paramSet = NULL)

# S4 method for optimizationResult
scores(object, paramSet = NULL, DoEIteration = NULL)

# S4 method for optimizationResult
experimentInfo(object, paramSet, DoEIteration)

Arguments

obj, x, object
An optimizationResult object.

use.names
Ignored.

paramSet
Numeric index of the parameter set (i.e. the first parameter set gets index 1). For some methods optional: if NULL the best will be selected.

DoEIteration
Numeric index specifying the DoE iteration within the specified paramSet. For some methods optional: if NULL the best will be selected.

paramsToPlot
Which parameters relations should be plot. If NULL all will be plot. Alternatively, a list containing one or more character vectors specifying each two parameters that should be plotted. Finally, if only one pair should be plotted, can be a character vector specifying both parameters.

maxCols
Multiple parameter pairs are plotted in a grid. The maximum number of columns can be set with this argument. Set to NULL for no limit.

type
The type of plots to be generated: "contour", "image" or "persp". The equally named functions will be called for plotting.

image
Passed to contour (if type="contour").

contours
Passed to persp (if type="persp").

...
Further arguments passed to contour, image or persp (depending on type).

Details

Objects from this class are returned by optimizeFeatureFinding and optimizeFeatureGrouping.

Methods (by generic)

algorithm: Returns the algorithm that was used for finding features.

length: Obtain total number of experimental design iterations performed.

lengths: Obtain number of experimental design iterations performed for each parameter set.

show: Shows summary information for this object.

plot: Generates response plots for all or a selected set of parameters.

optimizedParameters: Returns parameter set yielding optimal results. The paramSet and DoEIteration arguments can be NULL.

optimizedObject: Returns the object (i.e. a features or featureGroups object) that was generated with optimized parameters. The paramSet argument can be NULL.

scores: Returns optimization scores. The paramSet and DoEIteration arguments can be NULL.

experimentInfo: Returns a list with optimization information from an DoE iteration.

Slots

algorithm

A character specifying the algorithm that was optimized.

paramSets

A list with detailed results from each parameter set that was tested.

bestParamSet

Numeric index of the parameter set yielding the best response.

Examples

if (FALSE) {
# ftOpt is an optimization object.

# plot contour of all parameter pairs from the first parameter set/iteration.
plot(ftOpt, paramSet = 1, DoEIteration = 1)

# as above, but only plot two parameter pairs
plot(ftOpt, paramSet = 1, DoEIteration = 1,
     paramsToPlot = list(c("mzPPM", "chromFWHM"), c("chromFWHM", "chromSNR")))

# plot 3d perspective plots
plot(ftOpt, paramSet = 1, DoEIteration = 1, type = "persp")
}