Defunct functions.

These functions do not work anymore and may be updated in the future. future.

compoundViewer(fGroups, MSPeakLists, compounds)

# S4 method for class 'featureGroups,MSPeakLists,compounds'
compoundViewer(fGroups, MSPeakLists, compounds)

groupFeaturesScreening(...)

checkChromatograms(...)

Arguments

MSPeakLists

A MSPeakLists object.

compounds

A compounds object.

...

Ignore.

Details

The compoundViewer method is used to view compound identification results. It will display available candidate information such as scorings and identifiers, MS/MS spectra with explained peaks and chemical structures.

groupFeaturesScreening was replaced by screenSuspects. Please refer to this function and the handbook for the updated interface.

checkChromatograms was replaced by checkFeatures. Please refer to this function and the handbook for the updated interface.