Deprecated and renamed functions.

Please do not use these functions anymore since they may be removed in the future.

reportMD(...)

exportDAFiles(
  anaInfo,
  format = "mzML",
  exportLine = TRUE,
  outPath = anaInfo$path,
  overWrite = FALSE
)

plotEIC(obj, ...)

groups(object, ...)

plotSpec(obj, ...)

formulaTable(...)

compoundTable(...)

Arguments

...
Passed to successor function.

format
The output format of exported files. Should be either "mzXML", "mzML" or "mzData".

exportLine
Export line spectra (TRUE) or profile spectra (FALSE). Usually line spectra are preferred, since profile spectra use signficantly more disk space and increase required memory during processing.

outPath
Character vector of output paths for exported analyses. Will be recycled if necessary.

overWrite
If TRUE existing files will be overwritten.

Details

reportMD performs HTML reporting, please use reportHTML instead.

exportDAFiles will export a set of analyses either in .mzXML or .mzML formats.

Please use plotChroms instead.

Please use groupTable instead.

Please use plotSpectrum instead.

Please use annotations instead.

Please use annotations instead.