Workflow solutions for mass-spectrometry based non-target analysis.

Provides an easy-to-use interface to a mass spectrometry based non-target analysis workflow. Various (open-source) tools are combined which provide algorithms for extraction and grouping of features, extraction of MS and MS/MS data, automatic formula and compound annotation and grouping related features to components. In addition, various tools are provided for e.g. data preparation and cleanup, plotting results and automatic reporting.

Package options

The following package options (see options) can be set:

patRoon.cache.mode: A character setting the current caching mode: "save" and "load" will only save/load results to/from the cache, "both" (default) will do both and "none" to completely disable caching. This option can be changed anytime, which might be useful, for instance, to temporarily disable cached results before running a function.

patRoon.cache.fileName: a character specifying the name of the cache file (default is cache.sqlite).

patRoon.maxProcAmount: The maximum number of processes that should be initiated in parallel. A good starting point is the number of physical cores, which is the default as detected by detectCores.

patRoon.path.pwiz: The path in which the ProteoWizard binaries are installed. If unset an attempt is made to find this directory from the Windows registry and PATH environment variable.

patRoon.path.GenForm: The path to the GenForm executable. If not set (the default) the internal GenForm binary is used. Only set if you want to override the executable.

patRoon.path.MetFragCL: The complete file path to the MetFrag CL jar file that must be set when using generateCompoundsMetfrag. Example: "C:/MetFrag2.4.2-CL.jar".

patRoon.path.MetFragCompTox: The complete file path to the CompTox database csv file. See generateCompounds for more details.

patRoon.path.MetFragPubChemLite: The complete file path to the PubChem database csv file. See generateCompounds for more details.

patRoon.path.SIRIUS: The directory in which SIRIUS is installed. Unless the binaries can be located via the PATH environment variable, this must be set when using generateFormulasSIRIUS or generateCompoundsSIRIUS. Example: "C:/sirius-win64-3.5.1".

patRoon.path.OpenMS: The path in which the OpenMS binaries are installed. Usually the location is added to the PATH environment variable when OpenMS is installed, in which case this option can be left empty.

patRoon.path.pngquant: The path of the pngquant binary that is used when optimizing .png plots generated by reportHTML (with optimizePng set to TRUE). If the binary can be located through the PATH environment variable this option can remain empty. Note that some of the functionality of reportHTML only locates the binary through the PATH environment variable, hence, it is recommended to set up PATH instead.

patRoon.path.obabel: The path in which the OpenBabel binaries are installed. Usually the location is added to the PATH environment variable when OpenBabel is installed, in which case this option can be left empty.

See also