Workflow solutions for mass-spectrometry based non-target analysis.

Provides an easy-to-use interface to a mass spectrometry based non-target analysis workflow. Various (open-source) tools are combined which provide algorithms for extraction and grouping of features, extraction of MS and MS/MS data, automatic formula and compound annotation and grouping related features to components. In addition, various tools are provided for e.g. data preparation and cleanup, plotting results and automatic reporting.

Package options

The following package options (see options) can be set:

patRoon.checkCentroided: If set to TRUE (the default) then the analyses files are verified to be centroided before loading any MS data. While these checks are optimized and cached, it may be useful to set this option to FALSE when processing very large numbers of analyses.

patRoon.cache.mode: A character setting the current caching mode: "save" and "load" will only save/load results to/from the cache, "both" (default) will do both and "none" to completely disable caching. This option can be changed anytime, which might be useful, for instance, to temporarily disable cached results before running a function.

patRoon.cache.fileName: a character specifying the name of the cache file (default is cache.sqlite).

patRoon.MP.maxProcs: The maximum number of processes that should be initiated in parallel. A good starting point is the number of physical cores, which is the default as detected by detectCores. This option is only used when patRoon.MP.method="classic".

patRoon.MP.method: Either "classic" or "future". The former is the default and uses processx to execute multiple commands in parallel. When "future" the future.apply package is used for parallelization, which is especially useful for e.g. cluster computing.

patRoon.MP.futureSched: Sets the future.scheduling function argument for future_lapply. Only used if patRoon.MP.method="future".

patRoon.MP.logPath: The path used for logging of output from commands executed by multiprocess. Set to FALSE to disable logging.

patRoon.path.pwiz: The path in which the ProteoWizard binaries are installed. If unset an attempt is made to find this directory from the Windows registry and PATH environment variable.

patRoon.path.GenForm: The path to the GenForm executable. If not set (the default) the internal GenForm binary is used. Only set if you want to override the executable.

patRoon.path.MetFragCL: The complete file path to the MetFrag CL jar to be used by generateCompoundsMetFrag. Example: "C:/MetFrag2.4.2-CL.jar".

patRoon.path.MetFragCompTox: The complete file path to the CompTox database csv file. See generateCompounds for more details.

patRoon.path.MetFragPubChemLite: The complete file path to the PubChemLite database csv file. See generateCompounds for more details.

patRoon.path.SIRIUS: The directory in which the SIRIUS binaries are installed. Used by all functions that interface with SIRIUS, such as generateFormulasSIRIUS and generateCompoundsSIRIUS. Example: "C:/sirius-win64-3.5.1". Note that the location of the binaries differs for each operating system. are installed in different subdirectories for each location inside this differs for each operating system

patRoon.path.OpenMS: The path in which the OpenMS binaries are installed.

patRoon.path.pngquant: The path of the pngquant binary that is used when optimizing .png plots generated by reportHTML (with optimizePng set to TRUE). If the binary can be located through the PATH environment variable this option can remain empty. Note that some of the functionality of reportHTML only locates the binary through the PATH environment variable, hence, it is recommended to set up PATH instead.

patRoon.path.obabel: The path in which the OpenBabel binaries are installed.

patRoon.path.BiotransFormer The full file path to the biotransformer .jar command line utility. This needs to be set when generateTPsBioTransformer is used. For more details see https://bitbucket.org/djoumbou/biotransformer/src/master.

Most external dependencies are provided by patRoonExt or otherwise found in the system environment PATH variable. However, the patRoon.path.* options should be set if this fails or you want to override the location. The verifyDependencies function can be used to assess if dependencies are found.

See also

Author

Maintainer: Rick Helmus r.helmus@uva.nl (ORCID)

Other contributors:

Olaf Brock (ORCID) [contributor]

Vittorio Albergamo (ORCID) [contributor]

Andrea Brunner (ORCID) [contributor]

Emma Schymanski (ORCID) [contributor]

Bas van de Velde (ORCID) [contributor]

Leon Saal (ORCID) [contributor]