Report feature group data

Functionality to report data produced by most workflow steps such as features, feature groups, calculated chemical formulae and tentatively identified compounds.

# S4 method for featureGroups
reportCSV(
  fGroups,
  path = "report",
  reportFGroupsAsRows = TRUE,
  reportFGroupsAnalysisInfo = TRUE,
  reportFGroupsRetMz = TRUE,
  reportFeatures = FALSE,
  formulas = NULL,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore"),
  compsCluster = NULL,
  components = NULL,
  retMin = TRUE,
  clearPath = FALSE
)

# S4 method for featureGroups
reportPDF(
  fGroups,
  path = "report",
  reportFGroups = TRUE,
  formulas = NULL,
  formulasTopMost = 5,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  reportFormulaSpectra = TRUE,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore"),
  compoundsOnlyUsedScorings = TRUE,
  compoundsTopMost = 5,
  compsCluster = NULL,
  components = NULL,
  MSPeakLists = NULL,
  retMin = TRUE,
  EICGrid = c(2, 1),
  EICRtWindow = 20,
  EICMzWindow = 0.005,
  EICTopMost = NULL,
  EICOnlyPresent = TRUE,
  clearPath = FALSE
)

# S4 method for featureGroups
reportHTML(
  fGroups,
  path = "report",
  reportPlots = c("chord", "venn", "upset", "eics", "formulas"),
  formulas = NULL,
  formulasTopMost = 5,
  formulasNormalizeScores = "max",
  formulasExclNormScores = NULL,
  compounds = NULL,
  compoundsNormalizeScores = "max",
  compoundsExclNormScores = c("score", "individualMoNAScore"),
  compoundsOnlyUsedScorings = TRUE,
  compoundsTopMost = 5,
  compsCluster = NULL,
  includeMFWebLinks = "compounds",
  components = NULL,
  interactiveHeat = FALSE,
  MSPeakLists = NULL,
  retMin = TRUE,
  EICRtWindow = 20,
  EICMzWindow = 0.005,
  EICTopMost = NULL,
  EICOnlyPresent = TRUE,
  selfContained = TRUE,
  optimizePng = FALSE,
  maxProcAmount = getOption("patRoon.maxProcAmount"),
  clearPath = FALSE,
  openReport = TRUE,
  noDate = FALSE
)

Arguments

fGroups
The featureGroups object that should be used for reporting data.

path
The destination file path for files generated during reporting. Will be generated if needed.

reportFGroupsAsRows
Report feature groups as rows (instead of columns) within the resulting .csv file.

reportFGroupsAnalysisInfo
Include analyses information (reference and replicate groups) in the reported feature groups table .csv file.

reportFGroupsRetMz
Include feature group information (retention time and m/z) within the reported feature groups table .csv file.

reportFeatures
If set to TRUE then for each analysis a .csv file will be generated with information about its detected features.

formulas, compounds, compsCluster, components
Further objects (formulas, compounds, compoundsCluster, components) that should be reported. Specify NULL to skip reporting a particular object.

compoundsNormalizeScores, formulasNormalizeScores
A character that specifies how normalization of compound scorings occurs. Either "none" (no normalization), "max" (normalize to max value) or "minmax" (perform min-max normalization). Note that normalization of negative scores (e.g. output by SIRIUS) is always performed as min-max. Furthermore, currently normalization takes the original scoring min/max values into account when candidates were generated. Thus, normalization is not affected when candidate results were removed after they were generated (e.g. by use of filter).

compoundsExclNormScores, formulasExclNormScores
A character vector specifying any compound scoring names that should not be normalized. Set to NULL to normalize all scorings. Note that whether any normalization occurs is set by the compoundsExclNormScores,formulasExclNormScores argument.

For compounds: By default score and individualMoNAScore are set to mimic the behaviour of the MetFrag web interface.

retMin
If TRUE then report retention times in minutes (otehrwise seconds).

clearPath
If TRUE then the destination path will be (recursively) removed prior to reporting.

reportFGroups
If TRUE then feature group data will be reported.

formulasTopMost, compoundsTopMost
Only this amount of top ranked candidate formulae/compounds are reported. Lower values may significantly speed up reporting. Set to NULL to ignore.

reportFormulaSpectra
If TRUE then explained MS/MS spectra (if available) for candidate formulae will be reported. Specifying formulas and setting this argument to FALSE still allows further annotation of compound MS/MS spectra.

compoundsOnlyUsedScorings
If TRUE then only scorings are plotted that actually have been used to rank data (see the scoreTypes argument to generateCompoundsMetfrag for more details).

MSPeakLists
A MSPeakLists object that is mandatory when spectra for formulae and/or compounds will be reported.

EICGrid
An integer vector in the form c(columns, rows) that is used to determine the plotting grid when reporting EICs in PDF files.

EICRtWindow, EICMzWindow, EICTopMost, EICOnlyPresent
Plotting parameters passed to plotEIC (i.e. rtWindow, mzWindow, topMost and onlyPresent arguments).

reportPlots
A character vector specifying what should be plotted. Valid options are: "chord", "venn", "upset" (plot a chord, Venn and UpSet diagram, respectively), "eics" (plot EICs for individual feature groups) and "formulas" (plot annotated formula spectra). Set to "none" to plot none of these.

includeMFWebLinks
A character specifying to which feature groups a web link should be added in the annotation page to MetFragWeb. Options are: "compounds" (only to those with compounds results), "MSMS" (only to those with MSMS peak lists) or "none".

interactiveHeat
If TRUE an interactive heatmap HTML widget will be generated to display hierarchical clustering results. Set to FALSE for a 'regular' static plot.

selfContained
If TRUE the output will be a standalone HTML file which contains all graphics and script dependencies. When FALSE, the latter will be placed in an additional directory (report_files) which should remain present when viewing the output file. Especially on Windows, a non-self contained output might be desirable when reporting large amounts of data to prevent pandoc from running out of memory.

optimizePng
If TRUE then pngquant is used to reduce the size of generated graphics. A signficant reduction in disk space usage may be seen, however, at the cost additional processing time.

maxProcAmount
Maximum amount of pngquant commands to run in parallel. Higher numbers will decrease processing time, with an optimum usually close to the amount of CPU cores.

openReport
If set to TRUE then the output report file will be opened with the system browser.

noDate
If TRUE then the current date is not added to the report. This is mainly used for testing and its main purpose is to guarentees equal report files when `reportHTML()` is called multiple times with equal arguments.

Details

These functions are usually called at the very end of the workflow. It is used to report various data on features and feature groups. In addition, these functions may be used for reporting formulae and/or compounds that were generated for the specified feature groups. Data can be reported in tabular form (i.e. .csv files) by reportCSV or graphically by reportPDF and reportHTML. The latter functions will plot for instance chromatograms and annotated mass spectra, which are useful to get a graphical overview of results.

All functions have a wide variety of arguments that influence the reporting process. Nevertheless, most parameters are optional and only required to be given for fine tuning. In addition, only those objects (e.g. formulae, compounds, clustering) that are desired to be reported need to be specified.

reportCSV generates tabular data (i.e. .csv files) for given data to be reported. This may also be useful to allow import by other tools for post processing.

reportPDF will report graphical data (e.g. chromatograms and mass spectra) within PDF files. Compared to reportHTML this function may be faster and yield smaller report files, however, its functionality is a bit more basic and generated data is more 'scattered' around.

reportHTML will report graphical data (e.g. chromatograms and mass spectra) and summary information in an easy browsable HTML file using rmarkdown, flexdashboard and knitr.

Note

Any formulae and compounds for feature groups which are not present within fGroups (i.e. because it has been subset afterwards) will not be reported.

References

Creating MetFrag landing page URLs based on code from MetFamily R package.

Yihui Xie (2015) Dynamic Documents with R and knitr. 2nd edition. Chapman and Hall/CRC. ISBN 978-1498716963

Yihui Xie (2014) knitr: A Comprehensive Tool for Reproducible Research in R. In Victoria Stodden, Friedrich Leisch and Roger D. Peng, editors, Implementing Reproducible Computational Research. Chapman and Hall/CRC. ISBN 978-1466561595