Target and suspect screening

Utilities to screen for analytes with known or suspected identity.

# S4 method for featureGroups
screenSuspects(
  fGroups,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE,
  onlyHits = FALSE
)

numericIDLevel(level)

genIDLevelRulesFile(out, inLevels = NULL, exLevels = NULL)

importFeatureGroupsBrukerTASQ(path, analysisInfo, clusterRTWindow = 12)

Arguments

fGroups
The featureGroups object that should be screened.

suspects
A data.frame with suspect information. See the Suspect list format section below.

rtWindow, mzWindow
The retention time window (in seconds) and m/z window that will be used for matching a suspect (+/- feature data).

adduct
An adduct object (or something that can be converted to it with as.adduct). Examples: "[M-H]-", "[M+Na]+". May be NULL, see Suspect list format section below.

skipInvalid
If set to TRUE then suspects with invalid data (e.g. missing names or other missing data) will be ignored with a warning. Similarly, any suspects for which mass calculation failed (when no mz column is present in the suspect list), for instance, due to invalid SMILES, will be ignored with a warning.

onlyHits
If TRUE then all feature groups not matched by any of the suspects will be removed.

level
The identification level to be converted.

out
The file path to the target file.

inLevels, exLevels
A regular expression for the identification levels to include or exclude, respectively. For instance, exLevels="4|5" would exclude level 4 and 5 from the output file. Set to NULL to ignore.

path
The file path to an Excel export of the Global results table from TASQ, converted to .csv format.

analysisInfo
A table with analysis information.

clusterRTWindow
This retention time window (in seconds) is used to group hits across analyses together. See also the details section.

Value

screenSuspects returns a featureGroupsScreening object, which is a copy of the input fGroups object amended with additional screening information.

importFeatureGroupsBrukerTASQ returns a new featureGroups object containing converted screening results from Bruker TASQ.

Details

Besides 'full non-target analysis', where compounds may be identified with little to no prior knowledge, a common strategy is to screen for compounds with known or suspected identity. This may be a generally favorable approach if possible, as it can significantly reduce the load on data interpretation.

screenSuspects is used to perform suspect screening. The input featureGroups object will be screened for suspects by m/z values and optionally retention times. Afterwards, any feature groups not matched may be kept or removed, depending whether a full non-target analysis is desired.

numericIDLevel Extracts the numeric part of a given identification level (e.g. "3a" becomes 3).

genIDLevelRulesFile Generates a template YAML file that is used to configure the rules for automatic estimation of identification levels. This file can then be used as input for annotateSuspects.

importFeatureGroupsBrukerTASQ will convert screening results from Bruker TASQ to a featureGroups object. The feature groups across analyses are formed based on the name of suspects and their closeness in retention time. The latter is necessary because TASQ does not necessarily perform checks on retention times and may therefore assign a suspect to peaks with different retention times across analyses (or within a single analysis). Hence, suspects with equal names are hierarchically clustered on their retention times (using fastcluster) to form the feature groups. The cut-off value for this is specified by the clusterRTWindow argument. The input for this function is obtained by generating an Excel export of the 'global' results and subsequently converting the file to .csv format.

Note

Both screenSuspects and importFeatureGroupsBrukerTASQ may use the suspect names to base file names used for reporting, logging etc. Therefore, it is important that these are file-compatible names.

For screenSuspects in some cases you may need to install OpenBabel (e.g. when only InChI data is available for mass calculation).

importFeatureGroupsBrukerTASQ will use estimated min/max values for retention times and dummy min/max m/z values for conversion to features, since this information is not (readily) available. Hence, when plotting, for instance, extracted ion chromatograms (with plotChroms) the integrated chromatographic peak range shown is incorrect.

Suspect list format

the suspects argument for screenSuspects should be a data.frame with the following mandatory and optional columns:

name The suspect name. Must be file-compatible. (mandatory)

rt The retention time (in seconds) for the suspect. If specified the suspect will only be matched if its retention matches the experimental value (tolerance defined by the rtWindow argument). (optional)

neutralMass,formula,SMILES,InChI The neutral monoisotopic mass, chemical formula, SMILES or InChI for the suspect. (data from one of these columns are mandatory in case no value from the mz column is available for a suspect)

mz The ionized m/z of the suspect. (mandatory unless it can be calculated from one of the aforementioned columns)

adduct A character that can be converted with as.adduct. (mandatory unless data from the mz column is available or the adduct argument is set)

fragments_mz,fragments_formula One or more MS/MS fragments (specified as m/z or formulae, respectively). Multiple values can be specified by separating them with a semicolon (;). This data is used by annotateSuspects to report detected MS/MS fragments and calculate identification levels. (optional)

References

OBoyle NM, Banck M, James CA, Morley C, Vandermeersch T, Hutchison GR (2011). “Open Babel: An open chemical toolbox.” Journal of Cheminformatics, 3(1). doi: 10.1186/1758-2946-3-33 , https://doi.org/10.1186/1758-2946-3-33.

Daniel Müllner (2013). fastcluster: Fast Hierarchical, Agglomerative Clustering Routines for R and Python. Journal of Statistical Software, 53(9), 1-18. URL http://www.jstatsoft.org/v53/i09/.

See also