Target and suspect screening

Utilities to screen for analytes with known or suspected identity.

screenSuspects(
  obj,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE
)

# S4 method for features
screenSuspects(
  obj,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE
)

# S4 method for featureGroups
screenSuspects(
  obj,
  suspects,
  rtWindow = 12,
  mzWindow = 0.005,
  adduct = NULL,
  skipInvalid = TRUE
)

# S4 method for featureGroups
groupFeaturesScreening(fGroups, scr)

importFeatureGroupsBrukerTASQ(path, analysisInfo, clusterRTWindow = 12)

Arguments

obj
The object that should be screened (i.e. features or featureGroups).

suspects
A data.frame that must contain a "name" column (the analyte name) and at least a "mz", "neutralMass", "formula", "SMILES", "InChI" column (with the ionized m/z value, the neutral monoisotopic mass or the chemical formula/SMILES/InChI character string for the molecule, respectively). If an ion mass needs to be calculated (i.e. no valid data is available in the mz column) then data is tried to be used from the columns in the aforementioned order. Furthermore, if ion masses need to be calculated then the adduct must be specified either with the adduct function argument or by an "adduct" column containing a character that can be converted with as.adduct (e.g. "[M+H]+"). In addition, a column "rt" can be included to specify the retention time (if unspecified no retention times are checked).

rtWindow, mzWindow
The retention time window (in seconds) and m/z window that will be used for matching a suspect (+/- feature data).

adduct
An adduct object (or something that can be converted to it with as.adduct). Examples: "[M-H]-", "[M+Na]+". Can be NULL if either a "mz" or "adduct" column is present in the suspects argument.

skipInvalid
If set to TRUE then suspects with invalid data (e.g. missing names or other missing data) will be ignored with a warning. Similarly, any suspects for which mass calculation failed (when no mz column is present in the suspect list), for instance, due to invalid SMILES, will be ignored with a warning.

fGroups
The featureGroups object that should be transformed (and was used to obtain the screening results).

scr
The screening results table returned by screenSuspects.

path
The file path to an Excel export of the Global results table from TASQ, converted to .csv format.

analysisInfo
A table with analysis information.

clusterRTWindow
This retention time window (in seconds) is used to group hits across analyses together. See also the details section.

Value

screenSuspects will return a table (a data.table) with detected suspects and details such as retention and m/z values. If a suspect is matched on multiple features/feature groups then each hit is reported as a separate row.

groupFeaturesScreening returns a modified featureGroups object in which those feature groups that were not matched by any suspects are removed and others renamed by the respective suspect name. In case of duplicate suspect results, feature group names are made unique with make.unique.

importFeatureGroupsBrukerTASQ returns a new featureGroups object containing converted screening results from Bruker TASQ.

Details

Besides 'full non-target analysis', where compounds may be identitified with little to no prior knowledge, a common strategy is to screen for compounds with known or suspected identity. This may be a generally favourable approach if possible, as it can significantly reduce the load on data interpretation.

screenSuspects will screen a set of suspects (provided as a data.frame) within an features or featureGroups object.

groupFeaturesScreening uses results from screenSuspects to transform an existing featureGroups object by (1) renaming any matched feature groups by the respective name of the suspect and (2) filtering out any feature groups that were not matched. A common workflow is to first obtain and group features (with e.g. findFeatures and groupFeatures), screen them with screenSuspects, convert the featureGroups object that was used for screening with this method function and continue any further workflow steps such as compound identification as with 'regular' featureGroups.

importFeatureGroupsBrukerTASQ will convert screening results from Bruker TASQ to a featureGroups object. The feature groups across analyses are formed based on the name of suspects and their closeness in retention time. The latter is necessary because TASQ does not necessarily perform checks on retention times and may therefore assign a suspect to peaks with different retention times across analyses (or within a single analysis). Hence, suspects with equal names are hierarchically clustered on their retention times (using fastcluster) to form the feature groups. The cut-off value for this is specified by the clusterRTWindow argument. The input for this function is obtained by generating an Excel export of the 'global' results and subsequently converting the file to .csv format. Similar to groupFeaturesScreening, this method will return an object that is suitable for any further workflow processes.

Note

Both groupFeaturesScreening and importFeatureGroupsBrukerTASQ use names from targets/suspects as feature group names, therefore, it is important that these are file-compatible names when reporting data csv or pdf data.

For screenSuspects in some cases you may need to install OpenBabel (e.g. when only InChI data is available for mass calculation).

Please note that groupFeaturesScreening method can only transform the featureGroups object that was used to obtain the given screening results.

importFeatureGroupsBrukerTASQ will use estimated min/max values for retention times and dummy min/max m/z values for conversion to features, since this information is not (readily) available. Hence, when plotting, for instance, extracted ion chromatograms (with plotEIC) the integrated chromatographic peak range shown is incorrect.

References

Daniel Müllner (2013). fastcluster: Fast Hierarchical, Agglomerative Clustering Routines for R and Python. Journal of Statistical Software, 53(9), 1-18. URL http://www.jstatsoft.org/v53/i09/.