Target and suspect screening

Utilities to screen for targets with known or suspected identity.

screenTargets(obj, targets, rtWindow = 12, mzWindow = 0.005)

# S4 method for features
screenTargets(obj, targets, rtWindow = 12,
  mzWindow = 0.005)

# S4 method for featureGroups
screenTargets(obj, targets, rtWindow = 12,
  mzWindow = 0.005)

# S4 method for featureGroups
groupFeaturesScreening(fGroups, scr)

importFeatureGroupsBrukerTASQ(path, analysisInfo)

Arguments

obj
The object that should be screened (i.e. features or featureGroups).

targets
A data.frame consisting of mandatory columns "name" (the targeted analyte name) and "mz" (the m/z value of the target). In addition, a column "rt" can be used to specify the retention time (if unspecified no retention times are checked).

rtWindow, mzWindow
The retention time window (in seconds) and m/z window that will be used for matching a target (+/- feature data).

fGroups
The featureGroups object that should be transformed (and was used to obtain the screening results).

scr
The screening results table returned by screenTargets.

path
The file path to an Excel export of the Global results table from TASQ, converted to .csv format.

analysisInfo
A table with analysis information.

Value

screenTargets will return a table (a data.table) with detected targets and details such as retention and m/z values. If a target is matched on multiple features/feature groups then each hit is reported as a separate row. In case concentration values were specified in the analysis info then the featureGroups method will also report correlation coefficients ('RSQ' column) calculated from the linear regression between intensity values and specified concentrations.

groupFeaturesScreening returns a modified featureGroups object in which those feature groups that were not matched by any targets are removed and others renamed by the respective target name. In case of duplicate target results, feature group names are made unique with make.unique.

importFeatureGroupsBrukerTASQ returns a new featureGroups object containing converted screening results from Bruker TASQ.

Details

Besides 'full non-target analysis', where compounds may be identitified with little to no prior knowledge, a common strategy is to screen for compounds with known or suspected identity. This may be a generally favourable approach if possible, as it can significantly reduce the load on data interpretation.

screenTargets will screen a set of targets (provided as a data.frame) within an features or featureGroups object.

groupFeaturesScreening uses results from screenTargets to transform an existing featureGroups object by (1) renaming any matched feature groups by the respective name of the target and (2) filtering out any feature groups that were not matched. A common workflow is to first obtain and group features (with e.g. findFeatures and groupFeatures), screen them with screenTargets, convert the featureGroups object that was used for screening with this method function and continue any further workflow steps such as compound identification as with 'regular' featureGroups.

importFeatureGroupsBrukerTASQ will convert screening results from Bruker TASQ to a featureGroups object. Groups are made based on target analyte names and individual results per analysis are used to generate the features. The input for this function is obtained by generating an Excel export of the 'global' results and subsequently converting the file to .csv format. Similar to groupFeaturesScreening, this method will return an object that is suitable for any further workflow processes.

Note

Both groupFeaturesScreening and importFeatureGroupsBrukerTASQ use names from targets as feature group names, therefore, it is important that these are file-compatible names when reporting data.

Please note that groupFeaturesScreening method can only transform the featureGroups object that was used to obtain the given screening results.

importFeatureGroupsBrukerTASQ will use dummy values for retention time and m/z values for conversion to features, since this information is not available. Hence, when plotting, for instance, extracted ion chromatograms (with plotEIC) the integrated chromatographic peak range shown is incorrect.