Base transformation products (TP) class

Holds information for all TPs for a set of parents.

# S4 method for transformationProducts
parents(TPs)

# S4 method for transformationProducts
products(TPs)

# S4 method for transformationProducts
length(x)

# S4 method for transformationProducts
names(x)

# S4 method for transformationProducts
show(object)

# S4 method for transformationProducts,ANY,missing,missing
[(x, i, j, ..., drop = TRUE)

# S4 method for transformationProducts,ANY,missing
[[(x, i, j)

# S4 method for transformationProducts
$(x, name)

# S4 method for transformationProducts
as.data.table(x)

# S4 method for transformationProducts
convertToSuspects(TPs, includeParents = FALSE)

Arguments

TPs, x, object
transformationProducts object to be accessed

i
For [/[[: A numeric or character value which is used to select parents by their index or name, respectively (for the order/names see names()).

For [: Can also be logical to perform logical selection (similar to regular vectors). If missing all parents are selected.

For [[: should be a scalar value.

...
Unused.

drop, j
ignored.

name
The parent name (partially matched).

includeParents
If TRUE then parents are also included in the returned suspect list.

Details

This class holds all generated data for transformation products for a set of parents. The class is virtual and derived objects are created by TP generators.

The TP data in objects from this class include a retDir column. These are numeric values that hint what the the chromatographic retention order of a TP might be compared to its parent: a value of -1 means it will elute earlier, 1 it will elute later and 0 that there is no significant difference or the direction is unknown. These values are based on a typical reversed phase separation. When generateTPsBioTransformer or generateTPsLibrary was used to generate the data, the retDir values are based on calculated log P values of the parent and its TPs.

Methods (by generic)

parents: Accessor method for the parents slot of a transformationProducts class.

products: Accessor method for the products slot.

length: Obtain total number of transformation products.

names: Obtain the names of all parents in this object.

show: Show summary information for this object.

[: Subset on parents.

[[: Extracts a table with TPs for a parent.

$: Extracts a table with TPs for a parent.

as.data.table: Returns all TP data in a table.

convertToSuspects: Converts this object to a suspect list that can be used as input for screenSuspects.

Slots

parents

A data.table with metadata for all parents that have TPs in this object. Use the parents method for access.

products

A list with data.table entries with TP information for each parent. Use the products method for access.

S4 class hierarchy

workflowStep

transformationProducts

transformationProductsBT

transformationProductsLibrary

transformationProductsLogic

See also