(Virtual) Base class for all workflow objects.

All workflow objects (e.g. featureGroups, compounds, etc) are derived from this class. Objects from this class are never created directly.

# S4 method for workflowStep
algorithm(obj)

# S4 method for workflowStep
as.data.table(x, keep.rownames = FALSE, ...)

# S4 method for workflowStep
as.data.frame(x, row.names = NULL,
  optional = FALSE, ...)

# S4 method for workflowStep
show(object)

Arguments

obj, x, object
An object (derived from) this class.

keep.rownames
Ignored.

...
Method specific arguments. Please see the documentation of the derived classes.

row.names, optional
Ignored.

Methods (by generic)

algorithm: Returns the algorithm that was used to generate an object.

as.data.table: Summarizes the data in this object and returns this as a data.table.

as.data.frame: This method simply calls as.data.table and converts the result to a classic a data.frame.

show: Shows summary information for this object.

Slots

algorithm

The algorithm that was used to generate this object. Use the algorithm method for access.

S4 class hierarchy

workflowStep

features

featuresFromFeatGroups

featuresConsensus

featuresBruker

featuresEnviPick

featuresOpenMS

featuresBrukerTASQ

featuresXCMS

featuresXCMS3

featureGroups

featureGroupsBruker

featureGroupsConsensus

featureGroupsEnviMass

featureGroupsOpenMS

featureGroupsScreening

featureGroupsBrukerTASQ

featureGroupsXCMS

featureGroupsXCMS3

components

componentsReduced

componentsCamera

componentsIntClust

componentsNT

componentsRC

compounds

compoundsConsensus

compoundsMF

formulas

MSPeakLists